Extract estimated parameters from a list of BUGS outputs

From the output of the function fit_trend() (or read_bugs()), this function extracts estimated parameters into a data frame.

The resulting data frame has no particular use in popbayes but it can be useful for users.

bugs_to_df(data)

Arguments

data
a named list of BUGS outputs. The output of fit_trend() or read_bugs().

Value

A data frame.

Examples

if (FALSE) {
## Get count series names ----
popbayes::list_series()

## Import BUGS outputs for one count series ----
bugs <- popbayes::read_bugs(series = "garamba__alcelaphus_buselaphus")

## Extract estimated parameters ----
popbayes::bugs_to_df(bugs)
}