Counts conversion data

This dataset contains information to convert individual counts estimated from a field method to a preferred field method. The field method can be A (aerial counts) or G (ground counts). This dataset can be used by user to add missing species. See format_data() for further information.

conversion_data

Format

A data frame with 15 rows (species) and the following columns:

order

the order of the species

family

the family of the species

species

the species binomial name

english

the species English name

french

the species French name

category

the detectability category of the species. One of MLB for Medium-sized Light and Brown species (20-150kg), LLB for Large Light and Brown species (>150kg), LD for Large Dark (>150kg), Elephant, and Giraffe

pref_field_method

the preferred field method of the species. One of A for Aerial counts, and G for Ground counts

conversion_fact

the conversion factor (corresponding to the detectability category) used to convert counts to the preferred field method

Examples

data("conversion_data")
conversion_data

#>           order       family                species             english
#> 1  Artiodactyla      Bovidae     Aepyceros melampus              Impala
#> 2  Artiodactyla      Bovidae  Alcelaphus buselaphus          Hartebeest
#> 3  Artiodactyla      Bovidae  Connochaetes taurinus     Blue wildebeest
#> 4  Artiodactyla      Bovidae     Damaliscus lunatus               Tiang
#> 5  Artiodactyla      Bovidae     Eudorcas rufifrons Red-fronted gazelle
#> 6  Artiodactyla      Bovidae    Hippotragus equinus       Roan Antelope
#> 7  Artiodactyla      Bovidae   Kobus ellipsiprymnus           Waterbuck
#> 8  Artiodactyla      Bovidae              Kobus kob                 Kob
#> 9  Artiodactyla      Bovidae         Ourebia ourebi               Oribi
#> 10 Artiodactyla      Bovidae        Redunca redunca      Bohor reedbuck
#> 11 Artiodactyla      Bovidae        Syncerus caffer     African Buffalo
#> 12 Artiodactyla      Bovidae  Tragelaphus derbianus         Giant eland
#> 13 Artiodactyla      Bovidae   Tragelaphus scriptus            Bushbuck
#> 14 Artiodactyla   Giraffidae Giraffa camelopardalis             Giraffe
#> 15  Proboscidea Elephantidae     Loxodonta africana    Savanna Elephant
#>              french category pref_field_method conversion_fact
#> 1            Impala      MLB                 G           6.747
#> 2            Bubale      LLB                 G           2.302
#> 3         Gnou bleu      LLB                 G           2.302
#> 4              Topi      MLB                 G           6.747
#> 5  Gazelle Rufifron      MLB                 G           6.747
#> 6       Hippotrague      LLB                 G           2.302
#> 7      Cobe Defassa      MLB                 G           6.747
#> 8    Cobe de Buffon      MLB                 G           6.747
#> 9            Ourebi      MLB                 G           6.747
#> 10          Redunca      MLB                 G           6.747
#> 11           Buffle       LD                 A           0.561
#> 12   Eland de derby      LLB                 G           2.302
#> 13    Guib Harnache      MLB                 G           6.747
#> 14           Girafe  Giraffe                 A           3.011
#> 15         Elephant Elephant                 A           0.659