Extract the count series corresponding to a location and/or a species in a list

This function identifies the count series relative to a species and/or a location in a named list like the output of function format_data(). If both species and location are provided, the series of counts of the species at the specified location is extracted. Otherwise, all series corresponding to the specified criterion (species or location) are extracted.

filter_series(data, species = NULL, location = NULL)

Arguments

data
a named list. The output of function format_data().

species
a character string. A species name.

location
a character string. A site name.

Value

A subset of data, i.e. a named list.

Examples

if (FALSE) {
## Load Garamba dataset ----
file_path <- system.file("extdata", "garamba_survey.csv", 
                         package = "popbayes")
                         
garamba <- read.csv(file = file_path)

## Format dataset ----
garamba_formatted <- popbayes::format_data(garamba)

## Get series names ----
names(garamba_formatted)

## Or...
popbayes::list_series()

## Get data for Alcelaphus buselaphus (at all sites) ----
popbayes::filter_series(garamba_formatted, species = "Alcelaphus buselaphus")

## Get data at Garamba (for all species) ----
popbayes::filter_series(garamba_formatted, location = "Garamba")

## Get data for Alcelaphus buselaphus at Garamba only ----
popbayes::filter_series(garamba_formatted, location = "Garamba",
                        species = "Alcelaphus buselaphus")
}