Subset a series based on a location and/or a species

From the output of the function format_data(), this function subsets the list of series based on a species name and/or a site. It both the species and the location are provided, the resulting list will have a length of 1 (data for the species at the location), otherwise the list will have a length >= 1.

find_series(data, species = NULL, location = NULL)

Arguments

data
a named list. The output of the function format_data().

species
a character of length 1. A species name.

location
a character of length 1. A site name.

Value

A subset of data, i.e. a named list.

Examples

data(garamba)

## Format dataset ----
garamba_formatted <- format_data(data = garamba)

#> ✔ Conversion data found in 'popbayes'.

#> ✔ Detecting 10 series (with 1 locations and 10 species).

## Get series names ----
names(garamba_formatted)

#>  [1] "garamba__alcelaphus_buselaphus"  "garamba__giraffa_camelopardalis"
#>  [3] "garamba__hippotragus_equinus"    "garamba__kobus_ellipsiprymnus"  
#>  [5] "garamba__kobus_kob"              "garamba__loxodonta_africana"    
#>  [7] "garamba__ourebia_ourebi"         "garamba__redunca_redunca"       
#>  [9] "garamba__syncerus_caffer"        "garamba__tragelaphus_scriptus"  

## Get data for Alcelaphus buselaphus (at all sites) ----
find_series(data = garamba_formatted, species = "Alcelaphus buselaphus")

#> ✔ Found 1 series with 'Alcelaphus buselaphus'.

#> $garamba__alcelaphus_buselaphus
#> $garamba__alcelaphus_buselaphus$location
#> [1] "Garamba"
#> 
#> $garamba__alcelaphus_buselaphus$species
#> [1] "Alcelaphus buselaphus"
#> 
#> $garamba__alcelaphus_buselaphus$years
#>  [1] 1976 1983 1984 1986 1991 1993 1995 1998 2000 2002 2003 2004 2012 2014 2017
#> 
#> $garamba__alcelaphus_buselaphus$n_years
#> [1] 15
#> 
#> $garamba__alcelaphus_buselaphus$stat_methods
#> [1] "S" "T"
#> 
#> $garamba__alcelaphus_buselaphus$field_methods
#> [1] "A"
#> 
#> $garamba__alcelaphus_buselaphus$pref_field_method
#> [1] "G"
#> 
#> $garamba__alcelaphus_buselaphus$conversion_fact
#> [1] 2.302
#> 
#> $garamba__alcelaphus_buselaphus$data_original
#>    location               species year stat_method field_method counts_orig
#> 1   Garamba Alcelaphus buselaphus 1976           S            A        7750
#> 2   Garamba Alcelaphus buselaphus 1983           S            A        1932
#> 3   Garamba Alcelaphus buselaphus 1984           S            A        1224
#> 4   Garamba Alcelaphus buselaphus 1986           S            A        1705
#> 5   Garamba Alcelaphus buselaphus 1991           S            A         987
#> 6   Garamba Alcelaphus buselaphus 1993           S            A        3444
#> 7   Garamba Alcelaphus buselaphus 1995           S            A        2819
#> 8   Garamba Alcelaphus buselaphus 1998           S            A        1685
#> 9   Garamba Alcelaphus buselaphus 2000           S            A        1169
#> 10  Garamba Alcelaphus buselaphus 2002           S            A        1139
#> 11  Garamba Alcelaphus buselaphus 2003           S            A        1595
#> 12  Garamba Alcelaphus buselaphus 2004           S            A        1204
#> 13  Garamba Alcelaphus buselaphus 2012           T            A         552
#> 14  Garamba Alcelaphus buselaphus 2014           T            A         698
#> 15  Garamba Alcelaphus buselaphus 2017           T            A        1051
#>    lower_ci_orig upper_ci_orig
#> 1           6280          9220
#> 2           1120          2744
#> 3            782          1666
#> 4           1116          2294
#> 5            663          1311
#> 6           1290          5598
#> 7           1620          4018
#> 8           1287          2083
#> 9            945          1393
#> 10           907          1371
#> 11          1142          2048
#> 12           811          1597
#> 13            NA            NA
#> 14            NA            NA
#> 15            NA            NA
#> 
#> $garamba__alcelaphus_buselaphus$data_converted
#>    location               species year stat_method field_method counts_conv
#> 1   Garamba Alcelaphus buselaphus 1976           S            A   3366.6377
#> 2   Garamba Alcelaphus buselaphus 1983           S            A    839.2702
#> 3   Garamba Alcelaphus buselaphus 1984           S            A    531.7116
#> 4   Garamba Alcelaphus buselaphus 1986           S            A    740.6603
#> 5   Garamba Alcelaphus buselaphus 1991           S            A    428.7576
#> 6   Garamba Alcelaphus buselaphus 1993           S            A   1496.0904
#> 7   Garamba Alcelaphus buselaphus 1995           S            A   1224.5873
#> 8   Garamba Alcelaphus buselaphus 1998           S            A    731.9722
#> 9   Garamba Alcelaphus buselaphus 2000           S            A    507.8193
#> 10  Garamba Alcelaphus buselaphus 2002           S            A    494.7871
#> 11  Garamba Alcelaphus buselaphus 2003           S            A    692.8758
#> 12  Garamba Alcelaphus buselaphus 2004           S            A    523.0235
#> 13  Garamba Alcelaphus buselaphus 2012           T            A    239.7915
#> 14  Garamba Alcelaphus buselaphus 2014           T            A    303.2146
#> 15  Garamba Alcelaphus buselaphus 2017           T            A    456.5595
#>    lower_ci_conv upper_ci_conv field_method_conv
#> 1      2728.0626     4005.2129                 G
#> 2       486.5334     1192.0070                 G
#> 3       339.7046      723.7185                 G
#> 4       484.7958      996.5248                 G
#> 5       288.0104      569.5048                 G
#> 6       560.3823     2431.7984                 G
#> 7       703.7359     1745.4387                 G
#> 8       559.0791      904.8653                 G
#> 9       410.5126      605.1260                 G
#> 10      394.0052      595.5691                 G
#> 11      496.0904      889.6612                 G
#> 12      352.3023      693.7446                 G
#> 13      227.8019      287.7498                 G
#> 14      288.0539      363.8575                 G
#> 15      433.7315      547.8714                 G
#> 
#> 

## Get data for at Garamba (for all species) ----
find_series(data = garamba_formatted, location = "Garamba")

#> ✔ Found 10 series with 'Garamba'.

#> $garamba__alcelaphus_buselaphus
#> $garamba__alcelaphus_buselaphus$location
#> [1] "Garamba"
#> 
#> $garamba__alcelaphus_buselaphus$species
#> [1] "Alcelaphus buselaphus"
#> 
#> $garamba__alcelaphus_buselaphus$years
#>  [1] 1976 1983 1984 1986 1991 1993 1995 1998 2000 2002 2003 2004 2012 2014 2017
#> 
#> $garamba__alcelaphus_buselaphus$n_years
#> [1] 15
#> 
#> $garamba__alcelaphus_buselaphus$stat_methods
#> [1] "S" "T"
#> 
#> $garamba__alcelaphus_buselaphus$field_methods
#> [1] "A"
#> 
#> $garamba__alcelaphus_buselaphus$pref_field_method
#> [1] "G"
#> 
#> $garamba__alcelaphus_buselaphus$conversion_fact
#> [1] 2.302
#> 
#> $garamba__alcelaphus_buselaphus$data_original
#>    location               species year stat_method field_method counts_orig
#> 1   Garamba Alcelaphus buselaphus 1976           S            A        7750
#> 2   Garamba Alcelaphus buselaphus 1983           S            A        1932
#> 3   Garamba Alcelaphus buselaphus 1984           S            A        1224
#> 4   Garamba Alcelaphus buselaphus 1986           S            A        1705
#> 5   Garamba Alcelaphus buselaphus 1991           S            A         987
#> 6   Garamba Alcelaphus buselaphus 1993           S            A        3444
#> 7   Garamba Alcelaphus buselaphus 1995           S            A        2819
#> 8   Garamba Alcelaphus buselaphus 1998           S            A        1685
#> 9   Garamba Alcelaphus buselaphus 2000           S            A        1169
#> 10  Garamba Alcelaphus buselaphus 2002           S            A        1139
#> 11  Garamba Alcelaphus buselaphus 2003           S            A        1595
#> 12  Garamba Alcelaphus buselaphus 2004           S            A        1204
#> 13  Garamba Alcelaphus buselaphus 2012           T            A         552
#> 14  Garamba Alcelaphus buselaphus 2014           T            A         698
#> 15  Garamba Alcelaphus buselaphus 2017           T            A        1051
#>    lower_ci_orig upper_ci_orig
#> 1           6280          9220
#> 2           1120          2744
#> 3            782          1666
#> 4           1116          2294
#> 5            663          1311
#> 6           1290          5598
#> 7           1620          4018
#> 8           1287          2083
#> 9            945          1393
#> 10           907          1371
#> 11          1142          2048
#> 12           811          1597
#> 13            NA            NA
#> 14            NA            NA
#> 15            NA            NA
#> 
#> $garamba__alcelaphus_buselaphus$data_converted
#>    location               species year stat_method field_method counts_conv
#> 1   Garamba Alcelaphus buselaphus 1976           S            A   3366.6377
#> 2   Garamba Alcelaphus buselaphus 1983           S            A    839.2702
#> 3   Garamba Alcelaphus buselaphus 1984           S            A    531.7116
#> 4   Garamba Alcelaphus buselaphus 1986           S            A    740.6603
#> 5   Garamba Alcelaphus buselaphus 1991           S            A    428.7576
#> 6   Garamba Alcelaphus buselaphus 1993           S            A   1496.0904
#> 7   Garamba Alcelaphus buselaphus 1995           S            A   1224.5873
#> 8   Garamba Alcelaphus buselaphus 1998           S            A    731.9722
#> 9   Garamba Alcelaphus buselaphus 2000           S            A    507.8193
#> 10  Garamba Alcelaphus buselaphus 2002           S            A    494.7871
#> 11  Garamba Alcelaphus buselaphus 2003           S            A    692.8758
#> 12  Garamba Alcelaphus buselaphus 2004           S            A    523.0235
#> 13  Garamba Alcelaphus buselaphus 2012           T            A    239.7915
#> 14  Garamba Alcelaphus buselaphus 2014           T            A    303.2146
#> 15  Garamba Alcelaphus buselaphus 2017           T            A    456.5595
#>    lower_ci_conv upper_ci_conv field_method_conv
#> 1      2728.0626     4005.2129                 G
#> 2       486.5334     1192.0070                 G
#> 3       339.7046      723.7185                 G
#> 4       484.7958      996.5248                 G
#> 5       288.0104      569.5048                 G
#> 6       560.3823     2431.7984                 G
#> 7       703.7359     1745.4387                 G
#> 8       559.0791      904.8653                 G
#> 9       410.5126      605.1260                 G
#> 10      394.0052      595.5691                 G
#> 11      496.0904      889.6612                 G
#> 12      352.3023      693.7446                 G
#> 13      227.8019      287.7498                 G
#> 14      288.0539      363.8575                 G
#> 15      433.7315      547.8714                 G
#> 
#> 
#> $garamba__giraffa_camelopardalis
#> $garamba__giraffa_camelopardalis$location
#> [1] "Garamba"
#> 
#> $garamba__giraffa_camelopardalis$species
#> [1] "Giraffa camelopardalis"
#> 
#> $garamba__giraffa_camelopardalis$years
#>  [1] 1976 1983 1984 1986 1991 1993 1995 1998 2000 2002 2003 2004 2012 2014 2017
#> 
#> $garamba__giraffa_camelopardalis$n_years
#> [1] 15
#> 
#> $garamba__giraffa_camelopardalis$stat_methods
#> [1] "S" "T"
#> 
#> $garamba__giraffa_camelopardalis$field_methods
#> [1] "A"
#> 
#> $garamba__giraffa_camelopardalis$pref_field_method
#> [1] "A"
#> 
#> $garamba__giraffa_camelopardalis$conversion_fact
#> [1] 3.011
#> 
#> $garamba__giraffa_camelopardalis$data_original
#>    location                species year stat_method field_method counts_orig
#> 1   Garamba Giraffa camelopardalis 1976           S            A         350
#> 2   Garamba Giraffa camelopardalis 1983           S            A         175
#> 3   Garamba Giraffa camelopardalis 1984           S            A         237
#> 4   Garamba Giraffa camelopardalis 1986           S            A         153
#> 5   Garamba Giraffa camelopardalis 1991           S            A         346
#> 6   Garamba Giraffa camelopardalis 1993           S            A         347
#> 7   Garamba Giraffa camelopardalis 1995           S            A         178
#> 8   Garamba Giraffa camelopardalis 1998           S            A         144
#> 9   Garamba Giraffa camelopardalis 2000           S            A         118
#> 10  Garamba Giraffa camelopardalis 2002           S            A          62
#> 11  Garamba Giraffa camelopardalis 2003           S            A          62
#> 12  Garamba Giraffa camelopardalis 2004           S            A         185
#> 13  Garamba Giraffa camelopardalis 2012           T            A          11
#> 14  Garamba Giraffa camelopardalis 2014           T            A          42
#> 15  Garamba Giraffa camelopardalis 2017           T            A          42
#>    lower_ci_orig upper_ci_orig
#> 1            100           600
#> 2             12           338
#> 3             93           381
#> 4             13           293
#> 5            143           549
#> 6            347           766
#> 7            178           388
#> 8             71           217
#> 9             54           182
#> 10            49            75
#> 11            62           137
#> 12            33           337
#> 13            NA            NA
#> 14            NA            NA
#> 15            NA            NA
#> 
#> $garamba__giraffa_camelopardalis$data_converted
#>    location                species year stat_method field_method counts_conv
#> 1   Garamba Giraffa camelopardalis 1976           S            A         350
#> 2   Garamba Giraffa camelopardalis 1983           S            A         175
#> 3   Garamba Giraffa camelopardalis 1984           S            A         237
#> 4   Garamba Giraffa camelopardalis 1986           S            A         153
#> 5   Garamba Giraffa camelopardalis 1991           S            A         346
#> 6   Garamba Giraffa camelopardalis 1993           S            A         347
#> 7   Garamba Giraffa camelopardalis 1995           S            A         178
#> 8   Garamba Giraffa camelopardalis 1998           S            A         144
#> 9   Garamba Giraffa camelopardalis 2000           S            A         118
#> 10  Garamba Giraffa camelopardalis 2002           S            A          62
#> 11  Garamba Giraffa camelopardalis 2003           S            A          62
#> 12  Garamba Giraffa camelopardalis 2004           S            A         185
#> 13  Garamba Giraffa camelopardalis 2012           T            A          11
#> 14  Garamba Giraffa camelopardalis 2014           T            A          42
#> 15  Garamba Giraffa camelopardalis 2017           T            A          42
#>    lower_ci_conv upper_ci_conv field_method_conv
#> 1         100.00         600.0                 A
#> 2          12.00         338.0                 A
#> 3          93.00         381.0                 A
#> 4          13.00         293.0                 A
#> 5         143.00         549.0                 A
#> 6         347.00         766.0                 A
#> 7         178.00         388.0                 A
#> 8          71.00         217.0                 A
#> 9          54.00         182.0                 A
#> 10         49.00          75.0                 A
#> 11         62.00         137.0                 A
#> 12         33.00         337.0                 A
#> 13         10.45          13.2                 A
#> 14         39.90          50.4                 A
#> 15         39.90          50.4                 A
#> 
#> 
#> $garamba__hippotragus_equinus
#> $garamba__hippotragus_equinus$location
#> [1] "Garamba"
#> 
#> $garamba__hippotragus_equinus$species
#> [1] "Hippotragus equinus"
#> 
#> $garamba__hippotragus_equinus$years
#>  [1] 1976 1983 1984 1986 1991 1993 1995 1998 2000 2002 2003 2004 2012 2014 2017
#> 
#> $garamba__hippotragus_equinus$n_years
#> [1] 15
#> 
#> $garamba__hippotragus_equinus$stat_methods
#> [1] "S" "T"
#> 
#> $garamba__hippotragus_equinus$field_methods
#> [1] "A"
#> 
#> $garamba__hippotragus_equinus$pref_field_method
#> [1] "G"
#> 
#> $garamba__hippotragus_equinus$conversion_fact
#> [1] 2.302
#> 
#> $garamba__hippotragus_equinus$data_original
#>    location             species year stat_method field_method counts_orig
#> 1   Garamba Hippotragus equinus 1976           S            A         360
#> 2   Garamba Hippotragus equinus 1983           S            A          91
#> 3   Garamba Hippotragus equinus 1984           S            A           0
#> 4   Garamba Hippotragus equinus 1986           S            A          34
#> 5   Garamba Hippotragus equinus 1991           S            A          13
#> 6   Garamba Hippotragus equinus 1993           S            A         150
#> 7   Garamba Hippotragus equinus 1995           S            A          81
#> 8   Garamba Hippotragus equinus 1998           S            A           8
#> 9   Garamba Hippotragus equinus 2000           S            A           0
#> 10  Garamba Hippotragus equinus 2002           S            A           0
#> 11  Garamba Hippotragus equinus 2003           S            A          57
#> 12  Garamba Hippotragus equinus 2004           S            A         281
#> 13  Garamba Hippotragus equinus 2012           T            A          11
#> 14  Garamba Hippotragus equinus 2014           T            A          13
#> 15  Garamba Hippotragus equinus 2017           T            A          16
#>    lower_ci_orig upper_ci_orig
#> 1              0           890
#> 2             91           215
#> 3              0             0
#> 4              9            59
#> 5              1            25
#> 6            150           405
#> 7              3           159
#> 8              1            15
#> 9              0             0
#> 10             0             0
#> 11            57           124
#> 12           281           724
#> 13            NA            NA
#> 14            NA            NA
#> 15            NA            NA
#> 
#> $garamba__hippotragus_equinus$data_converted
#>    location             species year stat_method field_method counts_conv
#> 1   Garamba Hippotragus equinus 1976           S            A  156.385752
#> 2   Garamba Hippotragus equinus 1983           S            A   39.530843
#> 3   Garamba Hippotragus equinus 1984           S            A    0.000000
#> 4   Garamba Hippotragus equinus 1986           S            A   14.769765
#> 5   Garamba Hippotragus equinus 1991           S            A    5.647263
#> 6   Garamba Hippotragus equinus 1993           S            A   65.160730
#> 7   Garamba Hippotragus equinus 1995           S            A   35.186794
#> 8   Garamba Hippotragus equinus 1998           S            A    3.475239
#> 9   Garamba Hippotragus equinus 2000           S            A    0.000000
#> 10  Garamba Hippotragus equinus 2002           S            A    0.000000
#> 11  Garamba Hippotragus equinus 2003           S            A   24.761077
#> 12  Garamba Hippotragus equinus 2004           S            A  122.067767
#> 13  Garamba Hippotragus equinus 2012           T            A    4.778454
#> 14  Garamba Hippotragus equinus 2014           T            A    5.647263
#> 15  Garamba Hippotragus equinus 2017           T            A    6.950478
#>    lower_ci_conv upper_ci_conv field_method_conv
#> 1      0.0000000    386.620330                 G
#> 2     39.5308427     93.397046                 G
#> 3      0.0000000      0.000000                 G
#> 4      3.9096438     25.629887                 G
#> 5      0.4344049     10.860122                 G
#> 6     65.1607298    175.933970                 G
#> 7      1.3032146     69.070374                 G
#> 8      0.4344049      6.516073                 G
#> 9      0.0000000      0.000000                 G
#> 10     0.0000000      0.000000                 G
#> 11    24.7610773     53.866203                 G
#> 12   122.0677672    314.509123                 G
#> 13     4.5395308      5.734144                 G
#> 14     5.3649001      6.776716                 G
#> 15     6.6029540      8.340573                 G
#> 
#> 
#> $garamba__kobus_ellipsiprymnus
#> $garamba__kobus_ellipsiprymnus$location
#> [1] "Garamba"
#> 
#> $garamba__kobus_ellipsiprymnus$species
#> [1] "Kobus ellipsiprymnus"
#> 
#> $garamba__kobus_ellipsiprymnus$years
#>  [1] 1976 1983 1984 1986 1991 1993 1995 1998 2000 2002 2003 2004 2012 2014 2017
#> 
#> $garamba__kobus_ellipsiprymnus$n_years
#> [1] 15
#> 
#> $garamba__kobus_ellipsiprymnus$stat_methods
#> [1] "S" "T"
#> 
#> $garamba__kobus_ellipsiprymnus$field_methods
#> [1] "A"
#> 
#> $garamba__kobus_ellipsiprymnus$pref_field_method
#> [1] "G"
#> 
#> $garamba__kobus_ellipsiprymnus$conversion_fact
#> [1] 6.747
#> 
#> $garamba__kobus_ellipsiprymnus$data_original
#>    location              species year stat_method field_method counts_orig
#> 1   Garamba Kobus ellipsiprymnus 1976           S            A        3680
#> 2   Garamba Kobus ellipsiprymnus 1983           S            A        2215
#> 3   Garamba Kobus ellipsiprymnus 1984           S            A         568
#> 4   Garamba Kobus ellipsiprymnus 1986           S            A        1322
#> 5   Garamba Kobus ellipsiprymnus 1991           S            A         718
#> 6   Garamba Kobus ellipsiprymnus 1993           S            A        1113
#> 7   Garamba Kobus ellipsiprymnus 1995           S            A        1680
#> 8   Garamba Kobus ellipsiprymnus 1998           S            A        1362
#> 9   Garamba Kobus ellipsiprymnus 2000           S            A        1314
#> 10  Garamba Kobus ellipsiprymnus 2002           S            A         797
#> 11  Garamba Kobus ellipsiprymnus 2003           S            A         544
#> 12  Garamba Kobus ellipsiprymnus 2004           S            A         660
#> 13  Garamba Kobus ellipsiprymnus 2012           T            A         660
#> 14  Garamba Kobus ellipsiprymnus 2014           T            A        1120
#> 15  Garamba Kobus ellipsiprymnus 2017           T            A        1436
#>    lower_ci_orig upper_ci_orig
#> 1           2350          5010
#> 2            795          3635
#> 3            275           861
#> 4            866          1778
#> 5            507           929
#> 6            590          1636
#> 7            340          3020
#> 8            929          1795
#> 9            850          1778
#> 10           481          1113
#> 11           317           771
#> 12           374           946
#> 13            NA            NA
#> 14            NA            NA
#> 15            NA            NA
#> 
#> $garamba__kobus_ellipsiprymnus$data_converted
#>    location              species year stat_method field_method counts_conv
#> 1   Garamba Kobus ellipsiprymnus 1976           S            A   545.42760
#> 2   Garamba Kobus ellipsiprymnus 1983           S            A   328.29406
#> 3   Garamba Kobus ellipsiprymnus 1984           S            A    84.18556
#> 4   Garamba Kobus ellipsiprymnus 1986           S            A   195.93894
#> 5   Garamba Kobus ellipsiprymnus 1991           S            A   106.41767
#> 6   Garamba Kobus ellipsiprymnus 1993           S            A   164.96221
#> 7   Garamba Kobus ellipsiprymnus 1995           S            A   248.99956
#> 8   Garamba Kobus ellipsiprymnus 1998           S            A   201.86750
#> 9   Garamba Kobus ellipsiprymnus 2000           S            A   194.75322
#> 10  Garamba Kobus ellipsiprymnus 2002           S            A   118.12657
#> 11  Garamba Kobus ellipsiprymnus 2003           S            A    80.62843
#> 12  Garamba Kobus ellipsiprymnus 2004           S            A    97.82125
#> 13  Garamba Kobus ellipsiprymnus 2012           T            A    97.82125
#> 14  Garamba Kobus ellipsiprymnus 2014           T            A   165.99970
#> 15  Garamba Kobus ellipsiprymnus 2017           T            A   212.83533
#>    lower_ci_conv upper_ci_conv field_method_conv
#> 1      348.30295      742.5522                 G
#> 2      117.83015      538.7580                 G
#> 3       40.75886      127.6123                 G
#> 4      128.35334      263.5245                 G
#> 5       75.14451      137.6908                 G
#> 6       87.44627      242.4781                 G
#> 7       50.39277      447.6063                 G
#> 8      137.69083      266.0442                 G
#> 9      125.98192      263.5245                 G
#> 10      71.29094      164.9622                 G
#> 11      46.98384      114.2730                 G
#> 12      55.43204      140.2105                 G
#> 13      92.93019      117.3855                 G
#> 14     157.69972      199.1996                 G
#> 15     202.19357      255.4024                 G
#> 
#> 
#> $garamba__kobus_kob
#> $garamba__kobus_kob$location
#> [1] "Garamba"
#> 
#> $garamba__kobus_kob$species
#> [1] "Kobus kob"
#> 
#> $garamba__kobus_kob$years
#>  [1] 1976 1983 1984 1986 1991 1993 1995 1998 2000 2002 2003 2004 2012 2014 2017
#> 
#> $garamba__kobus_kob$n_years
#> [1] 15
#> 
#> $garamba__kobus_kob$stat_methods
#> [1] "S" "T"
#> 
#> $garamba__kobus_kob$field_methods
#> [1] "A"
#> 
#> $garamba__kobus_kob$pref_field_method
#> [1] "G"
#> 
#> $garamba__kobus_kob$conversion_fact
#> [1] 6.747
#> 
#> $garamba__kobus_kob$data_original
#>    location   species year stat_method field_method counts_orig lower_ci_orig
#> 1   Garamba Kobus kob 1976           S            A        7180          4880
#> 2   Garamba Kobus kob 1983           S            A        3978          1657
#> 3   Garamba Kobus kob 1984           S            A        3792          3578
#> 4   Garamba Kobus kob 1986           S            A        7222          4621
#> 5   Garamba Kobus kob 1991           S            A        3423          2440
#> 6   Garamba Kobus kob 1993           S            A        6738          3391
#> 7   Garamba Kobus kob 1995           S            A        6601          3568
#> 8   Garamba Kobus kob 1998           S            A        6505          4947
#> 9   Garamba Kobus kob 2000           S            A        3902          2918
#> 10  Garamba Kobus kob 2002           S            A        4728          3056
#> 11  Garamba Kobus kob 2003           S            A        6447          4323
#> 12  Garamba Kobus kob 2004           S            A        5730          4355
#> 13  Garamba Kobus kob 2012           T            A        2803            NA
#> 14  Garamba Kobus kob 2014           T            A        3860            NA
#> 15  Garamba Kobus kob 2017           T            A        3877            NA
#>    upper_ci_orig
#> 1           9480
#> 2           6299
#> 3           4006
#> 4           9823
#> 5           4406
#> 6          10085
#> 7           9634
#> 8           8063
#> 9           4886
#> 10          6400
#> 11          8571
#> 12          7105
#> 13            NA
#> 14            NA
#> 15            NA
#> 
#> $garamba__kobus_kob$data_converted
#>    location   species year stat_method field_method counts_conv lower_ci_conv
#> 1   Garamba Kobus kob 1976           S            A   1064.1767      723.2844
#> 2   Garamba Kobus kob 1983           S            A    589.5954      245.5906
#> 3   Garamba Kobus kob 1984           S            A    562.0276      530.3098
#> 4   Garamba Kobus kob 1986           S            A   1070.4017      684.8970
#> 5   Garamba Kobus kob 1991           S            A    507.3366      361.6422
#> 6   Garamba Kobus kob 1993           S            A    998.6661      502.5937
#> 7   Garamba Kobus kob 1995           S            A    978.3608      528.8276
#> 8   Garamba Kobus kob 1998           S            A    964.1322      733.2148
#> 9   Garamba Kobus kob 2000           S            A    578.3311      432.4885
#> 10  Garamba Kobus kob 2002           S            A    700.7559      452.9420
#> 11  Garamba Kobus kob 2003           S            A    955.5358      640.7292
#> 12  Garamba Kobus kob 2004           S            A    849.2663      645.4721
#> 13  Garamba Kobus kob 2012           T            A    415.4439      394.6717
#> 14  Garamba Kobus kob 2014           T            A    572.1061      543.5008
#> 15  Garamba Kobus kob 2017           T            A    574.6258      545.8945
#>    upper_ci_conv field_method_conv
#> 1      1405.0689                 G
#> 2       933.6001                 G
#> 3       593.7454                 G
#> 4      1455.9063                 G
#> 5       653.0310                 G
#> 6      1494.7384                 G
#> 7      1427.8939                 G
#> 8      1195.0497                 G
#> 9       724.1737                 G
#> 10      948.5697                 G
#> 11     1270.3424                 G
#> 12     1053.0606                 G
#> 13      498.5327                 G
#> 14      686.5273                 G
#> 15      689.5509                 G
#> 
#> 
#> $garamba__loxodonta_africana
#> $garamba__loxodonta_africana$location
#> [1] "Garamba"
#> 
#> $garamba__loxodonta_africana$species
#> [1] "Loxodonta africana"
#> 
#> $garamba__loxodonta_africana$years
#>  [1] 1976 1983 1984 1986 1991 1993 1995 1998 2000 2002 2003 2004 2012 2014 2017
#> 
#> $garamba__loxodonta_africana$n_years
#> [1] 15
#> 
#> $garamba__loxodonta_africana$stat_methods
#> [1] "S" "T"
#> 
#> $garamba__loxodonta_africana$field_methods
#> [1] "A"
#> 
#> $garamba__loxodonta_africana$pref_field_method
#> [1] "A"
#> 
#> $garamba__loxodonta_africana$conversion_fact
#> [1] 0.659
#> 
#> $garamba__loxodonta_africana$data_original
#>    location            species year stat_method field_method counts_orig
#> 1   Garamba Loxodonta africana 1976           S            A       22670
#> 2   Garamba Loxodonta africana 1983           S            A        7742
#> 3   Garamba Loxodonta africana 1984           S            A        3300
#> 4   Garamba Loxodonta africana 1986           S            A        4169
#> 5   Garamba Loxodonta africana 1991           S            A        7385
#> 6   Garamba Loxodonta africana 1993           S            A        8768
#> 7   Garamba Loxodonta africana 1995           S            A       11175
#> 8   Garamba Loxodonta africana 1998           S            A        5872
#> 9   Garamba Loxodonta africana 2000           S            A        6022
#> 10  Garamba Loxodonta africana 2002           S            A        5983
#> 11  Garamba Loxodonta africana 2003           S            A        6948
#> 12  Garamba Loxodonta africana 2004           S            A        6354
#> 13  Garamba Loxodonta africana 2012           T            A        1668
#> 14  Garamba Loxodonta africana 2014           T            A        1718
#> 15  Garamba Loxodonta africana 2017           T            A        1191
#>    lower_ci_orig upper_ci_orig
#> 1          10880         34460
#> 2           4052         11432
#> 3           2791          3809
#> 4           1672          6666
#> 5           4464         10306
#> 6           7184         10352
#> 7           7515         14835
#> 8           4533          7211
#> 9           4976          7068
#> 10          4799          7167
#> 11          4953          8943
#> 12          4273          8435
#> 13            NA            NA
#> 14            NA            NA
#> 15            NA            NA
#> 
#> $garamba__loxodonta_africana$data_converted
#>    location            species year stat_method field_method counts_conv
#> 1   Garamba Loxodonta africana 1976           S            A       22670
#> 2   Garamba Loxodonta africana 1983           S            A        7742
#> 3   Garamba Loxodonta africana 1984           S            A        3300
#> 4   Garamba Loxodonta africana 1986           S            A        4169
#> 5   Garamba Loxodonta africana 1991           S            A        7385
#> 6   Garamba Loxodonta africana 1993           S            A        8768
#> 7   Garamba Loxodonta africana 1995           S            A       11175
#> 8   Garamba Loxodonta africana 1998           S            A        5872
#> 9   Garamba Loxodonta africana 2000           S            A        6022
#> 10  Garamba Loxodonta africana 2002           S            A        5983
#> 11  Garamba Loxodonta africana 2003           S            A        6948
#> 12  Garamba Loxodonta africana 2004           S            A        6354
#> 13  Garamba Loxodonta africana 2012           T            A        1668
#> 14  Garamba Loxodonta africana 2014           T            A        1718
#> 15  Garamba Loxodonta africana 2017           T            A        1191
#>    lower_ci_conv upper_ci_conv field_method_conv
#> 1       10880.00       34460.0                 A
#> 2        4052.00       11432.0                 A
#> 3        2791.00        3809.0                 A
#> 4        1672.00        6666.0                 A
#> 5        4464.00       10306.0                 A
#> 6        7184.00       10352.0                 A
#> 7        7515.00       14835.0                 A
#> 8        4533.00        7211.0                 A
#> 9        4976.00        7068.0                 A
#> 10       4799.00        7167.0                 A
#> 11       4953.00        8943.0                 A
#> 12       4273.00        8435.0                 A
#> 13       1584.60        2001.6                 A
#> 14       1632.10        2061.6                 A
#> 15       1131.45        1429.2                 A
#> 
#> 
#> $garamba__ourebia_ourebi
#> $garamba__ourebia_ourebi$location
#> [1] "Garamba"
#> 
#> $garamba__ourebia_ourebi$species
#> [1] "Ourebia ourebi"
#> 
#> $garamba__ourebia_ourebi$years
#>  [1] 1976 1983 1984 1986 1991 1993 1995 1998 2000 2002 2003 2004
#> 
#> $garamba__ourebia_ourebi$n_years
#> [1] 12
#> 
#> $garamba__ourebia_ourebi$stat_methods
#> [1] "S"
#> 
#> $garamba__ourebia_ourebi$field_methods
#> [1] "A"
#> 
#> $garamba__ourebia_ourebi$pref_field_method
#> [1] "G"
#> 
#> $garamba__ourebia_ourebi$conversion_fact
#> [1] 6.747
#> 
#> $garamba__ourebia_ourebi$data_original
#>    location        species year stat_method field_method counts_orig
#> 1   Garamba Ourebia ourebi 1976           S            A         380
#> 2   Garamba Ourebia ourebi 1983           S            A         234
#> 3   Garamba Ourebia ourebi 1984           S            A         153
#> 4   Garamba Ourebia ourebi 1986           S            A         230
#> 5   Garamba Ourebia ourebi 1991           S            A          39
#> 6   Garamba Ourebia ourebi 1993           S            A          90
#> 7   Garamba Ourebia ourebi 1995           S            A           0
#> 8   Garamba Ourebia ourebi 1998           S            A          89
#> 9   Garamba Ourebia ourebi 2000           S            A          41
#> 10  Garamba Ourebia ourebi 2002           S            A          39
#> 11  Garamba Ourebia ourebi 2003           S            A          58
#> 12  Garamba Ourebia ourebi 2004           S            A         312
#>    lower_ci_orig upper_ci_orig
#> 1            230           530
#> 2             82           386
#> 3             50           256
#> 4            140           320
#> 5             12            66
#> 6             33           147
#> 7              0             0
#> 8             50           128
#> 9             22            60
#> 10             8            70
#> 11            37            79
#> 12           197           427
#> 
#> $garamba__ourebia_ourebi$data_converted
#>    location        species year stat_method field_method counts_conv
#> 1   Garamba Ourebia ourebi 1976           S            A   56.321328
#> 2   Garamba Ourebia ourebi 1983           S            A   34.682081
#> 3   Garamba Ourebia ourebi 1984           S            A   22.676745
#> 4   Garamba Ourebia ourebi 1986           S            A   34.089225
#> 5   Garamba Ourebia ourebi 1991           S            A    5.780347
#> 6   Garamba Ourebia ourebi 1993           S            A   13.339262
#> 7   Garamba Ourebia ourebi 1995           S            A    0.000000
#> 8   Garamba Ourebia ourebi 1998           S            A   13.191048
#> 9   Garamba Ourebia ourebi 2000           S            A    6.076775
#> 10  Garamba Ourebia ourebi 2002           S            A    5.780347
#> 11  Garamba Ourebia ourebi 2003           S            A    8.596413
#> 12  Garamba Ourebia ourebi 2004           S            A   46.242775
#>    lower_ci_conv upper_ci_conv field_method_conv
#> 1      34.089225     78.553431                 G
#> 2      12.153550     57.210612                 G
#> 3       7.410701     37.942789                 G
#> 4      20.749963     47.428487                 G
#> 5       1.778568      9.782125                 G
#> 6       4.891063     21.787461                 G
#> 7       0.000000      0.000000                 G
#> 8       7.410701     18.971395                 G
#> 9       3.260708      8.892841                 G
#> 10      1.185712     10.374981                 G
#> 11      5.483919     11.708908                 G
#> 12     29.198162     63.287387                 G
#> 
#> 
#> $garamba__redunca_redunca
#> $garamba__redunca_redunca$location
#> [1] "Garamba"
#> 
#> $garamba__redunca_redunca$species
#> [1] "Redunca redunca"
#> 
#> $garamba__redunca_redunca$years
#>  [1] 1976 1983 1984 1986 1991 1993 1995 2000 2002 2003 2004 2017
#> 
#> $garamba__redunca_redunca$n_years
#> [1] 12
#> 
#> $garamba__redunca_redunca$stat_methods
#> [1] "S" "T"
#> 
#> $garamba__redunca_redunca$field_methods
#> [1] "A"
#> 
#> $garamba__redunca_redunca$pref_field_method
#> [1] "G"
#> 
#> $garamba__redunca_redunca$conversion_fact
#> [1] 6.747
#> 
#> $garamba__redunca_redunca$data_original
#>    location         species year stat_method field_method counts_orig
#> 1   Garamba Redunca redunca 1976           S            A         640
#> 2   Garamba Redunca redunca 1983           S            A         183
#> 3   Garamba Redunca redunca 1984           S            A         175
#> 4   Garamba Redunca redunca 1986           S            A         328
#> 5   Garamba Redunca redunca 1991           S            A          57
#> 6   Garamba Redunca redunca 1993           S            A          75
#> 7   Garamba Redunca redunca 1995           S            A         271
#> 8   Garamba Redunca redunca 2000           S            A         103
#> 9   Garamba Redunca redunca 2002           S            A         193
#> 10  Garamba Redunca redunca 2003           S            A          54
#> 11  Garamba Redunca redunca 2004           S            A          80
#> 12  Garamba Redunca redunca 2017           T            A          81
#>    lower_ci_orig upper_ci_orig
#> 1            330           950
#> 2             76           290
#> 3             74           276
#> 4            192           464
#> 5             29            85
#> 6             75           166
#> 7             91           451
#> 8             59           147
#> 9             90           296
#> 10            28            80
#> 11            49           111
#> 12            NA            NA
#> 
#> $garamba__redunca_redunca$data_converted
#>    location         species year stat_method field_method counts_conv
#> 1   Garamba Redunca redunca 1976           S            A   94.856973
#> 2   Garamba Redunca redunca 1983           S            A   27.123166
#> 3   Garamba Redunca redunca 1984           S            A   25.937454
#> 4   Garamba Redunca redunca 1986           S            A   48.614199
#> 5   Garamba Redunca redunca 1991           S            A    8.448199
#> 6   Garamba Redunca redunca 1993           S            A   11.116052
#> 7   Garamba Redunca redunca 1995           S            A   40.166000
#> 8   Garamba Redunca redunca 2000           S            A   15.266044
#> 9   Garamba Redunca redunca 2002           S            A   28.605306
#> 10  Garamba Redunca redunca 2003           S            A    8.003557
#> 11  Garamba Redunca redunca 2004           S            A   11.857122
#> 12  Garamba Redunca redunca 2017           T            A   12.005336
#>    lower_ci_conv upper_ci_conv field_method_conv
#> 1      48.910627     140.80332                 G
#> 2      11.264266      42.98207                 G
#> 3      10.967838      40.90707                 G
#> 4      28.457092      68.77131                 G
#> 5       4.298207      12.59819                 G
#> 6      11.116052      24.60353                 G
#> 7      13.487476      66.84452                 G
#> 8       8.744627      21.78746                 G
#> 9      13.339262      43.87135                 G
#> 10      4.149993      11.85712                 G
#> 11      7.262487      16.45176                 G
#> 12     11.405069      14.40640                 G
#> 
#> 
#> $garamba__syncerus_caffer
#> $garamba__syncerus_caffer$location
#> [1] "Garamba"
#> 
#> $garamba__syncerus_caffer$species
#> [1] "Syncerus caffer"
#> 
#> $garamba__syncerus_caffer$years
#>  [1] 1976 1983 1984 1986 1991 1993 1995 1998 2000 2002 2003 2004 2012 2014 2017
#> 
#> $garamba__syncerus_caffer$n_years
#> [1] 15
#> 
#> $garamba__syncerus_caffer$stat_methods
#> [1] "S" "T"
#> 
#> $garamba__syncerus_caffer$field_methods
#> [1] "A"
#> 
#> $garamba__syncerus_caffer$pref_field_method
#> [1] "A"
#> 
#> $garamba__syncerus_caffer$conversion_fact
#> [1] 0.561
#> 
#> $garamba__syncerus_caffer$data_original
#>    location         species year stat_method field_method counts_orig
#> 1   Garamba Syncerus caffer 1976           S            A       53000
#> 2   Garamba Syncerus caffer 1983           S            A       53312
#> 3   Garamba Syncerus caffer 1984           S            A       48284
#> 4   Garamba Syncerus caffer 1986           S            A       29293
#> 5   Garamba Syncerus caffer 1991           S            A       33910
#> 6   Garamba Syncerus caffer 1993           S            A       30555
#> 7   Garamba Syncerus caffer 1995           S            A       25242
#> 8   Garamba Syncerus caffer 1998           S            A        7772
#> 9   Garamba Syncerus caffer 2000           S            A       13115
#> 10  Garamba Syncerus caffer 2002           S            A       13281
#> 11  Garamba Syncerus caffer 2003           S            A       14460
#> 12  Garamba Syncerus caffer 2004           S            A       12230
#> 13  Garamba Syncerus caffer 2012           T            A        5587
#> 14  Garamba Syncerus caffer 2014           T            A        7149
#> 15  Garamba Syncerus caffer 2017           T            A        6728
#>    lower_ci_orig upper_ci_orig
#> 1          10640         95360
#> 2          36352         70272
#> 3          42302         54266
#> 4          21537         37049
#> 5          24436         43384
#> 6          11757         49353
#> 7           8322         42162
#> 8           5709          9835
#> 9          10049         16181
#> 10          9351         17211
#> 11         10229         18691
#> 12          7513         16947
#> 13            NA            NA
#> 14            NA            NA
#> 15            NA            NA
#> 
#> $garamba__syncerus_caffer$data_converted
#>    location         species year stat_method field_method counts_conv
#> 1   Garamba Syncerus caffer 1976           S            A       53000
#> 2   Garamba Syncerus caffer 1983           S            A       53312
#> 3   Garamba Syncerus caffer 1984           S            A       48284
#> 4   Garamba Syncerus caffer 1986           S            A       29293
#> 5   Garamba Syncerus caffer 1991           S            A       33910
#> 6   Garamba Syncerus caffer 1993           S            A       30555
#> 7   Garamba Syncerus caffer 1995           S            A       25242
#> 8   Garamba Syncerus caffer 1998           S            A        7772
#> 9   Garamba Syncerus caffer 2000           S            A       13115
#> 10  Garamba Syncerus caffer 2002           S            A       13281
#> 11  Garamba Syncerus caffer 2003           S            A       14460
#> 12  Garamba Syncerus caffer 2004           S            A       12230
#> 13  Garamba Syncerus caffer 2012           T            A        5587
#> 14  Garamba Syncerus caffer 2014           T            A        7149
#> 15  Garamba Syncerus caffer 2017           T            A        6728
#>    lower_ci_conv upper_ci_conv field_method_conv
#> 1       10640.00       95360.0                 A
#> 2       36352.00       70272.0                 A
#> 3       42302.00       54266.0                 A
#> 4       21537.00       37049.0                 A
#> 5       24436.00       43384.0                 A
#> 6       11757.00       49353.0                 A
#> 7        8322.00       42162.0                 A
#> 8        5709.00        9835.0                 A
#> 9       10049.00       16181.0                 A
#> 10       9351.00       17211.0                 A
#> 11      10229.00       18691.0                 A
#> 12       7513.00       16947.0                 A
#> 13       5307.65        6704.4                 A
#> 14       6791.55        8578.8                 A
#> 15       6391.60        8073.6                 A
#> 
#> 
#> $garamba__tragelaphus_scriptus
#> $garamba__tragelaphus_scriptus$location
#> [1] "Garamba"
#> 
#> $garamba__tragelaphus_scriptus$species
#> [1] "Tragelaphus scriptus"
#> 
#> $garamba__tragelaphus_scriptus$years
#>  [1] 1976 1983 1984 1986 1991 1993 1995 1998 2000 2002 2003 2004
#> 
#> $garamba__tragelaphus_scriptus$n_years
#> [1] 12
#> 
#> $garamba__tragelaphus_scriptus$stat_methods
#> [1] "S"
#> 
#> $garamba__tragelaphus_scriptus$field_methods
#> [1] "A"
#> 
#> $garamba__tragelaphus_scriptus$pref_field_method
#> [1] "G"
#> 
#> $garamba__tragelaphus_scriptus$conversion_fact
#> [1] 6.747
#> 
#> $garamba__tragelaphus_scriptus$data_original
#>    location              species year stat_method field_method counts_orig
#> 1   Garamba Tragelaphus scriptus 1976           S            A         420
#> 2   Garamba Tragelaphus scriptus 1983           S            A          51
#> 3   Garamba Tragelaphus scriptus 1984           S            A         131
#> 4   Garamba Tragelaphus scriptus 1986           S            A         149
#> 5   Garamba Tragelaphus scriptus 1991           S            A          90
#> 6   Garamba Tragelaphus scriptus 1993           S            A          75
#> 7   Garamba Tragelaphus scriptus 1995           S            A          67
#> 8   Garamba Tragelaphus scriptus 1998           S            A         187
#> 9   Garamba Tragelaphus scriptus 2000           S            A          89
#> 10  Garamba Tragelaphus scriptus 2002           S            A         143
#> 11  Garamba Tragelaphus scriptus 2003           S            A          62
#> 12  Garamba Tragelaphus scriptus 2004           S            A         107
#>    lower_ci_orig upper_ci_orig
#> 1            290           550
#> 2              6            96
#> 3             74           188
#> 4             28           270
#> 5             45           135
#> 6             22           128
#> 7             24           110
#> 8            141           233
#> 9             58           120
#> 10            90           196
#> 11            28            96
#> 12            38           176
#> 
#> $garamba__tragelaphus_scriptus$data_converted
#>    location              species year stat_method field_method counts_conv
#> 1   Garamba Tragelaphus scriptus 1976           S            A   62.249889
#> 2   Garamba Tragelaphus scriptus 1983           S            A    7.558915
#> 3   Garamba Tragelaphus scriptus 1984           S            A   19.416037
#> 4   Garamba Tragelaphus scriptus 1986           S            A   22.083889
#> 5   Garamba Tragelaphus scriptus 1991           S            A   13.339262
#> 6   Garamba Tragelaphus scriptus 1993           S            A   11.116052
#> 7   Garamba Tragelaphus scriptus 1995           S            A    9.930339
#> 8   Garamba Tragelaphus scriptus 1998           S            A   27.716022
#> 9   Garamba Tragelaphus scriptus 2000           S            A   13.191048
#> 10  Garamba Tragelaphus scriptus 2002           S            A   21.194605
#> 11  Garamba Tragelaphus scriptus 2003           S            A    9.189269
#> 12  Garamba Tragelaphus scriptus 2004           S            A   15.858900
#>    lower_ci_conv upper_ci_conv field_method_conv
#> 1     42.9820661      81.51771                 G
#> 2      0.8892841      14.22855                 G
#> 3     10.9678376      27.86424                 G
#> 4      4.1499926      40.01779                 G
#> 5      6.6696309      20.00889                 G
#> 6      3.2607085      18.97139                 G
#> 7      3.5571365      16.30354                 G
#> 8     20.8981770      34.53387                 G
#> 9      8.5964132      17.78568                 G
#> 10    13.3392619      29.04995                 G
#> 11     4.1499926      14.22855                 G
#> 12     5.6321328      26.08567                 G
#> 
#> 

## Get data for Alcelaphus buselaphus at Garamba only ----
find_series(data = garamba_formatted, species = "Alcelaphus buselaphus", 
            location = "Garamba")

#> ✔ Found 1 series with 'Alcelaphus buselaphus' and 'Garamba'.

#> $garamba__alcelaphus_buselaphus
#> $garamba__alcelaphus_buselaphus$location
#> [1] "Garamba"
#> 
#> $garamba__alcelaphus_buselaphus$species
#> [1] "Alcelaphus buselaphus"
#> 
#> $garamba__alcelaphus_buselaphus$years
#>  [1] 1976 1983 1984 1986 1991 1993 1995 1998 2000 2002 2003 2004 2012 2014 2017
#> 
#> $garamba__alcelaphus_buselaphus$n_years
#> [1] 15
#> 
#> $garamba__alcelaphus_buselaphus$stat_methods
#> [1] "S" "T"
#> 
#> $garamba__alcelaphus_buselaphus$field_methods
#> [1] "A"
#> 
#> $garamba__alcelaphus_buselaphus$pref_field_method
#> [1] "G"
#> 
#> $garamba__alcelaphus_buselaphus$conversion_fact
#> [1] 2.302
#> 
#> $garamba__alcelaphus_buselaphus$data_original
#>    location               species year stat_method field_method counts_orig
#> 1   Garamba Alcelaphus buselaphus 1976           S            A        7750
#> 2   Garamba Alcelaphus buselaphus 1983           S            A        1932
#> 3   Garamba Alcelaphus buselaphus 1984           S            A        1224
#> 4   Garamba Alcelaphus buselaphus 1986           S            A        1705
#> 5   Garamba Alcelaphus buselaphus 1991           S            A         987
#> 6   Garamba Alcelaphus buselaphus 1993           S            A        3444
#> 7   Garamba Alcelaphus buselaphus 1995           S            A        2819
#> 8   Garamba Alcelaphus buselaphus 1998           S            A        1685
#> 9   Garamba Alcelaphus buselaphus 2000           S            A        1169
#> 10  Garamba Alcelaphus buselaphus 2002           S            A        1139
#> 11  Garamba Alcelaphus buselaphus 2003           S            A        1595
#> 12  Garamba Alcelaphus buselaphus 2004           S            A        1204
#> 13  Garamba Alcelaphus buselaphus 2012           T            A         552
#> 14  Garamba Alcelaphus buselaphus 2014           T            A         698
#> 15  Garamba Alcelaphus buselaphus 2017           T            A        1051
#>    lower_ci_orig upper_ci_orig
#> 1           6280          9220
#> 2           1120          2744
#> 3            782          1666
#> 4           1116          2294
#> 5            663          1311
#> 6           1290          5598
#> 7           1620          4018
#> 8           1287          2083
#> 9            945          1393
#> 10           907          1371
#> 11          1142          2048
#> 12           811          1597
#> 13            NA            NA
#> 14            NA            NA
#> 15            NA            NA
#> 
#> $garamba__alcelaphus_buselaphus$data_converted
#>    location               species year stat_method field_method counts_conv
#> 1   Garamba Alcelaphus buselaphus 1976           S            A   3366.6377
#> 2   Garamba Alcelaphus buselaphus 1983           S            A    839.2702
#> 3   Garamba Alcelaphus buselaphus 1984           S            A    531.7116
#> 4   Garamba Alcelaphus buselaphus 1986           S            A    740.6603
#> 5   Garamba Alcelaphus buselaphus 1991           S            A    428.7576
#> 6   Garamba Alcelaphus buselaphus 1993           S            A   1496.0904
#> 7   Garamba Alcelaphus buselaphus 1995           S            A   1224.5873
#> 8   Garamba Alcelaphus buselaphus 1998           S            A    731.9722
#> 9   Garamba Alcelaphus buselaphus 2000           S            A    507.8193
#> 10  Garamba Alcelaphus buselaphus 2002           S            A    494.7871
#> 11  Garamba Alcelaphus buselaphus 2003           S            A    692.8758
#> 12  Garamba Alcelaphus buselaphus 2004           S            A    523.0235
#> 13  Garamba Alcelaphus buselaphus 2012           T            A    239.7915
#> 14  Garamba Alcelaphus buselaphus 2014           T            A    303.2146
#> 15  Garamba Alcelaphus buselaphus 2017           T            A    456.5595
#>    lower_ci_conv upper_ci_conv field_method_conv
#> 1      2728.0626     4005.2129                 G
#> 2       486.5334     1192.0070                 G
#> 3       339.7046      723.7185                 G
#> 4       484.7958      996.5248                 G
#> 5       288.0104      569.5048                 G
#> 6       560.3823     2431.7984                 G
#> 7       703.7359     1745.4387                 G
#> 8       559.0791      904.8653                 G
#> 9       410.5126      605.1260                 G
#> 10      394.0052      595.5691                 G
#> 11      496.0904      889.6612                 G
#> 12      352.3023      693.7446                 G
#> 13      227.8019      287.7498                 G
#> 14      288.0539      363.8575                 G
#> 15      433.7315      547.8714                 G
#> 
#>