Import a list of BUGS outputs previously exported

This function imports a list of BUGS outputs previously exported by fit_trend(). Users can import one, several, or all models.

read_bugs(series = NULL, path = ".")

Arguments

series

a vector of character strings. One or several count series names. If NULL (default) BUGS outputs for all count series will be imported. Users can run list_series() to get the correct spelling of count series names.

path

a character string. The directory in which BUGS outputs have been saved by the function fit_trend().

Value

An n-element list (where n is the number of count series). See fit_trend() for further information.

Examples

## Load Garamba raw dataset ----
file_path <- system.file("extdata", "garamba_survey.csv", 
                         package = "popbayes")
                         
garamba <- read.csv(file = file_path)

## Create temporary folder ----
temp_path <- tempdir()

## Format dataset ----
garamba_formatted <- popbayes::format_data(
  data              = garamba, 
  path              = temp_path,
  field_method      = "field_method",
  pref_field_method = "pref_field_method",
  conversion_A2G    = "conversion_A2G",
  rmax              = "rmax")
#> ✔ Detecting 10 count series

## Select one serie ----
a_buselaphus <- popbayes::filter_series(garamba_formatted, 
                                        location = "Garamba",
                                        species  = "Alcelaphus buselaphus")
#> ✔ Found 1 series with 'Alcelaphus buselaphus' and 'Garamba'.
# \donttest{
## Fit population trends (requires JAGS) ----
a_buselaphus_mod <- popbayes::fit_trend(a_buselaphus, path = temp_path)
#> Compiling data graph
#>    Resolving undeclared variables
#>    Allocating nodes
#>    Initializing
#>    Reading data back into data table
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 15
#>    Unobserved stochastic nodes: 15
#>    Total graph size: 227
#> 
#> Initializing model
#> 

## Import BUGS outputs for one count series ----
popbayes::read_bugs(series = "garamba__alcelaphus_buselaphus", 
                    path   = temp_path)
#> $garamba__alcelaphus_buselaphus
#> Inference for Bugs model at "model.bug", fit using jags,
#>  2 chains, each with 50000 iterations (first 10000 discarded), n.thin = 3
#>  n.sims = 26666 iterations saved
#>            mu.vect  sd.vect      2.5%       25%       50%       75%     97.5%
#> N[1]     17644.626 1726.824 14256.716 16489.154 17642.636 18814.527 20994.297
#> N[2]      3877.981  490.188  2953.358  3541.194  3865.598  4200.203  4880.474
#> N[3]      3330.346  382.033  2591.290  3071.180  3327.413  3585.150  4087.282
#> N[4]      3014.734  431.065  2213.189  2720.792  2995.974  3289.427  3915.020
#> N[5]      2665.383  326.463  2025.653  2443.867  2663.573  2883.768  3312.573
#> N[6]      3218.381  555.064  2237.280  2829.120  3176.490  3572.799  4414.514
#> N[7]      3750.988  684.345  2543.417  3268.160  3702.249  4179.508  5251.142
#> N[8]      3477.335  367.061  2779.323  3222.953  3472.967  3723.224  4212.542
#> N[9]      2887.311  205.893  2492.916  2746.344  2884.853  3026.146  3299.115
#> N[10]     2750.307  206.133  2349.118  2609.770  2749.540  2888.226  3158.582
#> N[11]     2756.392  257.879  2265.064  2579.970  2753.702  2926.678  3277.484
#> N[12]     2628.831  298.935  2071.159  2423.681  2620.733  2826.193  3237.262
#> N[13]     1311.296   73.749  1168.581  1261.149  1310.440  1360.974  1456.460
#> N[14]     1570.535   89.479  1397.685  1509.945  1570.101  1631.334  1746.928
#> N[15]     2391.464  151.535  2094.955  2288.883  2391.209  2494.181  2687.239
#> meanr       -0.049    0.003    -0.054    -0.051    -0.049    -0.047    -0.043
#> r[1]        -0.217    0.022    -0.261    -0.232    -0.217    -0.202    -0.175
#> r[2]        -0.151    0.069    -0.286    -0.198    -0.151    -0.104    -0.015
#> r[3]        -0.052    0.060    -0.169    -0.093    -0.052    -0.011     0.067
#> r[4]        -0.024    0.034    -0.090    -0.047    -0.024    -0.002     0.041
#> r[5]         0.091    0.064    -0.038     0.048     0.091     0.134     0.219
#> r[6]         0.076    0.058    -0.038     0.037     0.076     0.114     0.190
#> r[7]        -0.022    0.052    -0.124    -0.057    -0.022     0.013     0.082
#> r[8]        -0.091    0.049    -0.186    -0.125    -0.092    -0.058     0.007
#> r[9]        -0.024    0.044    -0.111    -0.054    -0.024     0.005     0.060
#> r[10]        0.001    0.066    -0.130    -0.044     0.002     0.046     0.126
#> r[11]       -0.049    0.066    -0.179    -0.093    -0.049    -0.005     0.079
#> r[12]       -0.086    0.015    -0.116    -0.097    -0.087    -0.076    -0.055
#> r[13]        0.090    0.036     0.020     0.066     0.090     0.114     0.160
#> r[14]        0.140    0.027     0.086     0.122     0.140     0.158     0.193
#> sdr          0.112    0.009     0.094     0.106     0.111     0.118     0.131
#> vrrmax       0.460    0.038     0.389     0.435     0.459     0.485     0.538
#> deviance   236.085    5.108   227.487   232.428   235.619   239.192   247.393
#>           Rhat n.eff
#> N[1]     1.001  3200
#> N[2]     1.001 17000
#> N[3]     1.001 27000
#> N[4]     1.001 27000
#> N[5]     1.001 27000
#> N[6]     1.001 27000
#> N[7]     1.001  5700
#> N[8]     1.001  4500
#> N[9]     1.001 27000
#> N[10]    1.001 27000
#> N[11]    1.001 24000
#> N[12]    1.001 27000
#> N[13]    1.001 27000
#> N[14]    1.001 27000
#> N[15]    1.001 27000
#> meanr    1.001  3700
#> r[1]     1.001  3200
#> r[2]     1.001  7000
#> r[3]     1.001 27000
#> r[4]     1.001 27000
#> r[5]     1.001 14000
#> r[6]     1.001  6300
#> r[7]     1.001 27000
#> r[8]     1.001  5100
#> r[9]     1.001 27000
#> r[10]    1.001 16000
#> r[11]    1.001 27000
#> r[12]    1.001 16000
#> r[13]    1.001 27000
#> r[14]    1.001 27000
#> sdr      1.001  9600
#> vrrmax   1.001  9600
#> deviance 1.001 27000
#> 
#> For each parameter, n.eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor (at convergence, Rhat=1).
#> 
#> DIC info (using the rule, pD = var(deviance)/2)
#> pD = 13.0 and DIC = 249.1
#> DIC is an estimate of expected predictive error (lower deviance is better).
#> 

## Import BUGS outputs for all count series ----
popbayes::read_bugs(path = temp_path)
#> $garamba__alcelaphus_buselaphus
#> Inference for Bugs model at "model.bug", fit using jags,
#>  2 chains, each with 50000 iterations (first 10000 discarded), n.thin = 3
#>  n.sims = 26666 iterations saved
#>            mu.vect  sd.vect      2.5%       25%       50%       75%     97.5%
#> N[1]     17644.626 1726.824 14256.716 16489.154 17642.636 18814.527 20994.297
#> N[2]      3877.981  490.188  2953.358  3541.194  3865.598  4200.203  4880.474
#> N[3]      3330.346  382.033  2591.290  3071.180  3327.413  3585.150  4087.282
#> N[4]      3014.734  431.065  2213.189  2720.792  2995.974  3289.427  3915.020
#> N[5]      2665.383  326.463  2025.653  2443.867  2663.573  2883.768  3312.573
#> N[6]      3218.381  555.064  2237.280  2829.120  3176.490  3572.799  4414.514
#> N[7]      3750.988  684.345  2543.417  3268.160  3702.249  4179.508  5251.142
#> N[8]      3477.335  367.061  2779.323  3222.953  3472.967  3723.224  4212.542
#> N[9]      2887.311  205.893  2492.916  2746.344  2884.853  3026.146  3299.115
#> N[10]     2750.307  206.133  2349.118  2609.770  2749.540  2888.226  3158.582
#> N[11]     2756.392  257.879  2265.064  2579.970  2753.702  2926.678  3277.484
#> N[12]     2628.831  298.935  2071.159  2423.681  2620.733  2826.193  3237.262
#> N[13]     1311.296   73.749  1168.581  1261.149  1310.440  1360.974  1456.460
#> N[14]     1570.535   89.479  1397.685  1509.945  1570.101  1631.334  1746.928
#> N[15]     2391.464  151.535  2094.955  2288.883  2391.209  2494.181  2687.239
#> meanr       -0.049    0.003    -0.054    -0.051    -0.049    -0.047    -0.043
#> r[1]        -0.217    0.022    -0.261    -0.232    -0.217    -0.202    -0.175
#> r[2]        -0.151    0.069    -0.286    -0.198    -0.151    -0.104    -0.015
#> r[3]        -0.052    0.060    -0.169    -0.093    -0.052    -0.011     0.067
#> r[4]        -0.024    0.034    -0.090    -0.047    -0.024    -0.002     0.041
#> r[5]         0.091    0.064    -0.038     0.048     0.091     0.134     0.219
#> r[6]         0.076    0.058    -0.038     0.037     0.076     0.114     0.190
#> r[7]        -0.022    0.052    -0.124    -0.057    -0.022     0.013     0.082
#> r[8]        -0.091    0.049    -0.186    -0.125    -0.092    -0.058     0.007
#> r[9]        -0.024    0.044    -0.111    -0.054    -0.024     0.005     0.060
#> r[10]        0.001    0.066    -0.130    -0.044     0.002     0.046     0.126
#> r[11]       -0.049    0.066    -0.179    -0.093    -0.049    -0.005     0.079
#> r[12]       -0.086    0.015    -0.116    -0.097    -0.087    -0.076    -0.055
#> r[13]        0.090    0.036     0.020     0.066     0.090     0.114     0.160
#> r[14]        0.140    0.027     0.086     0.122     0.140     0.158     0.193
#> sdr          0.112    0.009     0.094     0.106     0.111     0.118     0.131
#> vrrmax       0.460    0.038     0.389     0.435     0.459     0.485     0.538
#> deviance   236.085    5.108   227.487   232.428   235.619   239.192   247.393
#>           Rhat n.eff
#> N[1]     1.001  3200
#> N[2]     1.001 17000
#> N[3]     1.001 27000
#> N[4]     1.001 27000
#> N[5]     1.001 27000
#> N[6]     1.001 27000
#> N[7]     1.001  5700
#> N[8]     1.001  4500
#> N[9]     1.001 27000
#> N[10]    1.001 27000
#> N[11]    1.001 24000
#> N[12]    1.001 27000
#> N[13]    1.001 27000
#> N[14]    1.001 27000
#> N[15]    1.001 27000
#> meanr    1.001  3700
#> r[1]     1.001  3200
#> r[2]     1.001  7000
#> r[3]     1.001 27000
#> r[4]     1.001 27000
#> r[5]     1.001 14000
#> r[6]     1.001  6300
#> r[7]     1.001 27000
#> r[8]     1.001  5100
#> r[9]     1.001 27000
#> r[10]    1.001 16000
#> r[11]    1.001 27000
#> r[12]    1.001 16000
#> r[13]    1.001 27000
#> r[14]    1.001 27000
#> sdr      1.001  9600
#> vrrmax   1.001  9600
#> deviance 1.001 27000
#> 
#> For each parameter, n.eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor (at convergence, Rhat=1).
#> 
#> DIC info (using the rule, pD = var(deviance)/2)
#> pD = 13.0 and DIC = 249.1
#> DIC is an estimate of expected predictive error (lower deviance is better).
#> 
# }