Import a list of BUGS outputs previously exported

This function imports a list of BUGS outputs previously exported by fit_trend(). Users can import one, several, or all models.

read_bugs(series = NULL, path = ".")

Arguments

series

a vector of character strings. One or several count series names. If NULL (default) BUGS outputs for all count series will be imported. Users can run list_series() to get the correct spelling of count series names.

path

a character string. The directory in which BUGS outputs have been saved by the function fit_trend().

Value

An n-element list (where n is the number of count series). See fit_trend() for further information.

Examples

## Load Garamba raw dataset ----
file_path <- system.file("extdata", "garamba_survey.csv", 
                         package = "popbayes")
                         
garamba <- read.csv(file = file_path)

## Create temporary folder ----
temp_path <- tempdir()

## Format dataset ----
garamba_formatted <- popbayes::format_data(garamba, path = temp_path)
#> ✔ Conversion data found in 'data'.
#> ✔ Detecting 10 count series

## Select one serie ----
a_buselaphus <- popbayes::filter_series(garamba_formatted, 
                                        location = "Garamba",
                                        species  = "Alcelaphus buselaphus")
#> ✔ Found 1 series with 'Alcelaphus buselaphus' and 'Garamba'.
# \donttest{
## Fit population trends (requires JAGS) ----
a_buselaphus_mod <- popbayes::fit_trend(a_buselaphus, path = temp_path)
#> Compiling data graph
#>    Resolving undeclared variables
#>    Allocating nodes
#>    Initializing
#>    Reading data back into data table
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 15
#>    Unobserved stochastic nodes: 15
#>    Total graph size: 227
#> 
#> Initializing model
#> 

## Import BUGS outputs for one count series ----
popbayes::read_bugs(series = "garamba__alcelaphus_buselaphus", 
                    path   = temp_path)
#> $garamba__alcelaphus_buselaphus
#> Inference for Bugs model at "model.bug", fit using jags,
#>  2 chains, each with 50000 iterations (first 10000 discarded), n.thin = 3
#>  n.sims = 26666 iterations saved
#>            mu.vect  sd.vect      2.5%       25%       50%       75%     97.5%
#> N[1]     17653.374 1709.875 14293.416 16500.262 17670.895 18788.744 21013.284
#> N[2]      3877.572  507.906  2924.494  3526.800  3859.274  4215.326  4910.025
#> N[3]      3334.692  393.875  2565.487  3064.603  3334.835  3602.807  4108.442
#> N[4]      3035.278  431.005  2235.066  2739.158  3021.576  3316.903  3922.510
#> N[5]      2658.386  328.199  2022.471  2434.520  2652.234  2879.177  3310.742
#> N[6]      3182.195  534.831  2234.229  2805.697  3143.962  3523.175  4328.668
#> N[7]      3700.213  662.011  2556.281  3232.193  3643.074  4113.770  5130.826
#> N[8]      3472.790  368.289  2773.647  3218.876  3466.432  3714.664  4210.885
#> N[9]      2887.415  208.366  2484.422  2746.145  2885.442  3026.124  3302.663
#> N[10]     2751.235  207.891  2343.517  2611.320  2749.842  2889.875  3163.045
#> N[11]     2755.960  261.105  2254.244  2578.023  2754.211  2925.636  3278.103
#> N[12]     2626.435  301.551  2054.402  2421.871  2617.986  2826.383  3234.788
#> N[13]     1311.203   73.273  1168.945  1261.812  1310.886  1360.306  1456.269
#> N[14]     1570.938   90.413  1392.626  1510.284  1570.689  1631.773  1748.795
#> N[15]     2394.487  151.500  2098.620  2292.305  2393.208  2496.723  2692.175
#> meanr       -0.049    0.003    -0.054    -0.051    -0.049    -0.047    -0.043
#> r[1]        -0.217    0.022    -0.262    -0.232    -0.217    -0.202    -0.174
#> r[2]        -0.149    0.070    -0.287    -0.197    -0.149    -0.102    -0.013
#> r[3]        -0.049    0.061    -0.168    -0.089    -0.048    -0.008     0.071
#> r[4]        -0.026    0.034    -0.091    -0.049    -0.026    -0.003     0.040
#> r[5]         0.087    0.063    -0.040     0.044     0.087     0.130     0.213
#> r[6]         0.075    0.058    -0.040     0.036     0.074     0.113     0.188
#> r[7]        -0.018    0.052    -0.119    -0.053    -0.018     0.017     0.082
#> r[8]        -0.091    0.050    -0.187    -0.125    -0.091    -0.058     0.010
#> r[9]        -0.024    0.044    -0.112    -0.054    -0.024     0.006     0.061
#> r[10]        0.000    0.066    -0.132    -0.043     0.001     0.045     0.128
#> r[11]       -0.050    0.066    -0.181    -0.094    -0.050    -0.006     0.079
#> r[12]       -0.086    0.016    -0.116    -0.097    -0.087    -0.076    -0.054
#> r[13]        0.090    0.036     0.020     0.066     0.090     0.114     0.160
#> r[14]        0.140    0.027     0.086     0.122     0.141     0.159     0.193
#> sdr          0.112    0.009     0.095     0.105     0.111     0.118     0.131
#> vrrmax       0.459    0.038     0.389     0.433     0.458     0.484     0.538
#> deviance   236.268    5.160   227.455   232.659   235.815   239.378   247.695
#>           Rhat n.eff
#> N[1]     1.001  6600
#> N[2]     1.001 27000
#> N[3]     1.002  1900
#> N[4]     1.003   600
#> N[5]     1.001 27000
#> N[6]     1.001  7800
#> N[7]     1.001  4300
#> N[8]     1.001  4700
#> N[9]     1.001 17000
#> N[10]    1.001  5200
#> N[11]    1.001  3600
#> N[12]    1.002  2100
#> N[13]    1.001 24000
#> N[14]    1.001 15000
#> N[15]    1.001 27000
#> meanr    1.001  9300
#> r[1]     1.001  6300
#> r[2]     1.002  1300
#> r[3]     1.002  1500
#> r[4]     1.003   840
#> r[5]     1.001  4800
#> r[6]     1.001 21000
#> r[7]     1.002  1300
#> r[8]     1.002  2300
#> r[9]     1.001  3100
#> r[10]    1.001 15000
#> r[11]    1.001  5000
#> r[12]    1.002  1900
#> r[13]    1.001 27000
#> r[14]    1.001 27000
#> sdr      1.001  3000
#> vrrmax   1.001  3000
#> deviance 1.001 27000
#> 
#> For each parameter, n.eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor (at convergence, Rhat=1).
#> 
#> DIC info (using the rule, pD = var(deviance)/2)
#> pD = 13.3 and DIC = 249.6
#> DIC is an estimate of expected predictive error (lower deviance is better).
#> 

## Import BUGS outputs for all count series ----
popbayes::read_bugs(path = temp_path)
#> $garamba__alcelaphus_buselaphus
#> Inference for Bugs model at "model.bug", fit using jags,
#>  2 chains, each with 50000 iterations (first 10000 discarded), n.thin = 3
#>  n.sims = 26666 iterations saved
#>            mu.vect  sd.vect      2.5%       25%       50%       75%     97.5%
#> N[1]     17653.374 1709.875 14293.416 16500.262 17670.895 18788.744 21013.284
#> N[2]      3877.572  507.906  2924.494  3526.800  3859.274  4215.326  4910.025
#> N[3]      3334.692  393.875  2565.487  3064.603  3334.835  3602.807  4108.442
#> N[4]      3035.278  431.005  2235.066  2739.158  3021.576  3316.903  3922.510
#> N[5]      2658.386  328.199  2022.471  2434.520  2652.234  2879.177  3310.742
#> N[6]      3182.195  534.831  2234.229  2805.697  3143.962  3523.175  4328.668
#> N[7]      3700.213  662.011  2556.281  3232.193  3643.074  4113.770  5130.826
#> N[8]      3472.790  368.289  2773.647  3218.876  3466.432  3714.664  4210.885
#> N[9]      2887.415  208.366  2484.422  2746.145  2885.442  3026.124  3302.663
#> N[10]     2751.235  207.891  2343.517  2611.320  2749.842  2889.875  3163.045
#> N[11]     2755.960  261.105  2254.244  2578.023  2754.211  2925.636  3278.103
#> N[12]     2626.435  301.551  2054.402  2421.871  2617.986  2826.383  3234.788
#> N[13]     1311.203   73.273  1168.945  1261.812  1310.886  1360.306  1456.269
#> N[14]     1570.938   90.413  1392.626  1510.284  1570.689  1631.773  1748.795
#> N[15]     2394.487  151.500  2098.620  2292.305  2393.208  2496.723  2692.175
#> meanr       -0.049    0.003    -0.054    -0.051    -0.049    -0.047    -0.043
#> r[1]        -0.217    0.022    -0.262    -0.232    -0.217    -0.202    -0.174
#> r[2]        -0.149    0.070    -0.287    -0.197    -0.149    -0.102    -0.013
#> r[3]        -0.049    0.061    -0.168    -0.089    -0.048    -0.008     0.071
#> r[4]        -0.026    0.034    -0.091    -0.049    -0.026    -0.003     0.040
#> r[5]         0.087    0.063    -0.040     0.044     0.087     0.130     0.213
#> r[6]         0.075    0.058    -0.040     0.036     0.074     0.113     0.188
#> r[7]        -0.018    0.052    -0.119    -0.053    -0.018     0.017     0.082
#> r[8]        -0.091    0.050    -0.187    -0.125    -0.091    -0.058     0.010
#> r[9]        -0.024    0.044    -0.112    -0.054    -0.024     0.006     0.061
#> r[10]        0.000    0.066    -0.132    -0.043     0.001     0.045     0.128
#> r[11]       -0.050    0.066    -0.181    -0.094    -0.050    -0.006     0.079
#> r[12]       -0.086    0.016    -0.116    -0.097    -0.087    -0.076    -0.054
#> r[13]        0.090    0.036     0.020     0.066     0.090     0.114     0.160
#> r[14]        0.140    0.027     0.086     0.122     0.141     0.159     0.193
#> sdr          0.112    0.009     0.095     0.105     0.111     0.118     0.131
#> vrrmax       0.459    0.038     0.389     0.433     0.458     0.484     0.538
#> deviance   236.268    5.160   227.455   232.659   235.815   239.378   247.695
#>           Rhat n.eff
#> N[1]     1.001  6600
#> N[2]     1.001 27000
#> N[3]     1.002  1900
#> N[4]     1.003   600
#> N[5]     1.001 27000
#> N[6]     1.001  7800
#> N[7]     1.001  4300
#> N[8]     1.001  4700
#> N[9]     1.001 17000
#> N[10]    1.001  5200
#> N[11]    1.001  3600
#> N[12]    1.002  2100
#> N[13]    1.001 24000
#> N[14]    1.001 15000
#> N[15]    1.001 27000
#> meanr    1.001  9300
#> r[1]     1.001  6300
#> r[2]     1.002  1300
#> r[3]     1.002  1500
#> r[4]     1.003   840
#> r[5]     1.001  4800
#> r[6]     1.001 21000
#> r[7]     1.002  1300
#> r[8]     1.002  2300
#> r[9]     1.001  3100
#> r[10]    1.001 15000
#> r[11]    1.001  5000
#> r[12]    1.002  1900
#> r[13]    1.001 27000
#> r[14]    1.001 27000
#> sdr      1.001  3000
#> vrrmax   1.001  3000
#> deviance 1.001 27000
#> 
#> For each parameter, n.eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor (at convergence, Rhat=1).
#> 
#> DIC info (using the rule, pD = var(deviance)/2)
#> pD = 13.3 and DIC = 249.6
#> DIC is an estimate of expected predictive error (lower deviance is better).
#> 
# }