Import a list of BUGS outputs previously exported

This function imports a list of BUGS outputs previously exported by fit_trend(). Users can import one, several, or all models.

read_bugs(series = NULL, path = ".")

Arguments

series

a vector of character strings. One or several count series names. If NULL (default) BUGS outputs for all count series will be imported. Users can run list_series() to get the correct spelling of count series names.

path

a character string. The directory in which BUGS outputs have been saved by the function fit_trend().

Value

An n-element list (where n is the number of count series). See fit_trend() for further information.

Examples

## Load Garamba raw dataset ----
file_path <- system.file("extdata", "garamba_survey.csv", 
                         package = "popbayes")
                         
garamba <- read.csv(file = file_path)

## Create temporary folder ----
temp_path <- tempdir()

## Format dataset ----
garamba_formatted <- popbayes::format_data(
  data              = garamba, 
  path              = temp_path,
  field_method      = "field_method",
  pref_field_method = "pref_field_method",
  conversion_A2G    = "conversion_A2G",
  rmax              = "rmax")
#> ✔ Detecting 10 count series

## Select one serie ----
a_buselaphus <- popbayes::filter_series(garamba_formatted, 
                                        location = "Garamba",
                                        species  = "Alcelaphus buselaphus")
#> ✔ Found 1 series with 'Alcelaphus buselaphus' and 'Garamba'.
# \donttest{
## Fit population trends (requires JAGS) ----
a_buselaphus_mod <- popbayes::fit_trend(a_buselaphus, path = temp_path)
#> Compiling data graph
#>    Resolving undeclared variables
#>    Allocating nodes
#>    Initializing
#>    Reading data back into data table
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 15
#>    Unobserved stochastic nodes: 15
#>    Total graph size: 227
#> 
#> Initializing model
#> 

## Import BUGS outputs for one count series ----
popbayes::read_bugs(series = "garamba__alcelaphus_buselaphus", 
                    path   = temp_path)
#> $garamba__alcelaphus_buselaphus
#> Inference for Bugs model at "model.bug", fit using jags,
#>  2 chains, each with 50000 iterations (first 10000 discarded), n.thin = 3
#>  n.sims = 26666 iterations saved
#>            mu.vect  sd.vect      2.5%       25%       50%       75%     97.5%
#> N[1]     17620.932 1747.094 14255.313 16420.335 17620.135 18805.003 21038.115
#> N[2]      3865.431  494.270  2956.638  3521.794  3846.224  4187.759  4888.394
#> N[3]      3322.818  385.649  2593.255  3059.681  3315.530  3578.203  4101.060
#> N[4]      3026.909  426.752  2243.175  2726.753  3014.709  3310.150  3902.103
#> N[5]      2669.378  328.055  2033.506  2445.901  2663.353  2888.671  3323.097
#> N[6]      3213.173  548.348  2236.434  2825.869  3179.175  3563.990  4375.893
#> N[7]      3741.243  672.120  2542.617  3268.406  3700.188  4166.541  5184.313
#> N[8]      3469.711  370.932  2762.834  3216.257  3461.591  3713.401  4220.141
#> N[9]      2885.791  207.264  2485.648  2744.921  2883.441  3025.089  3297.378
#> N[10]     2754.180  208.723  2350.956  2612.140  2751.951  2894.106  3167.585
#> N[11]     2760.563  262.793  2254.393  2581.689  2755.457  2936.375  3284.374
#> N[12]     2634.443  299.822  2064.975  2427.444  2626.120  2835.231  3242.785
#> N[13]     1310.312   73.384  1169.702  1260.563  1309.606  1360.093  1455.932
#> N[14]     1570.420   89.239  1398.159  1509.850  1569.490  1629.383  1748.613
#> N[15]     2394.252  152.993  2093.084  2291.155  2394.473  2497.294  2697.668
#> meanr       -0.049    0.003    -0.054    -0.051    -0.049    -0.047    -0.043
#> r[1]        -0.217    0.023    -0.262    -0.232    -0.217    -0.202    -0.173
#> r[2]        -0.150    0.069    -0.285    -0.196    -0.150    -0.103    -0.015
#> r[3]        -0.048    0.061    -0.168    -0.090    -0.048    -0.006     0.071
#> r[4]        -0.025    0.033    -0.089    -0.047    -0.025    -0.002     0.042
#> r[5]         0.089    0.063    -0.033     0.046     0.089     0.132     0.218
#> r[6]         0.075    0.059    -0.039     0.036     0.075     0.114     0.192
#> r[7]        -0.022    0.052    -0.121    -0.057    -0.023     0.013     0.082
#> r[8]        -0.091    0.050    -0.188    -0.125    -0.091    -0.057     0.010
#> r[9]        -0.023    0.044    -0.110    -0.053    -0.023     0.006     0.063
#> r[10]        0.001    0.066    -0.129    -0.044     0.001     0.045     0.131
#> r[11]       -0.049    0.065    -0.179    -0.093    -0.048    -0.005     0.077
#> r[12]       -0.087    0.016    -0.116    -0.097    -0.087    -0.076    -0.055
#> r[13]        0.091    0.035     0.021     0.067     0.091     0.115     0.159
#> r[14]        0.140    0.027     0.086     0.122     0.141     0.159     0.193
#> sdr          0.112    0.009     0.094     0.105     0.111     0.118     0.131
#> vrrmax       0.460    0.039     0.388     0.434     0.459     0.486     0.539
#> deviance   236.150    5.130   227.403   232.536   235.682   239.291   247.330
#>           Rhat n.eff
#> N[1]     1.002  1300
#> N[2]     1.002  1100
#> N[3]     1.003   890
#> N[4]     1.001  6200
#> N[5]     1.001 27000
#> N[6]     1.001 27000
#> N[7]     1.001 27000
#> N[8]     1.001  3700
#> N[9]     1.001 27000
#> N[10]    1.001  3600
#> N[11]    1.001  4000
#> N[12]    1.001 19000
#> N[13]    1.001 27000
#> N[14]    1.001 27000
#> N[15]    1.001 25000
#> meanr    1.002  2500
#> r[1]     1.004   560
#> r[2]     1.001 27000
#> r[3]     1.001  3300
#> r[4]     1.001 18000
#> r[5]     1.001 27000
#> r[6]     1.002 27000
#> r[7]     1.001 17000
#> r[8]     1.001  5100
#> r[9]     1.001  9400
#> r[10]    1.001 27000
#> r[11]    1.001  9200
#> r[12]    1.001 20000
#> r[13]    1.001 27000
#> r[14]    1.001 27000
#> sdr      1.004   560
#> vrrmax   1.004   560
#> deviance 1.002  1300
#> 
#> For each parameter, n.eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor (at convergence, Rhat=1).
#> 
#> DIC info (using the rule, pD = var(deviance)/2)
#> pD = 13.1 and DIC = 249.3
#> DIC is an estimate of expected predictive error (lower deviance is better).
#> 

## Import BUGS outputs for all count series ----
popbayes::read_bugs(path = temp_path)
#> $garamba__alcelaphus_buselaphus
#> Inference for Bugs model at "model.bug", fit using jags,
#>  2 chains, each with 50000 iterations (first 10000 discarded), n.thin = 3
#>  n.sims = 26666 iterations saved
#>            mu.vect  sd.vect      2.5%       25%       50%       75%     97.5%
#> N[1]     17620.932 1747.094 14255.313 16420.335 17620.135 18805.003 21038.115
#> N[2]      3865.431  494.270  2956.638  3521.794  3846.224  4187.759  4888.394
#> N[3]      3322.818  385.649  2593.255  3059.681  3315.530  3578.203  4101.060
#> N[4]      3026.909  426.752  2243.175  2726.753  3014.709  3310.150  3902.103
#> N[5]      2669.378  328.055  2033.506  2445.901  2663.353  2888.671  3323.097
#> N[6]      3213.173  548.348  2236.434  2825.869  3179.175  3563.990  4375.893
#> N[7]      3741.243  672.120  2542.617  3268.406  3700.188  4166.541  5184.313
#> N[8]      3469.711  370.932  2762.834  3216.257  3461.591  3713.401  4220.141
#> N[9]      2885.791  207.264  2485.648  2744.921  2883.441  3025.089  3297.378
#> N[10]     2754.180  208.723  2350.956  2612.140  2751.951  2894.106  3167.585
#> N[11]     2760.563  262.793  2254.393  2581.689  2755.457  2936.375  3284.374
#> N[12]     2634.443  299.822  2064.975  2427.444  2626.120  2835.231  3242.785
#> N[13]     1310.312   73.384  1169.702  1260.563  1309.606  1360.093  1455.932
#> N[14]     1570.420   89.239  1398.159  1509.850  1569.490  1629.383  1748.613
#> N[15]     2394.252  152.993  2093.084  2291.155  2394.473  2497.294  2697.668
#> meanr       -0.049    0.003    -0.054    -0.051    -0.049    -0.047    -0.043
#> r[1]        -0.217    0.023    -0.262    -0.232    -0.217    -0.202    -0.173
#> r[2]        -0.150    0.069    -0.285    -0.196    -0.150    -0.103    -0.015
#> r[3]        -0.048    0.061    -0.168    -0.090    -0.048    -0.006     0.071
#> r[4]        -0.025    0.033    -0.089    -0.047    -0.025    -0.002     0.042
#> r[5]         0.089    0.063    -0.033     0.046     0.089     0.132     0.218
#> r[6]         0.075    0.059    -0.039     0.036     0.075     0.114     0.192
#> r[7]        -0.022    0.052    -0.121    -0.057    -0.023     0.013     0.082
#> r[8]        -0.091    0.050    -0.188    -0.125    -0.091    -0.057     0.010
#> r[9]        -0.023    0.044    -0.110    -0.053    -0.023     0.006     0.063
#> r[10]        0.001    0.066    -0.129    -0.044     0.001     0.045     0.131
#> r[11]       -0.049    0.065    -0.179    -0.093    -0.048    -0.005     0.077
#> r[12]       -0.087    0.016    -0.116    -0.097    -0.087    -0.076    -0.055
#> r[13]        0.091    0.035     0.021     0.067     0.091     0.115     0.159
#> r[14]        0.140    0.027     0.086     0.122     0.141     0.159     0.193
#> sdr          0.112    0.009     0.094     0.105     0.111     0.118     0.131
#> vrrmax       0.460    0.039     0.388     0.434     0.459     0.486     0.539
#> deviance   236.150    5.130   227.403   232.536   235.682   239.291   247.330
#>           Rhat n.eff
#> N[1]     1.002  1300
#> N[2]     1.002  1100
#> N[3]     1.003   890
#> N[4]     1.001  6200
#> N[5]     1.001 27000
#> N[6]     1.001 27000
#> N[7]     1.001 27000
#> N[8]     1.001  3700
#> N[9]     1.001 27000
#> N[10]    1.001  3600
#> N[11]    1.001  4000
#> N[12]    1.001 19000
#> N[13]    1.001 27000
#> N[14]    1.001 27000
#> N[15]    1.001 25000
#> meanr    1.002  2500
#> r[1]     1.004   560
#> r[2]     1.001 27000
#> r[3]     1.001  3300
#> r[4]     1.001 18000
#> r[5]     1.001 27000
#> r[6]     1.002 27000
#> r[7]     1.001 17000
#> r[8]     1.001  5100
#> r[9]     1.001  9400
#> r[10]    1.001 27000
#> r[11]    1.001  9200
#> r[12]    1.001 20000
#> r[13]    1.001 27000
#> r[14]    1.001 27000
#> sdr      1.004   560
#> vrrmax   1.004   560
#> deviance 1.002  1300
#> 
#> For each parameter, n.eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor (at convergence, Rhat=1).
#> 
#> DIC info (using the rule, pD = var(deviance)/2)
#> pD = 13.1 and DIC = 249.3
#> DIC is an estimate of expected predictive error (lower deviance is better).
#> 
# }