Import a list of BUGS outputs previously exported

This function imports a list of BUGS outputs previously exported by fit_trend(). Users can import one, several, or all models.

read_bugs(series = NULL, path = ".")

Arguments

series

a vector of character strings. One or several count series names. If NULL (default) BUGS outputs for all count series will be imported. Users can run list_series() to get the correct spelling of count series names.

path

a character string. The directory in which BUGS outputs have been saved by the function fit_trend().

Value

An n-element list (where n is the number of count series). See fit_trend() for further information.

Examples

## Load Garamba raw dataset ----
file_path <- system.file("extdata", "garamba_survey.csv", 
                         package = "popbayes")
                         
garamba <- read.csv(file = file_path)

## Create temporary folder ----
temp_path <- tempdir()

## Format dataset ----
garamba_formatted <- popbayes::format_data(
  data              = garamba, 
  path              = temp_path,
  field_method      = "field_method",
  pref_field_method = "pref_field_method",
  conversion_A2G    = "conversion_A2G",
  rmax              = "rmax")
#> ✔ Detecting 10 count series

## Select one serie ----
a_buselaphus <- popbayes::filter_series(garamba_formatted, 
                                        location = "Garamba",
                                        species  = "Alcelaphus buselaphus")
#> ✔ Found 1 series with 'Alcelaphus buselaphus' and 'Garamba'.
# \donttest{
## Fit population trends (requires JAGS) ----
a_buselaphus_mod <- popbayes::fit_trend(a_buselaphus, path = temp_path)
#> Compiling data graph
#>    Resolving undeclared variables
#>    Allocating nodes
#>    Initializing
#>    Reading data back into data table
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 15
#>    Unobserved stochastic nodes: 15
#>    Total graph size: 227
#> 
#> Initializing model
#> 

## Import BUGS outputs for one count series ----
popbayes::read_bugs(series = "garamba__alcelaphus_buselaphus", 
                    path   = temp_path)
#> $garamba__alcelaphus_buselaphus
#> Inference for Bugs model at "model.bug", fit using jags,
#>  2 chains, each with 50000 iterations (first 10000 discarded), n.thin = 3
#>  n.sims = 26666 iterations saved
#>            mu.vect  sd.vect      2.5%       25%       50%       75%     97.5%
#> N[1]     17733.573 1751.708 14318.770 16543.963 17734.927 18923.997 21135.502
#> N[2]      3857.909  508.868  2912.695  3507.453  3836.585  4198.951  4887.510
#> N[3]      3318.777  396.861  2553.087  3051.035  3315.086  3584.608  4102.925
#> N[4]      3010.138  432.770  2201.459  2713.940  2997.397  3290.641  3900.209
#> N[5]      2655.506  333.607  2021.733  2423.442  2650.876  2878.475  3329.936
#> N[6]      3184.132  558.595  2208.056  2792.822  3148.380  3530.678  4392.204
#> N[7]      3702.583  687.588  2512.041  3221.133  3645.992  4134.220  5182.170
#> N[8]      3468.225  369.174  2759.855  3215.746  3461.028  3713.921  4213.583
#> N[9]      2891.960  211.085  2482.726  2747.273  2890.538  3032.497  3309.724
#> N[10]     2752.169  206.263  2353.782  2610.771  2751.502  2890.614  3162.199
#> N[11]     2755.358  259.342  2262.887  2576.301  2750.998  2927.067  3275.270
#> N[12]     2625.368  298.733  2063.628  2417.638  2617.993  2824.926  3231.844
#> N[13]     1311.228   73.779  1167.360  1261.069  1311.313  1360.803  1456.349
#> N[14]     1569.757   89.769  1394.347  1509.481  1568.721  1631.015  1746.592
#> N[15]     2393.636  151.900  2097.538  2290.760  2393.475  2494.786  2693.104
#> meanr       -0.049    0.003    -0.054    -0.051    -0.049    -0.047    -0.043
#> r[1]        -0.218    0.023    -0.263    -0.234    -0.219    -0.203    -0.174
#> r[2]        -0.149    0.070    -0.287    -0.196    -0.148    -0.102    -0.013
#> r[3]        -0.050    0.061    -0.170    -0.091    -0.050    -0.009     0.068
#> r[4]        -0.025    0.034    -0.089    -0.048    -0.025    -0.003     0.044
#> r[5]         0.087    0.065    -0.041     0.043     0.087     0.131     0.216
#> r[6]         0.074    0.058    -0.040     0.035     0.074     0.113     0.192
#> r[7]        -0.018    0.053    -0.121    -0.054    -0.017     0.017     0.085
#> r[8]        -0.089    0.050    -0.186    -0.124    -0.090    -0.056     0.011
#> r[9]        -0.025    0.044    -0.111    -0.055    -0.025     0.005     0.061
#> r[10]        0.000    0.066    -0.133    -0.045     0.000     0.044     0.129
#> r[11]       -0.050    0.066    -0.180    -0.095    -0.050    -0.006     0.077
#> r[12]       -0.086    0.015    -0.115    -0.097    -0.086    -0.076    -0.055
#> r[13]        0.090    0.036     0.019     0.066     0.090     0.114     0.161
#> r[14]        0.141    0.027     0.086     0.122     0.141     0.159     0.194
#> sdr          0.112    0.009     0.095     0.106     0.112     0.118     0.132
#> vrrmax       0.461    0.039     0.390     0.435     0.460     0.486     0.542
#> deviance   236.470    5.151   227.691   232.845   236.002   239.600   248.007
#>           Rhat n.eff
#> N[1]     1.001 27000
#> N[2]     1.003  2000
#> N[3]     1.003  2500
#> N[4]     1.003  4100
#> N[5]     1.001  3600
#> N[6]     1.004   430
#> N[7]     1.005   380
#> N[8]     1.002  2700
#> N[9]     1.001 27000
#> N[10]    1.001 27000
#> N[11]    1.001  6000
#> N[12]    1.002  2200
#> N[13]    1.001 27000
#> N[14]    1.001  5300
#> N[15]    1.001 27000
#> meanr    1.001 27000
#> r[1]     1.002  2400
#> r[2]     1.001 17000
#> r[3]     1.001 27000
#> r[4]     1.005  1500
#> r[5]     1.005   410
#> r[6]     1.001 13000
#> r[7]     1.004   460
#> r[8]     1.002  1900
#> r[9]     1.001 27000
#> r[10]    1.001  5200
#> r[11]    1.001  2900
#> r[12]    1.002  2700
#> r[13]    1.001  7200
#> r[14]    1.001 27000
#> sdr      1.003   840
#> vrrmax   1.003   840
#> deviance 1.001  4600
#> 
#> For each parameter, n.eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor (at convergence, Rhat=1).
#> 
#> DIC info (using the rule, pD = var(deviance)/2)
#> pD = 13.3 and DIC = 249.7
#> DIC is an estimate of expected predictive error (lower deviance is better).
#> 

## Import BUGS outputs for all count series ----
popbayes::read_bugs(path = temp_path)
#> $garamba__alcelaphus_buselaphus
#> Inference for Bugs model at "model.bug", fit using jags,
#>  2 chains, each with 50000 iterations (first 10000 discarded), n.thin = 3
#>  n.sims = 26666 iterations saved
#>            mu.vect  sd.vect      2.5%       25%       50%       75%     97.5%
#> N[1]     17733.573 1751.708 14318.770 16543.963 17734.927 18923.997 21135.502
#> N[2]      3857.909  508.868  2912.695  3507.453  3836.585  4198.951  4887.510
#> N[3]      3318.777  396.861  2553.087  3051.035  3315.086  3584.608  4102.925
#> N[4]      3010.138  432.770  2201.459  2713.940  2997.397  3290.641  3900.209
#> N[5]      2655.506  333.607  2021.733  2423.442  2650.876  2878.475  3329.936
#> N[6]      3184.132  558.595  2208.056  2792.822  3148.380  3530.678  4392.204
#> N[7]      3702.583  687.588  2512.041  3221.133  3645.992  4134.220  5182.170
#> N[8]      3468.225  369.174  2759.855  3215.746  3461.028  3713.921  4213.583
#> N[9]      2891.960  211.085  2482.726  2747.273  2890.538  3032.497  3309.724
#> N[10]     2752.169  206.263  2353.782  2610.771  2751.502  2890.614  3162.199
#> N[11]     2755.358  259.342  2262.887  2576.301  2750.998  2927.067  3275.270
#> N[12]     2625.368  298.733  2063.628  2417.638  2617.993  2824.926  3231.844
#> N[13]     1311.228   73.779  1167.360  1261.069  1311.313  1360.803  1456.349
#> N[14]     1569.757   89.769  1394.347  1509.481  1568.721  1631.015  1746.592
#> N[15]     2393.636  151.900  2097.538  2290.760  2393.475  2494.786  2693.104
#> meanr       -0.049    0.003    -0.054    -0.051    -0.049    -0.047    -0.043
#> r[1]        -0.218    0.023    -0.263    -0.234    -0.219    -0.203    -0.174
#> r[2]        -0.149    0.070    -0.287    -0.196    -0.148    -0.102    -0.013
#> r[3]        -0.050    0.061    -0.170    -0.091    -0.050    -0.009     0.068
#> r[4]        -0.025    0.034    -0.089    -0.048    -0.025    -0.003     0.044
#> r[5]         0.087    0.065    -0.041     0.043     0.087     0.131     0.216
#> r[6]         0.074    0.058    -0.040     0.035     0.074     0.113     0.192
#> r[7]        -0.018    0.053    -0.121    -0.054    -0.017     0.017     0.085
#> r[8]        -0.089    0.050    -0.186    -0.124    -0.090    -0.056     0.011
#> r[9]        -0.025    0.044    -0.111    -0.055    -0.025     0.005     0.061
#> r[10]        0.000    0.066    -0.133    -0.045     0.000     0.044     0.129
#> r[11]       -0.050    0.066    -0.180    -0.095    -0.050    -0.006     0.077
#> r[12]       -0.086    0.015    -0.115    -0.097    -0.086    -0.076    -0.055
#> r[13]        0.090    0.036     0.019     0.066     0.090     0.114     0.161
#> r[14]        0.141    0.027     0.086     0.122     0.141     0.159     0.194
#> sdr          0.112    0.009     0.095     0.106     0.112     0.118     0.132
#> vrrmax       0.461    0.039     0.390     0.435     0.460     0.486     0.542
#> deviance   236.470    5.151   227.691   232.845   236.002   239.600   248.007
#>           Rhat n.eff
#> N[1]     1.001 27000
#> N[2]     1.003  2000
#> N[3]     1.003  2500
#> N[4]     1.003  4100
#> N[5]     1.001  3600
#> N[6]     1.004   430
#> N[7]     1.005   380
#> N[8]     1.002  2700
#> N[9]     1.001 27000
#> N[10]    1.001 27000
#> N[11]    1.001  6000
#> N[12]    1.002  2200
#> N[13]    1.001 27000
#> N[14]    1.001  5300
#> N[15]    1.001 27000
#> meanr    1.001 27000
#> r[1]     1.002  2400
#> r[2]     1.001 17000
#> r[3]     1.001 27000
#> r[4]     1.005  1500
#> r[5]     1.005   410
#> r[6]     1.001 13000
#> r[7]     1.004   460
#> r[8]     1.002  1900
#> r[9]     1.001 27000
#> r[10]    1.001  5200
#> r[11]    1.001  2900
#> r[12]    1.002  2700
#> r[13]    1.001  7200
#> r[14]    1.001 27000
#> sdr      1.003   840
#> vrrmax   1.003   840
#> deviance 1.001  4600
#> 
#> For each parameter, n.eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor (at convergence, Rhat=1).
#> 
#> DIC info (using the rule, pD = var(deviance)/2)
#> pD = 13.3 and DIC = 249.7
#> DIC is an estimate of expected predictive error (lower deviance is better).
#> 
# }