Import a list of BUGS outputs previously exported

This function imports a list of BUGS outputs previously exported by fit_trend(). Users can import one, several, or all models.

read_bugs(series = NULL, path = ".")

Arguments

series

a vector of character strings. One or several count series names. If NULL (default) BUGS outputs for all count series will be imported. Users can run list_series() to get the correct spelling of count series names.

path

a character string. The directory in which BUGS outputs have been saved by the function fit_trend().

Value

An n-element list (where n is the number of count series). See fit_trend() for further information.

Examples

## Load Garamba raw dataset ----
file_path <- system.file("extdata", "garamba_survey.csv", 
                         package = "popbayes")
                         
garamba <- read.csv(file = file_path)

## Create temporary folder ----
temp_path <- tempdir()

## Format dataset ----
garamba_formatted <- popbayes::format_data(
  data              = garamba, 
  path              = temp_path,
  field_method      = "field_method",
  pref_field_method = "pref_field_method",
  conversion_A2G    = "conversion_A2G",
  rmax              = "rmax")
#> ✔ Detecting 10 count series

## Select one serie ----
a_buselaphus <- popbayes::filter_series(garamba_formatted, 
                                        location = "Garamba",
                                        species  = "Alcelaphus buselaphus")
#> ✔ Found 1 series with 'Alcelaphus buselaphus' and 'Garamba'.
# \donttest{
## Fit population trends (requires JAGS) ----
a_buselaphus_mod <- popbayes::fit_trend(a_buselaphus, path = temp_path)
#> Compiling data graph
#>    Resolving undeclared variables
#>    Allocating nodes
#>    Initializing
#>    Reading data back into data table
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 15
#>    Unobserved stochastic nodes: 15
#>    Total graph size: 227
#> 
#> Initializing model
#> 

## Import BUGS outputs for one count series ----
popbayes::read_bugs(series = "garamba__alcelaphus_buselaphus", 
                    path   = temp_path)
#> $garamba__alcelaphus_buselaphus
#> Inference for Bugs model at "model.bug", fit using jags,
#>  2 chains, each with 50000 iterations (first 10000 discarded), n.thin = 3
#>  n.sims = 26666 iterations saved
#>            mu.vect  sd.vect      2.5%       25%       50%       75%     97.5%
#> N[1]     17686.918 1695.593 14304.886 16553.375 17710.601 18826.209 20974.513
#> N[2]      3858.650  498.509  2931.782  3512.044  3843.312  4186.612  4877.821
#> N[3]      3320.354  387.272  2584.801  3053.645  3314.290  3580.561  4094.809
#> N[4]      3010.849  426.569  2233.754  2710.599  2995.539  3288.295  3891.037
#> N[5]      2668.111  326.106  2041.057  2444.242  2664.783  2890.394  3310.302
#> N[6]      3219.153  543.620  2248.237  2840.208  3185.437  3564.856  4376.098
#> N[7]      3757.558  681.573  2580.844  3277.017  3705.597  4190.295  5217.434
#> N[8]      3472.076  366.861  2784.686  3218.728  3463.832  3716.575  4217.945
#> N[9]      2885.481  206.856  2486.034  2745.571  2883.188  3023.250  3296.248
#> N[10]     2753.791  205.967  2351.312  2612.962  2752.936  2891.665  3159.461
#> N[11]     2758.905  260.165  2253.906  2584.451  2755.395  2931.896  3283.613
#> N[12]     2630.275  301.218  2061.602  2422.481  2621.997  2829.004  3238.756
#> N[13]     1310.384   73.743  1165.336  1260.884  1310.414  1359.775  1455.525
#> N[14]     1570.636   90.042  1394.535  1509.643  1570.498  1630.670  1747.117
#> N[15]     2394.920  152.075  2096.743  2293.229  2393.680  2497.637  2688.971
#> meanr       -0.049    0.003    -0.054    -0.051    -0.049    -0.047    -0.043
#> r[1]        -0.218    0.022    -0.262    -0.233    -0.218    -0.203    -0.174
#> r[2]        -0.149    0.069    -0.283    -0.195    -0.148    -0.102    -0.013
#> r[3]        -0.051    0.061    -0.170    -0.091    -0.050    -0.010     0.068
#> r[4]        -0.024    0.033    -0.088    -0.046    -0.024    -0.002     0.041
#> r[5]         0.091    0.064    -0.037     0.047     0.091     0.135     0.216
#> r[6]         0.076    0.058    -0.036     0.038     0.076     0.114     0.191
#> r[7]        -0.023    0.052    -0.125    -0.058    -0.022     0.012     0.079
#> r[8]        -0.091    0.051    -0.188    -0.125    -0.092    -0.057     0.010
#> r[9]        -0.023    0.043    -0.110    -0.052    -0.023     0.006     0.061
#> r[10]        0.000    0.067    -0.133    -0.044     0.001     0.045     0.130
#> r[11]       -0.050    0.065    -0.178    -0.094    -0.050    -0.006     0.075
#> r[12]       -0.086    0.016    -0.116    -0.097    -0.087    -0.076    -0.055
#> r[13]        0.091    0.035     0.021     0.067     0.091     0.114     0.160
#> r[14]        0.140    0.027     0.087     0.122     0.140     0.159     0.194
#> sdr          0.112    0.009     0.095     0.106     0.112     0.118     0.131
#> vrrmax       0.462    0.038     0.390     0.436     0.461     0.486     0.539
#> deviance   236.067    5.144   227.316   232.375   235.634   239.228   247.347
#>           Rhat n.eff
#> N[1]     1.001 18000
#> N[2]     1.001 27000
#> N[3]     1.001 27000
#> N[4]     1.002  1300
#> N[5]     1.001  4300
#> N[6]     1.001 27000
#> N[7]     1.001 27000
#> N[8]     1.001 20000
#> N[9]     1.001  6900
#> N[10]    1.001 24000
#> N[11]    1.001 27000
#> N[12]    1.001 27000
#> N[13]    1.001 27000
#> N[14]    1.001 27000
#> N[15]    1.001 27000
#> meanr    1.001 27000
#> r[1]     1.001 17000
#> r[2]     1.001  3300
#> r[3]     1.002  1400
#> r[4]     1.001  6300
#> r[5]     1.002  2200
#> r[6]     1.001 16000
#> r[7]     1.001 27000
#> r[8]     1.001 27000
#> r[9]     1.001 27000
#> r[10]    1.001 27000
#> r[11]    1.001 27000
#> r[12]    1.001 26000
#> r[13]    1.001 27000
#> r[14]    1.001 27000
#> sdr      1.001  3600
#> vrrmax   1.001  3600
#> deviance 1.001 21000
#> 
#> For each parameter, n.eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor (at convergence, Rhat=1).
#> 
#> DIC info (using the rule, pD = var(deviance)/2)
#> pD = 13.2 and DIC = 249.3
#> DIC is an estimate of expected predictive error (lower deviance is better).
#> 

## Import BUGS outputs for all count series ----
popbayes::read_bugs(path = temp_path)
#> $garamba__alcelaphus_buselaphus
#> Inference for Bugs model at "model.bug", fit using jags,
#>  2 chains, each with 50000 iterations (first 10000 discarded), n.thin = 3
#>  n.sims = 26666 iterations saved
#>            mu.vect  sd.vect      2.5%       25%       50%       75%     97.5%
#> N[1]     17686.918 1695.593 14304.886 16553.375 17710.601 18826.209 20974.513
#> N[2]      3858.650  498.509  2931.782  3512.044  3843.312  4186.612  4877.821
#> N[3]      3320.354  387.272  2584.801  3053.645  3314.290  3580.561  4094.809
#> N[4]      3010.849  426.569  2233.754  2710.599  2995.539  3288.295  3891.037
#> N[5]      2668.111  326.106  2041.057  2444.242  2664.783  2890.394  3310.302
#> N[6]      3219.153  543.620  2248.237  2840.208  3185.437  3564.856  4376.098
#> N[7]      3757.558  681.573  2580.844  3277.017  3705.597  4190.295  5217.434
#> N[8]      3472.076  366.861  2784.686  3218.728  3463.832  3716.575  4217.945
#> N[9]      2885.481  206.856  2486.034  2745.571  2883.188  3023.250  3296.248
#> N[10]     2753.791  205.967  2351.312  2612.962  2752.936  2891.665  3159.461
#> N[11]     2758.905  260.165  2253.906  2584.451  2755.395  2931.896  3283.613
#> N[12]     2630.275  301.218  2061.602  2422.481  2621.997  2829.004  3238.756
#> N[13]     1310.384   73.743  1165.336  1260.884  1310.414  1359.775  1455.525
#> N[14]     1570.636   90.042  1394.535  1509.643  1570.498  1630.670  1747.117
#> N[15]     2394.920  152.075  2096.743  2293.229  2393.680  2497.637  2688.971
#> meanr       -0.049    0.003    -0.054    -0.051    -0.049    -0.047    -0.043
#> r[1]        -0.218    0.022    -0.262    -0.233    -0.218    -0.203    -0.174
#> r[2]        -0.149    0.069    -0.283    -0.195    -0.148    -0.102    -0.013
#> r[3]        -0.051    0.061    -0.170    -0.091    -0.050    -0.010     0.068
#> r[4]        -0.024    0.033    -0.088    -0.046    -0.024    -0.002     0.041
#> r[5]         0.091    0.064    -0.037     0.047     0.091     0.135     0.216
#> r[6]         0.076    0.058    -0.036     0.038     0.076     0.114     0.191
#> r[7]        -0.023    0.052    -0.125    -0.058    -0.022     0.012     0.079
#> r[8]        -0.091    0.051    -0.188    -0.125    -0.092    -0.057     0.010
#> r[9]        -0.023    0.043    -0.110    -0.052    -0.023     0.006     0.061
#> r[10]        0.000    0.067    -0.133    -0.044     0.001     0.045     0.130
#> r[11]       -0.050    0.065    -0.178    -0.094    -0.050    -0.006     0.075
#> r[12]       -0.086    0.016    -0.116    -0.097    -0.087    -0.076    -0.055
#> r[13]        0.091    0.035     0.021     0.067     0.091     0.114     0.160
#> r[14]        0.140    0.027     0.087     0.122     0.140     0.159     0.194
#> sdr          0.112    0.009     0.095     0.106     0.112     0.118     0.131
#> vrrmax       0.462    0.038     0.390     0.436     0.461     0.486     0.539
#> deviance   236.067    5.144   227.316   232.375   235.634   239.228   247.347
#>           Rhat n.eff
#> N[1]     1.001 18000
#> N[2]     1.001 27000
#> N[3]     1.001 27000
#> N[4]     1.002  1300
#> N[5]     1.001  4300
#> N[6]     1.001 27000
#> N[7]     1.001 27000
#> N[8]     1.001 20000
#> N[9]     1.001  6900
#> N[10]    1.001 24000
#> N[11]    1.001 27000
#> N[12]    1.001 27000
#> N[13]    1.001 27000
#> N[14]    1.001 27000
#> N[15]    1.001 27000
#> meanr    1.001 27000
#> r[1]     1.001 17000
#> r[2]     1.001  3300
#> r[3]     1.002  1400
#> r[4]     1.001  6300
#> r[5]     1.002  2200
#> r[6]     1.001 16000
#> r[7]     1.001 27000
#> r[8]     1.001 27000
#> r[9]     1.001 27000
#> r[10]    1.001 27000
#> r[11]    1.001 27000
#> r[12]    1.001 26000
#> r[13]    1.001 27000
#> r[14]    1.001 27000
#> sdr      1.001  3600
#> vrrmax   1.001  3600
#> deviance 1.001 21000
#> 
#> For each parameter, n.eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor (at convergence, Rhat=1).
#> 
#> DIC info (using the rule, pD = var(deviance)/2)
#> pD = 13.2 and DIC = 249.3
#> DIC is an estimate of expected predictive error (lower deviance is better).
#> 
# }