Import a list of BUGS outputs previously exported

This function imports a list of BUGS outputs previously exported by fit_trend(). Users can import one, several, or all models.

read_bugs(series = NULL, path = ".")

Arguments

series

a vector of character strings. One or several count series names. If NULL (default) BUGS outputs for all count series will be imported. Users can run list_series() to get the correct spelling of count series names.

path

a character string. The directory in which BUGS outputs have been saved by the function fit_trend().

Value

An n-element list (where n is the number of count series). See fit_trend() for further information.

Examples

## Load Garamba raw dataset ----
file_path <- system.file("extdata", "garamba_survey.csv", 
                         package = "popbayes")
                         
garamba <- read.csv(file = file_path)

## Create temporary folder ----
temp_path <- tempdir()

## Format dataset ----
garamba_formatted <- popbayes::format_data(
  data              = garamba, 
  path              = temp_path,
  field_method      = "field_method",
  pref_field_method = "pref_field_method",
  conversion_A2G    = "conversion_A2G",
  rmax              = "rmax")
#> ✔ Detecting 10 count series

## Select one serie ----
a_buselaphus <- popbayes::filter_series(garamba_formatted, 
                                        location = "Garamba",
                                        species  = "Alcelaphus buselaphus")
#> ✔ Found 1 series with 'Alcelaphus buselaphus' and 'Garamba'.
# \donttest{
## Fit population trends (requires JAGS) ----
a_buselaphus_mod <- popbayes::fit_trend(a_buselaphus, path = temp_path)
#> Compiling data graph
#>    Resolving undeclared variables
#>    Allocating nodes
#>    Initializing
#>    Reading data back into data table
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 15
#>    Unobserved stochastic nodes: 15
#>    Total graph size: 227
#> 
#> Initializing model
#> 

## Import BUGS outputs for one count series ----
popbayes::read_bugs(series = "garamba__alcelaphus_buselaphus", 
                    path   = temp_path)
#> $garamba__alcelaphus_buselaphus
#> Inference for Bugs model at "model.bug", fit using jags,
#>  2 chains, each with 50000 iterations (first 10000 discarded), n.thin = 3
#>  n.sims = 26666 iterations saved
#>            mu.vect  sd.vect      2.5%       25%       50%       75%     97.5%
#> N[1]     17681.059 1740.197 14251.208 16512.325 17675.767 18846.443 21085.835
#> N[2]      3844.707  501.674  2883.671  3503.604  3831.501  4170.255  4868.737
#> N[3]      3307.145  396.661  2542.824  3043.713  3297.831  3568.231  4100.702
#> N[4]      2991.051  430.165  2190.115  2694.061  2974.183  3275.168  3871.672
#> N[5]      2675.855  330.878  2040.079  2451.142  2670.347  2897.806  3331.546
#> N[6]      3231.527  551.365  2229.701  2849.896  3200.999  3584.445  4392.719
#> N[7]      3761.422  685.372  2548.135  3287.532  3711.448  4185.506  5240.120
#> N[8]      3477.820  371.817  2762.856  3226.666  3474.507  3726.792  4209.727
#> N[9]      2887.506  209.600  2483.995  2744.780  2886.785  3028.693  3303.183
#> N[10]     2746.548  205.054  2348.313  2606.457  2744.470  2882.883  3156.470
#> N[11]     2750.242  258.535  2256.987  2574.264  2746.841  2920.513  3272.113
#> N[12]     2620.422  297.900  2058.903  2417.746  2614.402  2815.653  3223.308
#> N[13]     1312.409   73.443  1168.210  1263.196  1311.788  1362.392  1455.081
#> N[14]     1570.226   89.120  1399.030  1509.313  1570.031  1630.301  1746.584
#> N[15]     2392.673  152.785  2092.148  2288.835  2392.308  2495.552  2690.842
#> meanr       -0.049    0.003    -0.054    -0.051    -0.049    -0.047    -0.043
#> r[1]        -0.219    0.023    -0.265    -0.233    -0.218    -0.203    -0.174
#> r[2]        -0.149    0.069    -0.284    -0.196    -0.149    -0.103    -0.013
#> r[3]        -0.052    0.061    -0.172    -0.092    -0.052    -0.011     0.066
#> r[4]        -0.022    0.034    -0.089    -0.044    -0.022     0.001     0.045
#> r[5]         0.091    0.065    -0.037     0.047     0.090     0.135     0.222
#> r[6]         0.075    0.059    -0.038     0.035     0.074     0.114     0.194
#> r[7]        -0.023    0.052    -0.122    -0.058    -0.023     0.012     0.081
#> r[8]        -0.091    0.051    -0.191    -0.126    -0.092    -0.058     0.011
#> r[9]        -0.025    0.044    -0.112    -0.055    -0.025     0.005     0.062
#> r[10]        0.000    0.067    -0.134    -0.045     0.001     0.045     0.127
#> r[11]       -0.050    0.065    -0.181    -0.093    -0.050    -0.007     0.077
#> r[12]       -0.086    0.015    -0.115    -0.096    -0.086    -0.076    -0.055
#> r[13]        0.090    0.036     0.020     0.066     0.090     0.114     0.159
#> r[14]        0.140    0.027     0.086     0.122     0.140     0.159     0.193
#> sdr          0.112    0.009     0.095     0.106     0.112     0.118     0.132
#> vrrmax       0.462    0.039     0.390     0.435     0.461     0.487     0.542
#> deviance   236.313    5.266   227.425   232.580   235.870   239.480   248.024
#>           Rhat n.eff
#> N[1]     1.001 11000
#> N[2]     1.001  5900
#> N[3]     1.001 27000
#> N[4]     1.003 10000
#> N[5]     1.001  3500
#> N[6]     1.001 12000
#> N[7]     1.001  5600
#> N[8]     1.001 27000
#> N[9]     1.001  5300
#> N[10]    1.002  1000
#> N[11]    1.004   480
#> N[12]    1.003   620
#> N[13]    1.001 10000
#> N[14]    1.001 27000
#> N[15]    1.001 27000
#> meanr    1.001 27000
#> r[1]     1.001  3600
#> r[2]     1.001  3100
#> r[3]     1.001 16000
#> r[4]     1.001 27000
#> r[5]     1.002  1200
#> r[6]     1.001 19000
#> r[7]     1.001  2900
#> r[8]     1.001 27000
#> r[9]     1.003   700
#> r[10]    1.002  1200
#> r[11]    1.001 27000
#> r[12]    1.004   580
#> r[13]    1.001 24000
#> r[14]    1.001 27000
#> sdr      1.003   970
#> vrrmax   1.003   970
#> deviance 1.002  1000
#> 
#> For each parameter, n.eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor (at convergence, Rhat=1).
#> 
#> DIC info (using the rule, pD = var(deviance)/2)
#> pD = 13.9 and DIC = 250.2
#> DIC is an estimate of expected predictive error (lower deviance is better).
#> 

## Import BUGS outputs for all count series ----
popbayes::read_bugs(path = temp_path)
#> $garamba__alcelaphus_buselaphus
#> Inference for Bugs model at "model.bug", fit using jags,
#>  2 chains, each with 50000 iterations (first 10000 discarded), n.thin = 3
#>  n.sims = 26666 iterations saved
#>            mu.vect  sd.vect      2.5%       25%       50%       75%     97.5%
#> N[1]     17681.059 1740.197 14251.208 16512.325 17675.767 18846.443 21085.835
#> N[2]      3844.707  501.674  2883.671  3503.604  3831.501  4170.255  4868.737
#> N[3]      3307.145  396.661  2542.824  3043.713  3297.831  3568.231  4100.702
#> N[4]      2991.051  430.165  2190.115  2694.061  2974.183  3275.168  3871.672
#> N[5]      2675.855  330.878  2040.079  2451.142  2670.347  2897.806  3331.546
#> N[6]      3231.527  551.365  2229.701  2849.896  3200.999  3584.445  4392.719
#> N[7]      3761.422  685.372  2548.135  3287.532  3711.448  4185.506  5240.120
#> N[8]      3477.820  371.817  2762.856  3226.666  3474.507  3726.792  4209.727
#> N[9]      2887.506  209.600  2483.995  2744.780  2886.785  3028.693  3303.183
#> N[10]     2746.548  205.054  2348.313  2606.457  2744.470  2882.883  3156.470
#> N[11]     2750.242  258.535  2256.987  2574.264  2746.841  2920.513  3272.113
#> N[12]     2620.422  297.900  2058.903  2417.746  2614.402  2815.653  3223.308
#> N[13]     1312.409   73.443  1168.210  1263.196  1311.788  1362.392  1455.081
#> N[14]     1570.226   89.120  1399.030  1509.313  1570.031  1630.301  1746.584
#> N[15]     2392.673  152.785  2092.148  2288.835  2392.308  2495.552  2690.842
#> meanr       -0.049    0.003    -0.054    -0.051    -0.049    -0.047    -0.043
#> r[1]        -0.219    0.023    -0.265    -0.233    -0.218    -0.203    -0.174
#> r[2]        -0.149    0.069    -0.284    -0.196    -0.149    -0.103    -0.013
#> r[3]        -0.052    0.061    -0.172    -0.092    -0.052    -0.011     0.066
#> r[4]        -0.022    0.034    -0.089    -0.044    -0.022     0.001     0.045
#> r[5]         0.091    0.065    -0.037     0.047     0.090     0.135     0.222
#> r[6]         0.075    0.059    -0.038     0.035     0.074     0.114     0.194
#> r[7]        -0.023    0.052    -0.122    -0.058    -0.023     0.012     0.081
#> r[8]        -0.091    0.051    -0.191    -0.126    -0.092    -0.058     0.011
#> r[9]        -0.025    0.044    -0.112    -0.055    -0.025     0.005     0.062
#> r[10]        0.000    0.067    -0.134    -0.045     0.001     0.045     0.127
#> r[11]       -0.050    0.065    -0.181    -0.093    -0.050    -0.007     0.077
#> r[12]       -0.086    0.015    -0.115    -0.096    -0.086    -0.076    -0.055
#> r[13]        0.090    0.036     0.020     0.066     0.090     0.114     0.159
#> r[14]        0.140    0.027     0.086     0.122     0.140     0.159     0.193
#> sdr          0.112    0.009     0.095     0.106     0.112     0.118     0.132
#> vrrmax       0.462    0.039     0.390     0.435     0.461     0.487     0.542
#> deviance   236.313    5.266   227.425   232.580   235.870   239.480   248.024
#>           Rhat n.eff
#> N[1]     1.001 11000
#> N[2]     1.001  5900
#> N[3]     1.001 27000
#> N[4]     1.003 10000
#> N[5]     1.001  3500
#> N[6]     1.001 12000
#> N[7]     1.001  5600
#> N[8]     1.001 27000
#> N[9]     1.001  5300
#> N[10]    1.002  1000
#> N[11]    1.004   480
#> N[12]    1.003   620
#> N[13]    1.001 10000
#> N[14]    1.001 27000
#> N[15]    1.001 27000
#> meanr    1.001 27000
#> r[1]     1.001  3600
#> r[2]     1.001  3100
#> r[3]     1.001 16000
#> r[4]     1.001 27000
#> r[5]     1.002  1200
#> r[6]     1.001 19000
#> r[7]     1.001  2900
#> r[8]     1.001 27000
#> r[9]     1.003   700
#> r[10]    1.002  1200
#> r[11]    1.001 27000
#> r[12]    1.004   580
#> r[13]    1.001 24000
#> r[14]    1.001 27000
#> sdr      1.003   970
#> vrrmax   1.003   970
#> deviance 1.002  1000
#> 
#> For each parameter, n.eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor (at convergence, Rhat=1).
#> 
#> DIC info (using the rule, pD = var(deviance)/2)
#> pD = 13.9 and DIC = 250.2
#> DIC is an estimate of expected predictive error (lower deviance is better).
#> 
# }