Import a list of BUGS outputs previously exported

This function imports a list of BUGS outputs previously exported by fit_trend(). Users can import one, several, or all models.

read_bugs(series = NULL, path = ".")

Arguments

series

a vector of character strings. One or several count series names. If NULL (default) BUGS outputs for all count series will be imported. Users can run list_series() to get the correct spelling of count series names.

path

a character string. The directory in which BUGS outputs have been saved by the function fit_trend().

Value

An n-element list (where n is the number of count series). See fit_trend() for further information.

Examples

## Load Garamba raw dataset ----
file_path <- system.file("extdata", "garamba_survey.csv", 
                         package = "popbayes")
                         
garamba <- read.csv(file = file_path)

## Create temporary folder ----
temp_path <- tempdir()

## Format dataset ----
garamba_formatted <- popbayes::format_data(
  data              = garamba, 
  path              = temp_path,
  field_method      = "field_method",
  pref_field_method = "pref_field_method",
  conversion_A2G    = "conversion_A2G",
  rmax              = "rmax")
#> ✔ Detecting 10 count series.

## Select one serie ----
a_buselaphus <- popbayes::filter_series(garamba_formatted, 
                                        location = "Garamba",
                                        species  = "Alcelaphus buselaphus")
#> ✔ Found 1 series with "Alcelaphus buselaphus" and "Garamba".
# \donttest{
## Fit population trends (requires JAGS) ----
a_buselaphus_mod <- popbayes::fit_trend(a_buselaphus, path = temp_path)
#> Compiling data graph
#>    Resolving undeclared variables
#>    Allocating nodes
#>    Initializing
#>    Reading data back into data table
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 15
#>    Unobserved stochastic nodes: 15
#>    Total graph size: 227
#> 
#> Initializing model
#> 

## Import BUGS outputs for one count series ----
popbayes::read_bugs(series = "garamba__alcelaphus_buselaphus", 
                    path   = temp_path)
#> $garamba__alcelaphus_buselaphus
#> Inference for Bugs model at "model.bug", fit using jags,
#>  2 chains, each with 50000 iterations (first 10000 discarded), n.thin = 3
#>  n.sims = 26666 iterations saved
#>            mu.vect  sd.vect      2.5%       25%       50%       75%     97.5%
#> N[1]     17722.076 1710.658 14321.167 16585.850 17736.096 18859.797 21066.199
#> N[2]      3855.123  496.941  2931.932  3511.800  3841.709  4183.929  4862.683
#> N[3]      3317.596  387.275  2569.655  3055.299  3316.870  3578.810  4077.283
#> N[4]      3025.798  422.113  2241.447  2737.125  3010.210  3303.343  3891.711
#> N[5]      2672.887  320.569  2058.607  2451.913  2666.287  2890.196  3312.771
#> N[6]      3236.485  544.340  2271.684  2857.773  3189.665  3579.589  4412.983
#> N[7]      3782.196  686.179  2552.608  3304.917  3738.049  4204.195  5275.816
#> N[8]      3475.951  367.288  2782.686  3220.010  3468.453  3724.091  4221.741
#> N[9]      2887.353  209.818  2486.550  2744.898  2882.567  3029.051  3305.855
#> N[10]     2748.101  206.734  2345.741  2606.775  2746.873  2886.469  3158.058
#> N[11]     2751.209  259.378  2259.313  2571.082  2746.882  2925.867  3266.590
#> N[12]     2624.841  302.213  2060.401  2413.027  2616.364  2827.269  3241.629
#> N[13]     1312.417   73.518  1168.662  1263.003  1312.446  1361.289  1457.364
#> N[14]     1570.722   89.498  1395.485  1509.634  1570.806  1631.468  1743.057
#> N[15]     2393.904  151.641  2100.203  2291.621  2393.327  2495.098  2693.941
#> meanr       -0.049    0.003    -0.054    -0.051    -0.049    -0.047    -0.043
#> r[1]        -0.218    0.023    -0.263    -0.233    -0.218    -0.203    -0.174
#> r[2]        -0.149    0.069    -0.284    -0.195    -0.148    -0.102    -0.014
#> r[3]        -0.047    0.060    -0.169    -0.088    -0.046    -0.007     0.068
#> r[4]        -0.024    0.032    -0.088    -0.046    -0.025    -0.003     0.039
#> r[5]         0.092    0.064    -0.034     0.049     0.093     0.136     0.222
#> r[6]         0.077    0.057    -0.034     0.038     0.076     0.115     0.191
#> r[7]        -0.025    0.052    -0.125    -0.060    -0.025     0.010     0.079
#> r[8]        -0.091    0.050    -0.188    -0.126    -0.092    -0.058     0.008
#> r[9]        -0.025    0.044    -0.113    -0.054    -0.025     0.005     0.063
#> r[10]        0.000    0.067    -0.133    -0.046     0.000     0.045     0.132
#> r[11]       -0.049    0.066    -0.178    -0.094    -0.049    -0.005     0.079
#> r[12]       -0.086    0.016    -0.116    -0.097    -0.086    -0.076    -0.055
#> r[13]        0.090    0.036     0.019     0.066     0.090     0.114     0.160
#> r[14]        0.140    0.027     0.087     0.122     0.140     0.159     0.193
#> sdr          0.112    0.009     0.095     0.106     0.112     0.118     0.132
#> vrrmax       0.462    0.039     0.390     0.435     0.461     0.487     0.542
#> deviance   235.986    5.170   227.159   232.318   235.542   239.184   247.386
#>           Rhat n.eff
#> N[1]     1.006   340
#> N[2]     1.004   500
#> N[3]     1.003   880
#> N[4]     1.001  2800
#> N[5]     1.001 27000
#> N[6]     1.002  1700
#> N[7]     1.004   590
#> N[8]     1.002  2400
#> N[9]     1.001 12000
#> N[10]    1.001 27000
#> N[11]    1.001 22000
#> N[12]    1.001 27000
#> N[13]    1.001  4800
#> N[14]    1.001 27000
#> N[15]    1.001 27000
#> meanr    1.004   530
#> r[1]     1.008   200
#> r[2]     1.002  1400
#> r[3]     1.001  8000
#> r[4]     1.001  3500
#> r[5]     1.002  1000
#> r[6]     1.002  1100
#> r[7]     1.003   860
#> r[8]     1.001  4400
#> r[9]     1.001 27000
#> r[10]    1.001 27000
#> r[11]    1.001 27000
#> r[12]    1.001 14000
#> r[13]    1.001  4300
#> r[14]    1.001 27000
#> sdr      1.007   250
#> vrrmax   1.007   250
#> deviance 1.004   570
#> 
#> For each parameter, n.eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor (at convergence, Rhat=1).
#> 
#> DIC info (using the rule, pD = var(deviance)/2)
#> pD = 13.3 and DIC = 249.3
#> DIC is an estimate of expected predictive error (lower deviance is better).
#> 

## Import BUGS outputs for all count series ----
popbayes::read_bugs(path = temp_path)
#> $garamba__alcelaphus_buselaphus
#> Inference for Bugs model at "model.bug", fit using jags,
#>  2 chains, each with 50000 iterations (first 10000 discarded), n.thin = 3
#>  n.sims = 26666 iterations saved
#>            mu.vect  sd.vect      2.5%       25%       50%       75%     97.5%
#> N[1]     17722.076 1710.658 14321.167 16585.850 17736.096 18859.797 21066.199
#> N[2]      3855.123  496.941  2931.932  3511.800  3841.709  4183.929  4862.683
#> N[3]      3317.596  387.275  2569.655  3055.299  3316.870  3578.810  4077.283
#> N[4]      3025.798  422.113  2241.447  2737.125  3010.210  3303.343  3891.711
#> N[5]      2672.887  320.569  2058.607  2451.913  2666.287  2890.196  3312.771
#> N[6]      3236.485  544.340  2271.684  2857.773  3189.665  3579.589  4412.983
#> N[7]      3782.196  686.179  2552.608  3304.917  3738.049  4204.195  5275.816
#> N[8]      3475.951  367.288  2782.686  3220.010  3468.453  3724.091  4221.741
#> N[9]      2887.353  209.818  2486.550  2744.898  2882.567  3029.051  3305.855
#> N[10]     2748.101  206.734  2345.741  2606.775  2746.873  2886.469  3158.058
#> N[11]     2751.209  259.378  2259.313  2571.082  2746.882  2925.867  3266.590
#> N[12]     2624.841  302.213  2060.401  2413.027  2616.364  2827.269  3241.629
#> N[13]     1312.417   73.518  1168.662  1263.003  1312.446  1361.289  1457.364
#> N[14]     1570.722   89.498  1395.485  1509.634  1570.806  1631.468  1743.057
#> N[15]     2393.904  151.641  2100.203  2291.621  2393.327  2495.098  2693.941
#> meanr       -0.049    0.003    -0.054    -0.051    -0.049    -0.047    -0.043
#> r[1]        -0.218    0.023    -0.263    -0.233    -0.218    -0.203    -0.174
#> r[2]        -0.149    0.069    -0.284    -0.195    -0.148    -0.102    -0.014
#> r[3]        -0.047    0.060    -0.169    -0.088    -0.046    -0.007     0.068
#> r[4]        -0.024    0.032    -0.088    -0.046    -0.025    -0.003     0.039
#> r[5]         0.092    0.064    -0.034     0.049     0.093     0.136     0.222
#> r[6]         0.077    0.057    -0.034     0.038     0.076     0.115     0.191
#> r[7]        -0.025    0.052    -0.125    -0.060    -0.025     0.010     0.079
#> r[8]        -0.091    0.050    -0.188    -0.126    -0.092    -0.058     0.008
#> r[9]        -0.025    0.044    -0.113    -0.054    -0.025     0.005     0.063
#> r[10]        0.000    0.067    -0.133    -0.046     0.000     0.045     0.132
#> r[11]       -0.049    0.066    -0.178    -0.094    -0.049    -0.005     0.079
#> r[12]       -0.086    0.016    -0.116    -0.097    -0.086    -0.076    -0.055
#> r[13]        0.090    0.036     0.019     0.066     0.090     0.114     0.160
#> r[14]        0.140    0.027     0.087     0.122     0.140     0.159     0.193
#> sdr          0.112    0.009     0.095     0.106     0.112     0.118     0.132
#> vrrmax       0.462    0.039     0.390     0.435     0.461     0.487     0.542
#> deviance   235.986    5.170   227.159   232.318   235.542   239.184   247.386
#>           Rhat n.eff
#> N[1]     1.006   340
#> N[2]     1.004   500
#> N[3]     1.003   880
#> N[4]     1.001  2800
#> N[5]     1.001 27000
#> N[6]     1.002  1700
#> N[7]     1.004   590
#> N[8]     1.002  2400
#> N[9]     1.001 12000
#> N[10]    1.001 27000
#> N[11]    1.001 22000
#> N[12]    1.001 27000
#> N[13]    1.001  4800
#> N[14]    1.001 27000
#> N[15]    1.001 27000
#> meanr    1.004   530
#> r[1]     1.008   200
#> r[2]     1.002  1400
#> r[3]     1.001  8000
#> r[4]     1.001  3500
#> r[5]     1.002  1000
#> r[6]     1.002  1100
#> r[7]     1.003   860
#> r[8]     1.001  4400
#> r[9]     1.001 27000
#> r[10]    1.001 27000
#> r[11]    1.001 27000
#> r[12]    1.001 14000
#> r[13]    1.001  4300
#> r[14]    1.001 27000
#> sdr      1.007   250
#> vrrmax   1.007   250
#> deviance 1.004   570
#> 
#> For each parameter, n.eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor (at convergence, Rhat=1).
#> 
#> DIC info (using the rule, pD = var(deviance)/2)
#> pD = 13.3 and DIC = 249.3
#> DIC is an estimate of expected predictive error (lower deviance is better).
#> 
# }