Import a list of BUGS outputs previously exported

This function imports a list of BUGS outputs previously exported by fit_trend(). Users can import one, several, or all models.

read_bugs(series = NULL, path = ".")

Arguments

series

a vector of character strings. One or several count series names. If NULL (default) BUGS outputs for all count series will be imported. Users can run list_series() to get the correct spelling of count series names.

path

a character string. The directory in which BUGS outputs have been saved by the function fit_trend().

Value

An n-element list (where n is the number of count series). See fit_trend() for further information.

Examples

## Load Garamba raw dataset ----
file_path <- system.file("extdata", "garamba_survey.csv", 
                         package = "popbayes")
                         
garamba <- read.csv(file = file_path)

## Create temporary folder ----
temp_path <- tempdir()

## Format dataset ----
garamba_formatted <- popbayes::format_data(
  data              = garamba, 
  path              = temp_path,
  field_method      = "field_method",
  pref_field_method = "pref_field_method",
  conversion_A2G    = "conversion_A2G",
  rmax              = "rmax")
#> ✔ Detecting 10 count series

## Select one serie ----
a_buselaphus <- popbayes::filter_series(garamba_formatted, 
                                        location = "Garamba",
                                        species  = "Alcelaphus buselaphus")
#> ✔ Found 1 series with 'Alcelaphus buselaphus' and 'Garamba'.
# \donttest{
## Fit population trends (requires JAGS) ----
a_buselaphus_mod <- popbayes::fit_trend(a_buselaphus, path = temp_path)
#> Compiling data graph
#>    Resolving undeclared variables
#>    Allocating nodes
#>    Initializing
#>    Reading data back into data table
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 15
#>    Unobserved stochastic nodes: 15
#>    Total graph size: 227
#> 
#> Initializing model
#> 

## Import BUGS outputs for one count series ----
popbayes::read_bugs(series = "garamba__alcelaphus_buselaphus", 
                    path   = temp_path)
#> $garamba__alcelaphus_buselaphus
#> Inference for Bugs model at "model.bug", fit using jags,
#>  2 chains, each with 50000 iterations (first 10000 discarded), n.thin = 3
#>  n.sims = 26666 iterations saved
#>            mu.vect  sd.vect      2.5%       25%       50%       75%     97.5%
#> N[1]     17680.130 1720.230 14282.115 16511.212 17683.486 18837.937 21069.473
#> N[2]      3867.864  501.200  2934.633  3520.689  3850.719  4205.499  4886.091
#> N[3]      3327.177  389.227  2567.674  3062.307  3322.447  3588.449  4100.157
#> N[4]      3023.246  429.867  2219.171  2728.394  3011.217  3298.496  3910.030
#> N[5]      2657.546  331.893  2015.879  2430.948  2657.209  2884.211  3307.792
#> N[6]      3183.807  550.729  2166.016  2804.482  3155.593  3538.467  4338.702
#> N[7]      3696.770  688.126  2469.993  3218.295  3649.672  4132.701  5168.849
#> N[8]      3459.097  374.962  2740.949  3202.892  3453.521  3705.493  4212.421
#> N[9]      2889.755  208.819  2485.687  2747.831  2887.959  3030.115  3302.362
#> N[10]     2753.182  205.502  2360.157  2612.094  2752.466  2891.964  3161.217
#> N[11]     2760.712  257.558  2272.155  2585.163  2755.811  2932.508  3272.941
#> N[12]     2633.547  295.676  2072.808  2433.602  2624.892  2828.420  3231.760
#> N[13]     1310.958   73.582  1168.704  1260.771  1310.733  1360.883  1454.498
#> N[14]     1570.451   89.537  1397.946  1509.100  1569.443  1630.934  1747.967
#> N[15]     2392.420  152.373  2094.661  2289.749  2392.126  2495.358  2690.959
#> meanr       -0.049    0.003    -0.054    -0.051    -0.049    -0.047    -0.043
#> r[1]        -0.218    0.023    -0.263    -0.233    -0.218    -0.202    -0.174
#> r[2]        -0.149    0.068    -0.280    -0.195    -0.149    -0.103    -0.012
#> r[3]        -0.050    0.061    -0.170    -0.091    -0.049    -0.009     0.072
#> r[4]        -0.025    0.033    -0.091    -0.047    -0.025    -0.003     0.039
#> r[5]         0.087    0.065    -0.044     0.044     0.088     0.131     0.215
#> r[6]         0.074    0.059    -0.042     0.034     0.074     0.114     0.186
#> r[7]        -0.018    0.054    -0.120    -0.055    -0.019     0.017     0.088
#> r[8]        -0.088    0.052    -0.186    -0.123    -0.089    -0.054     0.016
#> r[9]        -0.024    0.044    -0.111    -0.054    -0.024     0.005     0.061
#> r[10]        0.001    0.067    -0.133    -0.043     0.002     0.046     0.131
#> r[11]       -0.049    0.065    -0.177    -0.093    -0.049    -0.005     0.076
#> r[12]       -0.087    0.015    -0.116    -0.097    -0.087    -0.077    -0.055
#> r[13]        0.090    0.036     0.021     0.066     0.090     0.114     0.161
#> r[14]        0.140    0.027     0.086     0.122     0.140     0.159     0.193
#> sdr          0.112    0.009     0.094     0.105     0.111     0.118     0.131
#> vrrmax       0.460    0.039     0.387     0.434     0.459     0.485     0.539
#> deviance   236.352    5.211   227.459   232.655   235.879   239.486   247.815
#>           Rhat n.eff
#> N[1]     1.001  6900
#> N[2]     1.001 17000
#> N[3]     1.001  5400
#> N[4]     1.001 27000
#> N[5]     1.002  1200
#> N[6]     1.003   700
#> N[7]     1.004   440
#> N[8]     1.001 14000
#> N[9]     1.001 13000
#> N[10]    1.001 27000
#> N[11]    1.001 11000
#> N[12]    1.001 27000
#> N[13]    1.001 27000
#> N[14]    1.001 12000
#> N[15]    1.001 27000
#> meanr    1.001 13000
#> r[1]     1.001 27000
#> r[2]     1.001 13000
#> r[3]     1.001 15000
#> r[4]     1.002  1500
#> r[5]     1.002  1900
#> r[6]     1.002  2600
#> r[7]     1.007   270
#> r[8]     1.001 27000
#> r[9]     1.001 24000
#> r[10]    1.001  4700
#> r[11]    1.001  4800
#> r[12]    1.001 27000
#> r[13]    1.001 27000
#> r[14]    1.001 13000
#> sdr      1.001  7400
#> vrrmax   1.001  7400
#> deviance 1.001  8600
#> 
#> For each parameter, n.eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor (at convergence, Rhat=1).
#> 
#> DIC info (using the rule, pD = var(deviance)/2)
#> pD = 13.6 and DIC = 249.9
#> DIC is an estimate of expected predictive error (lower deviance is better).
#> 

## Import BUGS outputs for all count series ----
popbayes::read_bugs(path = temp_path)
#> $garamba__alcelaphus_buselaphus
#> Inference for Bugs model at "model.bug", fit using jags,
#>  2 chains, each with 50000 iterations (first 10000 discarded), n.thin = 3
#>  n.sims = 26666 iterations saved
#>            mu.vect  sd.vect      2.5%       25%       50%       75%     97.5%
#> N[1]     17680.130 1720.230 14282.115 16511.212 17683.486 18837.937 21069.473
#> N[2]      3867.864  501.200  2934.633  3520.689  3850.719  4205.499  4886.091
#> N[3]      3327.177  389.227  2567.674  3062.307  3322.447  3588.449  4100.157
#> N[4]      3023.246  429.867  2219.171  2728.394  3011.217  3298.496  3910.030
#> N[5]      2657.546  331.893  2015.879  2430.948  2657.209  2884.211  3307.792
#> N[6]      3183.807  550.729  2166.016  2804.482  3155.593  3538.467  4338.702
#> N[7]      3696.770  688.126  2469.993  3218.295  3649.672  4132.701  5168.849
#> N[8]      3459.097  374.962  2740.949  3202.892  3453.521  3705.493  4212.421
#> N[9]      2889.755  208.819  2485.687  2747.831  2887.959  3030.115  3302.362
#> N[10]     2753.182  205.502  2360.157  2612.094  2752.466  2891.964  3161.217
#> N[11]     2760.712  257.558  2272.155  2585.163  2755.811  2932.508  3272.941
#> N[12]     2633.547  295.676  2072.808  2433.602  2624.892  2828.420  3231.760
#> N[13]     1310.958   73.582  1168.704  1260.771  1310.733  1360.883  1454.498
#> N[14]     1570.451   89.537  1397.946  1509.100  1569.443  1630.934  1747.967
#> N[15]     2392.420  152.373  2094.661  2289.749  2392.126  2495.358  2690.959
#> meanr       -0.049    0.003    -0.054    -0.051    -0.049    -0.047    -0.043
#> r[1]        -0.218    0.023    -0.263    -0.233    -0.218    -0.202    -0.174
#> r[2]        -0.149    0.068    -0.280    -0.195    -0.149    -0.103    -0.012
#> r[3]        -0.050    0.061    -0.170    -0.091    -0.049    -0.009     0.072
#> r[4]        -0.025    0.033    -0.091    -0.047    -0.025    -0.003     0.039
#> r[5]         0.087    0.065    -0.044     0.044     0.088     0.131     0.215
#> r[6]         0.074    0.059    -0.042     0.034     0.074     0.114     0.186
#> r[7]        -0.018    0.054    -0.120    -0.055    -0.019     0.017     0.088
#> r[8]        -0.088    0.052    -0.186    -0.123    -0.089    -0.054     0.016
#> r[9]        -0.024    0.044    -0.111    -0.054    -0.024     0.005     0.061
#> r[10]        0.001    0.067    -0.133    -0.043     0.002     0.046     0.131
#> r[11]       -0.049    0.065    -0.177    -0.093    -0.049    -0.005     0.076
#> r[12]       -0.087    0.015    -0.116    -0.097    -0.087    -0.077    -0.055
#> r[13]        0.090    0.036     0.021     0.066     0.090     0.114     0.161
#> r[14]        0.140    0.027     0.086     0.122     0.140     0.159     0.193
#> sdr          0.112    0.009     0.094     0.105     0.111     0.118     0.131
#> vrrmax       0.460    0.039     0.387     0.434     0.459     0.485     0.539
#> deviance   236.352    5.211   227.459   232.655   235.879   239.486   247.815
#>           Rhat n.eff
#> N[1]     1.001  6900
#> N[2]     1.001 17000
#> N[3]     1.001  5400
#> N[4]     1.001 27000
#> N[5]     1.002  1200
#> N[6]     1.003   700
#> N[7]     1.004   440
#> N[8]     1.001 14000
#> N[9]     1.001 13000
#> N[10]    1.001 27000
#> N[11]    1.001 11000
#> N[12]    1.001 27000
#> N[13]    1.001 27000
#> N[14]    1.001 12000
#> N[15]    1.001 27000
#> meanr    1.001 13000
#> r[1]     1.001 27000
#> r[2]     1.001 13000
#> r[3]     1.001 15000
#> r[4]     1.002  1500
#> r[5]     1.002  1900
#> r[6]     1.002  2600
#> r[7]     1.007   270
#> r[8]     1.001 27000
#> r[9]     1.001 24000
#> r[10]    1.001  4700
#> r[11]    1.001  4800
#> r[12]    1.001 27000
#> r[13]    1.001 27000
#> r[14]    1.001 13000
#> sdr      1.001  7400
#> vrrmax   1.001  7400
#> deviance 1.001  8600
#> 
#> For each parameter, n.eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor (at convergence, Rhat=1).
#> 
#> DIC info (using the rule, pD = var(deviance)/2)
#> pD = 13.6 and DIC = 249.9
#> DIC is an estimate of expected predictive error (lower deviance is better).
#> 
# }