Import a list of BUGS outputs previously exported

This function imports a list of BUGS outputs previously exported by fit_trend(). Users can import one, several, or all models.

read_bugs(series = NULL, path = ".")

Arguments

series

a vector of character strings. One or several count series names. If NULL (default) BUGS outputs for all count series will be imported. Users can run list_series() to get the correct spelling of count series names.

path

a character string. The directory in which BUGS outputs have been saved by the function fit_trend().

Value

An n-element list (where n is the number of count series). See fit_trend() for further information.

Examples

## Load Garamba raw dataset ----
file_path <- system.file("extdata", "garamba_survey.csv", 
                         package = "popbayes")
                         
garamba <- read.csv(file = file_path)

## Create temporary folder ----
temp_path <- tempdir()

## Format dataset ----
garamba_formatted <- popbayes::format_data(
  data              = garamba, 
  path              = temp_path,
  field_method      = "field_method",
  pref_field_method = "pref_field_method",
  conversion_A2G    = "conversion_A2G",
  rmax              = "rmax")
#> ✔ Detecting 10 count series

## Select one serie ----
a_buselaphus <- popbayes::filter_series(garamba_formatted, 
                                        location = "Garamba",
                                        species  = "Alcelaphus buselaphus")
#> ✔ Found 1 series with 'Alcelaphus buselaphus' and 'Garamba'.
# \donttest{
## Fit population trends (requires JAGS) ----
a_buselaphus_mod <- popbayes::fit_trend(a_buselaphus, path = temp_path)
#> Compiling data graph
#>    Resolving undeclared variables
#>    Allocating nodes
#>    Initializing
#>    Reading data back into data table
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 15
#>    Unobserved stochastic nodes: 15
#>    Total graph size: 227
#> 
#> Initializing model
#> 

## Import BUGS outputs for one count series ----
popbayes::read_bugs(series = "garamba__alcelaphus_buselaphus", 
                    path   = temp_path)
#> $garamba__alcelaphus_buselaphus
#> Inference for Bugs model at "model.bug", fit using jags,
#>  2 chains, each with 50000 iterations (first 10000 discarded), n.thin = 3
#>  n.sims = 26666 iterations saved
#>            mu.vect  sd.vect      2.5%       25%       50%       75%     97.5%
#> N[1]     17644.554 1715.392 14374.291 16450.572 17619.062 18805.399 21063.906
#> N[2]      3881.083  501.308  2937.984  3539.917  3863.400  4206.890  4915.753
#> N[3]      3335.215  386.961  2581.571  3073.678  3329.665  3592.628  4099.286
#> N[4]      3021.546  426.238  2223.121  2728.193  3008.142  3297.542  3911.779
#> N[5]      2669.300  329.218  2028.427  2450.870  2664.601  2889.550  3325.059
#> N[6]      3197.981  546.094  2226.829  2812.473  3167.461  3550.650  4351.792
#> N[7]      3719.958  679.798  2516.498  3241.490  3672.621  4157.556  5164.870
#> N[8]      3463.835  371.710  2758.981  3209.014  3456.248  3714.347  4208.222
#> N[9]      2884.375  205.496  2487.714  2743.928  2881.634  3022.412  3294.304
#> N[10]     2755.467  204.691  2357.056  2617.848  2754.033  2893.760  3160.728
#> N[11]     2760.493  259.728  2261.818  2582.629  2757.353  2934.175  3275.384
#> N[12]     2630.859  299.806  2063.785  2424.011  2622.668  2828.968  3234.892
#> N[13]     1310.488   73.476  1168.460  1260.644  1310.077  1359.664  1454.222
#> N[14]     1570.846   89.934  1397.722  1509.437  1570.312  1631.329  1749.443
#> N[15]     2393.229  152.119  2095.372  2290.627  2391.478  2494.557  2691.737
#> meanr       -0.049    0.003    -0.054    -0.051    -0.049    -0.047    -0.043
#> r[1]        -0.217    0.023    -0.262    -0.232    -0.216    -0.202    -0.174
#> r[2]        -0.150    0.070    -0.287    -0.197    -0.149    -0.103    -0.014
#> r[3]        -0.051    0.061    -0.173    -0.091    -0.050    -0.010     0.066
#> r[4]        -0.024    0.033    -0.090    -0.047    -0.024    -0.002     0.040
#> r[5]         0.087    0.064    -0.042     0.044     0.087     0.131     0.214
#> r[6]         0.075    0.058    -0.038     0.036     0.074     0.113     0.189
#> r[7]        -0.020    0.052    -0.119    -0.056    -0.022     0.015     0.085
#> r[8]        -0.090    0.051    -0.187    -0.125    -0.090    -0.056     0.012
#> r[9]        -0.023    0.044    -0.109    -0.052    -0.023     0.007     0.061
#> r[10]        0.000    0.067    -0.132    -0.045     0.001     0.046     0.128
#> r[11]       -0.050    0.066    -0.180    -0.094    -0.050    -0.006     0.076
#> r[12]       -0.086    0.015    -0.116    -0.097    -0.087    -0.076    -0.055
#> r[13]        0.091    0.035     0.022     0.066     0.091     0.114     0.161
#> r[14]        0.140    0.027     0.086     0.122     0.140     0.159     0.194
#> sdr          0.112    0.009     0.094     0.105     0.111     0.118     0.131
#> vrrmax       0.459    0.038     0.388     0.433     0.458     0.484     0.540
#> deviance   236.252    5.174   227.494   232.558   235.793   239.378   247.781
#>           Rhat n.eff
#> N[1]     1.001 16000
#> N[2]     1.002  1600
#> N[3]     1.001  4300
#> N[4]     1.001 27000
#> N[5]     1.001  9000
#> N[6]     1.002  1500
#> N[7]     1.004   510
#> N[8]     1.001  6400
#> N[9]     1.001 27000
#> N[10]    1.001 27000
#> N[11]    1.001 27000
#> N[12]    1.001 13000
#> N[13]    1.001 13000
#> N[14]    1.001 18000
#> N[15]    1.001  6800
#> meanr    1.001 27000
#> r[1]     1.002  1600
#> r[2]     1.002  2400
#> r[3]     1.001  3500
#> r[4]     1.001 24000
#> r[5]     1.002  1700
#> r[6]     1.003   960
#> r[7]     1.004   530
#> r[8]     1.001  3300
#> r[9]     1.001 27000
#> r[10]    1.001  6200
#> r[11]    1.001 13000
#> r[12]    1.001 27000
#> r[13]    1.001  6000
#> r[14]    1.001  5300
#> sdr      1.003   870
#> vrrmax   1.003   870
#> deviance 1.002  1900
#> 
#> For each parameter, n.eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor (at convergence, Rhat=1).
#> 
#> DIC info (using the rule, pD = var(deviance)/2)
#> pD = 13.4 and DIC = 249.6
#> DIC is an estimate of expected predictive error (lower deviance is better).
#> 

## Import BUGS outputs for all count series ----
popbayes::read_bugs(path = temp_path)
#> $garamba__alcelaphus_buselaphus
#> Inference for Bugs model at "model.bug", fit using jags,
#>  2 chains, each with 50000 iterations (first 10000 discarded), n.thin = 3
#>  n.sims = 26666 iterations saved
#>            mu.vect  sd.vect      2.5%       25%       50%       75%     97.5%
#> N[1]     17644.554 1715.392 14374.291 16450.572 17619.062 18805.399 21063.906
#> N[2]      3881.083  501.308  2937.984  3539.917  3863.400  4206.890  4915.753
#> N[3]      3335.215  386.961  2581.571  3073.678  3329.665  3592.628  4099.286
#> N[4]      3021.546  426.238  2223.121  2728.193  3008.142  3297.542  3911.779
#> N[5]      2669.300  329.218  2028.427  2450.870  2664.601  2889.550  3325.059
#> N[6]      3197.981  546.094  2226.829  2812.473  3167.461  3550.650  4351.792
#> N[7]      3719.958  679.798  2516.498  3241.490  3672.621  4157.556  5164.870
#> N[8]      3463.835  371.710  2758.981  3209.014  3456.248  3714.347  4208.222
#> N[9]      2884.375  205.496  2487.714  2743.928  2881.634  3022.412  3294.304
#> N[10]     2755.467  204.691  2357.056  2617.848  2754.033  2893.760  3160.728
#> N[11]     2760.493  259.728  2261.818  2582.629  2757.353  2934.175  3275.384
#> N[12]     2630.859  299.806  2063.785  2424.011  2622.668  2828.968  3234.892
#> N[13]     1310.488   73.476  1168.460  1260.644  1310.077  1359.664  1454.222
#> N[14]     1570.846   89.934  1397.722  1509.437  1570.312  1631.329  1749.443
#> N[15]     2393.229  152.119  2095.372  2290.627  2391.478  2494.557  2691.737
#> meanr       -0.049    0.003    -0.054    -0.051    -0.049    -0.047    -0.043
#> r[1]        -0.217    0.023    -0.262    -0.232    -0.216    -0.202    -0.174
#> r[2]        -0.150    0.070    -0.287    -0.197    -0.149    -0.103    -0.014
#> r[3]        -0.051    0.061    -0.173    -0.091    -0.050    -0.010     0.066
#> r[4]        -0.024    0.033    -0.090    -0.047    -0.024    -0.002     0.040
#> r[5]         0.087    0.064    -0.042     0.044     0.087     0.131     0.214
#> r[6]         0.075    0.058    -0.038     0.036     0.074     0.113     0.189
#> r[7]        -0.020    0.052    -0.119    -0.056    -0.022     0.015     0.085
#> r[8]        -0.090    0.051    -0.187    -0.125    -0.090    -0.056     0.012
#> r[9]        -0.023    0.044    -0.109    -0.052    -0.023     0.007     0.061
#> r[10]        0.000    0.067    -0.132    -0.045     0.001     0.046     0.128
#> r[11]       -0.050    0.066    -0.180    -0.094    -0.050    -0.006     0.076
#> r[12]       -0.086    0.015    -0.116    -0.097    -0.087    -0.076    -0.055
#> r[13]        0.091    0.035     0.022     0.066     0.091     0.114     0.161
#> r[14]        0.140    0.027     0.086     0.122     0.140     0.159     0.194
#> sdr          0.112    0.009     0.094     0.105     0.111     0.118     0.131
#> vrrmax       0.459    0.038     0.388     0.433     0.458     0.484     0.540
#> deviance   236.252    5.174   227.494   232.558   235.793   239.378   247.781
#>           Rhat n.eff
#> N[1]     1.001 16000
#> N[2]     1.002  1600
#> N[3]     1.001  4300
#> N[4]     1.001 27000
#> N[5]     1.001  9000
#> N[6]     1.002  1500
#> N[7]     1.004   510
#> N[8]     1.001  6400
#> N[9]     1.001 27000
#> N[10]    1.001 27000
#> N[11]    1.001 27000
#> N[12]    1.001 13000
#> N[13]    1.001 13000
#> N[14]    1.001 18000
#> N[15]    1.001  6800
#> meanr    1.001 27000
#> r[1]     1.002  1600
#> r[2]     1.002  2400
#> r[3]     1.001  3500
#> r[4]     1.001 24000
#> r[5]     1.002  1700
#> r[6]     1.003   960
#> r[7]     1.004   530
#> r[8]     1.001  3300
#> r[9]     1.001 27000
#> r[10]    1.001  6200
#> r[11]    1.001 13000
#> r[12]    1.001 27000
#> r[13]    1.001  6000
#> r[14]    1.001  5300
#> sdr      1.003   870
#> vrrmax   1.003   870
#> deviance 1.002  1900
#> 
#> For each parameter, n.eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor (at convergence, Rhat=1).
#> 
#> DIC info (using the rule, pD = var(deviance)/2)
#> pD = 13.4 and DIC = 249.6
#> DIC is an estimate of expected predictive error (lower deviance is better).
#> 
# }