Import a list of BUGS outputs previously exported

This function imports a list of BUGS outputs previously exported by fit_trend(). Users can import one, several, or all models.

read_bugs(series = NULL, path = ".")

Arguments

series

a vector of character strings. One or several count series names. If NULL (default) BUGS outputs for all count series will be imported. Users can run list_series() to get the correct spelling of count series names.

path

a character string. The directory in which BUGS outputs have been saved by the function fit_trend().

Value

An n-element list (where n is the number of count series). See fit_trend() for further information.

Examples

## Load Garamba raw dataset ----
file_path <- system.file("extdata", "garamba_survey.csv", 
                         package = "popbayes")
                         
garamba <- read.csv(file = file_path)

## Create temporary folder ----
temp_path <- tempdir()

## Format dataset ----
garamba_formatted <- popbayes::format_data(garamba, path = temp_path)
#> ✔ Conversion data found in 'data'.
#> ✔ Detecting 10 count series

## Select one serie ----
a_buselaphus <- popbayes::filter_series(garamba_formatted, 
                                        location = "Garamba",
                                        species  = "Alcelaphus buselaphus")
#> ✔ Found 1 series with 'Alcelaphus buselaphus' and 'Garamba'.
# \donttest{
## Fit population trends (requires JAGS) ----
a_buselaphus_mod <- popbayes::fit_trend(a_buselaphus, path = temp_path)
#> Compiling data graph
#>    Resolving undeclared variables
#>    Allocating nodes
#>    Initializing
#>    Reading data back into data table
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 15
#>    Unobserved stochastic nodes: 15
#>    Total graph size: 227
#> 
#> Initializing model
#> 

## Import BUGS outputs for one count series ----
popbayes::read_bugs(series = "garamba__alcelaphus_buselaphus", 
                    path   = temp_path)
#> $garamba__alcelaphus_buselaphus
#> Inference for Bugs model at "model.bug", fit using jags,
#>  2 chains, each with 50000 iterations (first 10000 discarded), n.thin = 3
#>  n.sims = 26666 iterations saved
#>            mu.vect  sd.vect      2.5%       25%       50%       75%     97.5%
#> N[1]     17693.833 1718.077 14318.018 16549.879 17695.430 18847.151 21063.777
#> N[2]      3891.334  495.411  2959.654  3548.046  3882.006  4216.805  4881.814
#> N[3]      3342.181  383.888  2602.837  3079.466  3337.225  3595.668  4105.253
#> N[4]      3032.074  431.554  2255.972  2727.318  3008.611  3312.052  3930.056
#> N[5]      2662.314  322.620  2045.393  2442.665  2658.298  2877.588  3308.681
#> N[6]      3198.193  534.318  2240.028  2818.643  3169.216  3549.839  4321.805
#> N[7]      3731.681  666.341  2556.447  3258.892  3684.828  4161.398  5148.666
#> N[8]      3466.778  377.498  2753.475  3209.541  3457.866  3715.885  4228.615
#> N[9]      2888.785  206.524  2492.180  2748.554  2886.318  3027.012  3300.993
#> N[10]     2747.021  206.473  2346.531  2606.596  2745.740  2886.494  3155.859
#> N[11]     2751.660  258.728  2254.887  2576.689  2747.793  2923.179  3267.391
#> N[12]     2623.443  295.757  2061.152  2419.757  2616.037  2818.567  3220.716
#> N[13]     1311.168   73.602  1168.381  1260.821  1311.659  1360.697  1454.710
#> N[14]     1571.292   90.384  1398.542  1509.019  1570.959  1632.094  1749.307
#> N[15]     2394.436  153.222  2097.756  2290.757  2393.497  2496.879  2699.965
#> meanr       -0.049    0.003    -0.054    -0.051    -0.049    -0.047    -0.043
#> r[1]        -0.217    0.022    -0.262    -0.231    -0.216    -0.202    -0.174
#> r[2]        -0.151    0.069    -0.284    -0.197    -0.151    -0.105    -0.013
#> r[3]        -0.050    0.060    -0.168    -0.091    -0.051    -0.010     0.066
#> r[4]        -0.025    0.033    -0.091    -0.048    -0.026    -0.003     0.040
#> r[5]         0.088    0.064    -0.037     0.045     0.089     0.132     0.216
#> r[6]         0.076    0.058    -0.039     0.038     0.077     0.115     0.188
#> r[7]        -0.021    0.051    -0.118    -0.057    -0.022     0.013     0.081
#> r[8]        -0.089    0.051    -0.189    -0.124    -0.090    -0.056     0.014
#> r[9]        -0.025    0.044    -0.112    -0.055    -0.025     0.005     0.061
#> r[10]        0.000    0.066    -0.132    -0.044     0.001     0.045     0.128
#> r[11]       -0.050    0.065    -0.177    -0.093    -0.049    -0.006     0.077
#> r[12]       -0.086    0.015    -0.115    -0.097    -0.086    -0.076    -0.055
#> r[13]        0.090    0.036     0.020     0.066     0.090     0.114     0.161
#> r[14]        0.140    0.028     0.086     0.122     0.140     0.159     0.195
#> sdr          0.112    0.009     0.094     0.105     0.111     0.118     0.131
#> vrrmax       0.459    0.038     0.388     0.434     0.458     0.484     0.538
#> deviance   236.208    5.172   227.267   232.557   235.758   239.424   247.579
#>           Rhat n.eff
#> N[1]     1.001  9100
#> N[2]     1.001 27000
#> N[3]     1.001 27000
#> N[4]     1.001 11000
#> N[5]     1.002  1600
#> N[6]     1.001  4100
#> N[7]     1.002  2000
#> N[8]     1.002  2000
#> N[9]     1.001 12000
#> N[10]    1.001 27000
#> N[11]    1.001 27000
#> N[12]    1.001 27000
#> N[13]    1.001 14000
#> N[14]    1.001 27000
#> N[15]    1.001 27000
#> meanr    1.001 27000
#> r[1]     1.001 27000
#> r[2]     1.001 27000
#> r[3]     1.001 21000
#> r[4]     1.002  1500
#> r[5]     1.004   500
#> r[6]     1.001  6900
#> r[7]     1.001  9600
#> r[8]     1.002  1100
#> r[9]     1.001 27000
#> r[10]    1.001 27000
#> r[11]    1.001 27000
#> r[12]    1.001 27000
#> r[13]    1.001 27000
#> r[14]    1.001 26000
#> sdr      1.003   890
#> vrrmax   1.003   890
#> deviance 1.001  3500
#> 
#> For each parameter, n.eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor (at convergence, Rhat=1).
#> 
#> DIC info (using the rule, pD = var(deviance)/2)
#> pD = 13.4 and DIC = 249.6
#> DIC is an estimate of expected predictive error (lower deviance is better).
#> 

## Import BUGS outputs for all count series ----
popbayes::read_bugs(path = temp_path)
#> $garamba__alcelaphus_buselaphus
#> Inference for Bugs model at "model.bug", fit using jags,
#>  2 chains, each with 50000 iterations (first 10000 discarded), n.thin = 3
#>  n.sims = 26666 iterations saved
#>            mu.vect  sd.vect      2.5%       25%       50%       75%     97.5%
#> N[1]     17693.833 1718.077 14318.018 16549.879 17695.430 18847.151 21063.777
#> N[2]      3891.334  495.411  2959.654  3548.046  3882.006  4216.805  4881.814
#> N[3]      3342.181  383.888  2602.837  3079.466  3337.225  3595.668  4105.253
#> N[4]      3032.074  431.554  2255.972  2727.318  3008.611  3312.052  3930.056
#> N[5]      2662.314  322.620  2045.393  2442.665  2658.298  2877.588  3308.681
#> N[6]      3198.193  534.318  2240.028  2818.643  3169.216  3549.839  4321.805
#> N[7]      3731.681  666.341  2556.447  3258.892  3684.828  4161.398  5148.666
#> N[8]      3466.778  377.498  2753.475  3209.541  3457.866  3715.885  4228.615
#> N[9]      2888.785  206.524  2492.180  2748.554  2886.318  3027.012  3300.993
#> N[10]     2747.021  206.473  2346.531  2606.596  2745.740  2886.494  3155.859
#> N[11]     2751.660  258.728  2254.887  2576.689  2747.793  2923.179  3267.391
#> N[12]     2623.443  295.757  2061.152  2419.757  2616.037  2818.567  3220.716
#> N[13]     1311.168   73.602  1168.381  1260.821  1311.659  1360.697  1454.710
#> N[14]     1571.292   90.384  1398.542  1509.019  1570.959  1632.094  1749.307
#> N[15]     2394.436  153.222  2097.756  2290.757  2393.497  2496.879  2699.965
#> meanr       -0.049    0.003    -0.054    -0.051    -0.049    -0.047    -0.043
#> r[1]        -0.217    0.022    -0.262    -0.231    -0.216    -0.202    -0.174
#> r[2]        -0.151    0.069    -0.284    -0.197    -0.151    -0.105    -0.013
#> r[3]        -0.050    0.060    -0.168    -0.091    -0.051    -0.010     0.066
#> r[4]        -0.025    0.033    -0.091    -0.048    -0.026    -0.003     0.040
#> r[5]         0.088    0.064    -0.037     0.045     0.089     0.132     0.216
#> r[6]         0.076    0.058    -0.039     0.038     0.077     0.115     0.188
#> r[7]        -0.021    0.051    -0.118    -0.057    -0.022     0.013     0.081
#> r[8]        -0.089    0.051    -0.189    -0.124    -0.090    -0.056     0.014
#> r[9]        -0.025    0.044    -0.112    -0.055    -0.025     0.005     0.061
#> r[10]        0.000    0.066    -0.132    -0.044     0.001     0.045     0.128
#> r[11]       -0.050    0.065    -0.177    -0.093    -0.049    -0.006     0.077
#> r[12]       -0.086    0.015    -0.115    -0.097    -0.086    -0.076    -0.055
#> r[13]        0.090    0.036     0.020     0.066     0.090     0.114     0.161
#> r[14]        0.140    0.028     0.086     0.122     0.140     0.159     0.195
#> sdr          0.112    0.009     0.094     0.105     0.111     0.118     0.131
#> vrrmax       0.459    0.038     0.388     0.434     0.458     0.484     0.538
#> deviance   236.208    5.172   227.267   232.557   235.758   239.424   247.579
#>           Rhat n.eff
#> N[1]     1.001  9100
#> N[2]     1.001 27000
#> N[3]     1.001 27000
#> N[4]     1.001 11000
#> N[5]     1.002  1600
#> N[6]     1.001  4100
#> N[7]     1.002  2000
#> N[8]     1.002  2000
#> N[9]     1.001 12000
#> N[10]    1.001 27000
#> N[11]    1.001 27000
#> N[12]    1.001 27000
#> N[13]    1.001 14000
#> N[14]    1.001 27000
#> N[15]    1.001 27000
#> meanr    1.001 27000
#> r[1]     1.001 27000
#> r[2]     1.001 27000
#> r[3]     1.001 21000
#> r[4]     1.002  1500
#> r[5]     1.004   500
#> r[6]     1.001  6900
#> r[7]     1.001  9600
#> r[8]     1.002  1100
#> r[9]     1.001 27000
#> r[10]    1.001 27000
#> r[11]    1.001 27000
#> r[12]    1.001 27000
#> r[13]    1.001 27000
#> r[14]    1.001 26000
#> sdr      1.003   890
#> vrrmax   1.003   890
#> deviance 1.001  3500
#> 
#> For each parameter, n.eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor (at convergence, Rhat=1).
#> 
#> DIC info (using the rule, pD = var(deviance)/2)
#> pD = 13.4 and DIC = 249.6
#> DIC is an estimate of expected predictive error (lower deviance is better).
#> 
# }