Import a list of BUGS outputs previously exported

This function imports a list of BUGS outputs previously exported by fit_trend(). Users can import one, several, or all models.

read_bugs(series = NULL, path = ".")

Arguments

series

a vector of character strings. One or several count series names. If NULL (default) BUGS outputs for all count series will be imported. Users can run list_series() to get the correct spelling of count series names.

path

a character string. The directory in which BUGS outputs have been saved by the function fit_trend().

Value

An n-element list (where n is the number of count series). See fit_trend() for further information.

Examples

## Load Garamba raw dataset ----
file_path <- system.file("extdata", "garamba_survey.csv", 
                         package = "popbayes")
                         
garamba <- read.csv(file = file_path)

## Create temporary folder ----
temp_path <- tempdir()

## Format dataset ----
garamba_formatted <- popbayes::format_data(
  data              = garamba, 
  path              = temp_path,
  field_method      = "field_method",
  pref_field_method = "pref_field_method",
  conversion_A2G    = "conversion_A2G",
  rmax              = "rmax")
#> ✔ Detecting 10 count series

## Select one serie ----
a_buselaphus <- popbayes::filter_series(garamba_formatted, 
                                        location = "Garamba",
                                        species  = "Alcelaphus buselaphus")
#> ✔ Found 1 series with 'Alcelaphus buselaphus' and 'Garamba'.
# \donttest{
## Fit population trends (requires JAGS) ----
a_buselaphus_mod <- popbayes::fit_trend(a_buselaphus, path = temp_path)
#> Compiling data graph
#>    Resolving undeclared variables
#>    Allocating nodes
#>    Initializing
#>    Reading data back into data table
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 15
#>    Unobserved stochastic nodes: 15
#>    Total graph size: 227
#> 
#> Initializing model
#> 

## Import BUGS outputs for one count series ----
popbayes::read_bugs(series = "garamba__alcelaphus_buselaphus", 
                    path   = temp_path)
#> $garamba__alcelaphus_buselaphus
#> Inference for Bugs model at "model.bug", fit using jags,
#>  2 chains, each with 50000 iterations (first 10000 discarded), n.thin = 3
#>  n.sims = 26666 iterations saved
#>            mu.vect  sd.vect      2.5%       25%       50%       75%     97.5%
#> N[1]     17661.401 1693.917 14392.262 16504.810 17663.979 18796.966 21004.947
#> N[2]      3880.066  490.899  2964.074  3537.990  3866.842  4205.265  4879.447
#> N[3]      3329.854  382.309  2592.360  3071.695  3326.870  3584.333  4085.362
#> N[4]      3006.045  428.694  2214.570  2708.222  2990.010  3285.981  3878.534
#> N[5]      2673.199  329.408  2039.011  2449.259  2667.701  2893.516  3337.452
#> N[6]      3226.636  548.032  2265.517  2840.543  3185.081  3565.972  4410.007
#> N[7]      3752.527  687.021  2575.552  3263.989  3696.129  4177.339  5257.520
#> N[8]      3473.753  370.520  2763.540  3219.575  3471.466  3722.568  4210.008
#> N[9]      2886.372  206.824  2486.637  2744.886  2884.229  3024.577  3295.868
#> N[10]     2748.902  205.055  2356.708  2609.518  2748.164  2885.135  3157.853
#> N[11]     2754.622  258.787  2258.755  2577.650  2750.206  2926.205  3282.283
#> N[12]     2626.778  297.028  2078.697  2422.069  2616.438  2822.624  3240.976
#> N[13]     1311.116   73.444  1167.679  1261.388  1310.880  1360.213  1456.362
#> N[14]     1571.175   89.836  1395.467  1511.042  1570.356  1630.707  1749.075
#> N[15]     2394.884  152.838  2096.410  2292.255  2395.074  2497.931  2694.965
#> meanr       -0.049    0.003    -0.054    -0.051    -0.049    -0.047    -0.043
#> r[1]        -0.217    0.022    -0.260    -0.232    -0.217    -0.202    -0.174
#> r[2]        -0.152    0.068    -0.285    -0.198    -0.151    -0.105    -0.018
#> r[3]        -0.053    0.059    -0.170    -0.093    -0.053    -0.012     0.062
#> r[4]        -0.023    0.034    -0.088    -0.046    -0.023    -0.001     0.044
#> r[5]         0.091    0.063    -0.032     0.048     0.091     0.132     0.220
#> r[6]         0.074    0.058    -0.036     0.034     0.074     0.113     0.189
#> r[7]        -0.022    0.052    -0.125    -0.057    -0.022     0.013     0.081
#> r[8]        -0.091    0.051    -0.189    -0.126    -0.092    -0.057     0.010
#> r[9]        -0.025    0.044    -0.111    -0.054    -0.024     0.005     0.060
#> r[10]        0.000    0.067    -0.130    -0.045     0.001     0.046     0.130
#> r[11]       -0.050    0.066    -0.180    -0.094    -0.049    -0.004     0.077
#> r[12]       -0.086    0.015    -0.116    -0.097    -0.086    -0.076    -0.056
#> r[13]        0.090    0.035     0.021     0.067     0.090     0.114     0.160
#> r[14]        0.140    0.027     0.086     0.122     0.140     0.158     0.194
#> sdr          0.112    0.009     0.094     0.105     0.111     0.118     0.131
#> vrrmax       0.460    0.039     0.389     0.433     0.459     0.485     0.540
#> deviance   236.139    5.095   227.273   232.541   235.758   239.296   247.209
#>           Rhat n.eff
#> N[1]     1.001 27000
#> N[2]     1.002  1300
#> N[3]     1.001  6300
#> N[4]     1.001  4700
#> N[5]     1.001  8300
#> N[6]     1.001  6500
#> N[7]     1.003   810
#> N[8]     1.001  3200
#> N[9]     1.001 27000
#> N[10]    1.001  6800
#> N[11]    1.001  5200
#> N[12]    1.002  2400
#> N[13]    1.001 27000
#> N[14]    1.001 27000
#> N[15]    1.001 27000
#> meanr    1.001 27000
#> r[1]     1.002  1900
#> r[2]     1.002  1100
#> r[3]     1.002  1100
#> r[4]     1.002  2300
#> r[5]     1.001 27000
#> r[6]     1.003   720
#> r[7]     1.002  1200
#> r[8]     1.002  2600
#> r[9]     1.001 12000
#> r[10]    1.001 27000
#> r[11]    1.001  4400
#> r[12]    1.002  2400
#> r[13]    1.001 27000
#> r[14]    1.001 27000
#> sdr      1.001 27000
#> vrrmax   1.001 27000
#> deviance 1.001 27000
#> 
#> For each parameter, n.eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor (at convergence, Rhat=1).
#> 
#> DIC info (using the rule, pD = var(deviance)/2)
#> pD = 13.0 and DIC = 249.1
#> DIC is an estimate of expected predictive error (lower deviance is better).
#> 

## Import BUGS outputs for all count series ----
popbayes::read_bugs(path = temp_path)
#> $garamba__alcelaphus_buselaphus
#> Inference for Bugs model at "model.bug", fit using jags,
#>  2 chains, each with 50000 iterations (first 10000 discarded), n.thin = 3
#>  n.sims = 26666 iterations saved
#>            mu.vect  sd.vect      2.5%       25%       50%       75%     97.5%
#> N[1]     17661.401 1693.917 14392.262 16504.810 17663.979 18796.966 21004.947
#> N[2]      3880.066  490.899  2964.074  3537.990  3866.842  4205.265  4879.447
#> N[3]      3329.854  382.309  2592.360  3071.695  3326.870  3584.333  4085.362
#> N[4]      3006.045  428.694  2214.570  2708.222  2990.010  3285.981  3878.534
#> N[5]      2673.199  329.408  2039.011  2449.259  2667.701  2893.516  3337.452
#> N[6]      3226.636  548.032  2265.517  2840.543  3185.081  3565.972  4410.007
#> N[7]      3752.527  687.021  2575.552  3263.989  3696.129  4177.339  5257.520
#> N[8]      3473.753  370.520  2763.540  3219.575  3471.466  3722.568  4210.008
#> N[9]      2886.372  206.824  2486.637  2744.886  2884.229  3024.577  3295.868
#> N[10]     2748.902  205.055  2356.708  2609.518  2748.164  2885.135  3157.853
#> N[11]     2754.622  258.787  2258.755  2577.650  2750.206  2926.205  3282.283
#> N[12]     2626.778  297.028  2078.697  2422.069  2616.438  2822.624  3240.976
#> N[13]     1311.116   73.444  1167.679  1261.388  1310.880  1360.213  1456.362
#> N[14]     1571.175   89.836  1395.467  1511.042  1570.356  1630.707  1749.075
#> N[15]     2394.884  152.838  2096.410  2292.255  2395.074  2497.931  2694.965
#> meanr       -0.049    0.003    -0.054    -0.051    -0.049    -0.047    -0.043
#> r[1]        -0.217    0.022    -0.260    -0.232    -0.217    -0.202    -0.174
#> r[2]        -0.152    0.068    -0.285    -0.198    -0.151    -0.105    -0.018
#> r[3]        -0.053    0.059    -0.170    -0.093    -0.053    -0.012     0.062
#> r[4]        -0.023    0.034    -0.088    -0.046    -0.023    -0.001     0.044
#> r[5]         0.091    0.063    -0.032     0.048     0.091     0.132     0.220
#> r[6]         0.074    0.058    -0.036     0.034     0.074     0.113     0.189
#> r[7]        -0.022    0.052    -0.125    -0.057    -0.022     0.013     0.081
#> r[8]        -0.091    0.051    -0.189    -0.126    -0.092    -0.057     0.010
#> r[9]        -0.025    0.044    -0.111    -0.054    -0.024     0.005     0.060
#> r[10]        0.000    0.067    -0.130    -0.045     0.001     0.046     0.130
#> r[11]       -0.050    0.066    -0.180    -0.094    -0.049    -0.004     0.077
#> r[12]       -0.086    0.015    -0.116    -0.097    -0.086    -0.076    -0.056
#> r[13]        0.090    0.035     0.021     0.067     0.090     0.114     0.160
#> r[14]        0.140    0.027     0.086     0.122     0.140     0.158     0.194
#> sdr          0.112    0.009     0.094     0.105     0.111     0.118     0.131
#> vrrmax       0.460    0.039     0.389     0.433     0.459     0.485     0.540
#> deviance   236.139    5.095   227.273   232.541   235.758   239.296   247.209
#>           Rhat n.eff
#> N[1]     1.001 27000
#> N[2]     1.002  1300
#> N[3]     1.001  6300
#> N[4]     1.001  4700
#> N[5]     1.001  8300
#> N[6]     1.001  6500
#> N[7]     1.003   810
#> N[8]     1.001  3200
#> N[9]     1.001 27000
#> N[10]    1.001  6800
#> N[11]    1.001  5200
#> N[12]    1.002  2400
#> N[13]    1.001 27000
#> N[14]    1.001 27000
#> N[15]    1.001 27000
#> meanr    1.001 27000
#> r[1]     1.002  1900
#> r[2]     1.002  1100
#> r[3]     1.002  1100
#> r[4]     1.002  2300
#> r[5]     1.001 27000
#> r[6]     1.003   720
#> r[7]     1.002  1200
#> r[8]     1.002  2600
#> r[9]     1.001 12000
#> r[10]    1.001 27000
#> r[11]    1.001  4400
#> r[12]    1.002  2400
#> r[13]    1.001 27000
#> r[14]    1.001 27000
#> sdr      1.001 27000
#> vrrmax   1.001 27000
#> deviance 1.001 27000
#> 
#> For each parameter, n.eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor (at convergence, Rhat=1).
#> 
#> DIC info (using the rule, pD = var(deviance)/2)
#> pD = 13.0 and DIC = 249.1
#> DIC is an estimate of expected predictive error (lower deviance is better).
#> 
# }