Import a list of BUGS outputs previously exported

This function imports a list of BUGS outputs previously exported by fit_trend(). Users can import one, several, or all models.

read_bugs(series = NULL, path = ".")

Arguments

series
a vector of character strings. One or several count series names. If NULL (default) BUGS outputs for all count series will be imported. Users can run list_series() to get the correct spelling of count series names.

path
a character string. The directory in which BUGS outputs have been saved by the function fit_trend().

Value

An n-element list (where n is the number of count series). See fit_trend() for further information.

Examples

## Load Garamba raw dataset ----
file_path <- system.file("extdata", "garamba_survey.csv", 
                         package = "popbayes")
                         
garamba <- read.csv(file = file_path)

## Create temporary folder ----
temp_path <- tempdir()

## Format dataset ----
garamba_formatted <- popbayes::format_data(garamba, path = temp_path)
#> ✔ Conversion data found in 'data'.
#> ✔ Detecting 10 count series

## Select one serie ----
a_buselaphus <- popbayes::filter_series(garamba_formatted, 
                                        location = "Garamba",
                                        species  = "Alcelaphus buselaphus")
#> ✔ Found 1 series with 'Alcelaphus buselaphus' and 'Garamba'.
# \donttest{
## Fit population trends (requires JAGS) ----
a_buselaphus_mod <- popbayes::fit_trend(a_buselaphus, path = temp_path)
#> Compiling data graph
#>    Resolving undeclared variables
#>    Allocating nodes
#>    Initializing
#>    Reading data back into data table
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 15
#>    Unobserved stochastic nodes: 15
#>    Total graph size: 227
#> 
#> Initializing model
#> 

## Import BUGS outputs for one count series ----
popbayes::read_bugs(series = "garamba__alcelaphus_buselaphus", 
                    path   = temp_path)
#> $garamba__alcelaphus_buselaphus
#> Inference for Bugs model at "model.bug", fit using jags,
#>  2 chains, each with 50000 iterations (first 10000 discarded), n.thin = 3
#>  n.sims = 26666 iterations saved
#>            mu.vect  sd.vect      2.5%       25%       50%       75%     97.5%
#> N[1]     17685.981 1733.998 14277.150 16508.855 17717.666 18846.160 21063.507
#> N[2]      3865.329  512.525  2891.789  3516.437  3855.529  4198.655  4906.995
#> N[3]      3322.320  398.899  2549.491  3057.826  3321.147  3586.658  4113.046
#> N[4]      3010.473  427.929  2217.944  2713.434  2995.194  3289.426  3882.440
#> N[5]      2675.482  331.545  2041.604  2447.240  2671.309  2900.134  3330.760
#> N[6]      3224.325  544.322  2248.000  2839.607  3189.280  3573.923  4389.428
#> N[7]      3755.939  668.736  2559.061  3281.074  3721.329  4188.211  5163.674
#> N[8]      3483.815  365.680  2799.321  3230.259  3478.492  3728.039  4214.172
#> N[9]      2888.850  206.927  2490.742  2748.531  2888.379  3027.213  3301.309
#> N[10]     2748.587  206.509  2349.815  2607.414  2747.427  2886.491  3157.499
#> N[11]     2753.478  260.659  2257.112  2573.372  2749.665  2926.622  3270.544
#> N[12]     2624.659  300.613  2061.996  2412.867  2615.560  2824.720  3234.690
#> N[13]     1311.974   73.704  1168.594  1262.254  1311.874  1361.720  1456.847
#> N[14]     1571.090   89.536  1398.506  1510.792  1569.996  1630.734  1750.616
#> N[15]     2394.431  153.041  2095.715  2292.351  2393.482  2497.532  2696.191
#> meanr       -0.049    0.003    -0.054    -0.051    -0.049    -0.047    -0.043
#> r[1]        -0.218    0.023    -0.263    -0.233    -0.218    -0.202    -0.173
#> r[2]        -0.150    0.069    -0.285    -0.196    -0.150    -0.103    -0.013
#> r[3]        -0.051    0.061    -0.171    -0.091    -0.051    -0.010     0.069
#> r[4]        -0.023    0.034    -0.090    -0.046    -0.023    -0.001     0.044
#> r[5]         0.090    0.064    -0.036     0.047     0.091     0.133     0.216
#> r[6]         0.075    0.058    -0.039     0.036     0.076     0.115     0.189
#> r[7]        -0.022    0.051    -0.120    -0.056    -0.022     0.012     0.081
#> r[8]        -0.092    0.050    -0.186    -0.126    -0.093    -0.058     0.006
#> r[9]        -0.025    0.044    -0.111    -0.055    -0.025     0.005     0.061
#> r[10]        0.000    0.067    -0.134    -0.044     0.001     0.046     0.128
#> r[11]       -0.050    0.066    -0.180    -0.094    -0.050    -0.005     0.079
#> r[12]       -0.086    0.016    -0.115    -0.097    -0.086    -0.076    -0.055
#> r[13]        0.090    0.036     0.021     0.066     0.090     0.114     0.160
#> r[14]        0.140    0.027     0.087     0.122     0.141     0.159     0.193
#> sdr          0.112    0.009     0.094     0.106     0.112     0.118     0.131
#> vrrmax       0.461    0.038     0.388     0.435     0.460     0.486     0.539
#> deviance   236.237    5.158   227.500   232.585   235.764   239.411   247.580
#>           Rhat n.eff
#> N[1]     1.004   550
#> N[2]     1.001  5400
#> N[3]     1.001 27000
#> N[4]     1.002  2500
#> N[5]     1.001 10000
#> N[6]     1.001 14000
#> N[7]     1.001 27000
#> N[8]     1.002  1500
#> N[9]     1.001 27000
#> N[10]    1.001 27000
#> N[11]    1.001  2800
#> N[12]    1.003   820
#> N[13]    1.001 27000
#> N[14]    1.001 27000
#> N[15]    1.001 27000
#> meanr    1.003   840
#> r[1]     1.003   700
#> r[2]     1.002   990
#> r[3]     1.002  2300
#> r[4]     1.001 11000
#> r[5]     1.001 27000
#> r[6]     1.001  5700
#> r[7]     1.001  3300
#> r[8]     1.002  1200
#> r[9]     1.001 27000
#> r[10]    1.003   890
#> r[11]    1.003   870
#> r[12]    1.002  1100
#> r[13]    1.001 20000
#> r[14]    1.001 27000
#> sdr      1.001 12000
#> vrrmax   1.001 12000
#> deviance 1.001 27000
#> 
#> For each parameter, n.eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor (at convergence, Rhat=1).
#> 
#> DIC info (using the rule, pD = var(deviance)/2)
#> pD = 13.3 and DIC = 249.5
#> DIC is an estimate of expected predictive error (lower deviance is better).
#> 

## Import BUGS outputs for all count series ----
popbayes::read_bugs(path = temp_path)
#> $garamba__alcelaphus_buselaphus
#> Inference for Bugs model at "model.bug", fit using jags,
#>  2 chains, each with 50000 iterations (first 10000 discarded), n.thin = 3
#>  n.sims = 26666 iterations saved
#>            mu.vect  sd.vect      2.5%       25%       50%       75%     97.5%
#> N[1]     17685.981 1733.998 14277.150 16508.855 17717.666 18846.160 21063.507
#> N[2]      3865.329  512.525  2891.789  3516.437  3855.529  4198.655  4906.995
#> N[3]      3322.320  398.899  2549.491  3057.826  3321.147  3586.658  4113.046
#> N[4]      3010.473  427.929  2217.944  2713.434  2995.194  3289.426  3882.440
#> N[5]      2675.482  331.545  2041.604  2447.240  2671.309  2900.134  3330.760
#> N[6]      3224.325  544.322  2248.000  2839.607  3189.280  3573.923  4389.428
#> N[7]      3755.939  668.736  2559.061  3281.074  3721.329  4188.211  5163.674
#> N[8]      3483.815  365.680  2799.321  3230.259  3478.492  3728.039  4214.172
#> N[9]      2888.850  206.927  2490.742  2748.531  2888.379  3027.213  3301.309
#> N[10]     2748.587  206.509  2349.815  2607.414  2747.427  2886.491  3157.499
#> N[11]     2753.478  260.659  2257.112  2573.372  2749.665  2926.622  3270.544
#> N[12]     2624.659  300.613  2061.996  2412.867  2615.560  2824.720  3234.690
#> N[13]     1311.974   73.704  1168.594  1262.254  1311.874  1361.720  1456.847
#> N[14]     1571.090   89.536  1398.506  1510.792  1569.996  1630.734  1750.616
#> N[15]     2394.431  153.041  2095.715  2292.351  2393.482  2497.532  2696.191
#> meanr       -0.049    0.003    -0.054    -0.051    -0.049    -0.047    -0.043
#> r[1]        -0.218    0.023    -0.263    -0.233    -0.218    -0.202    -0.173
#> r[2]        -0.150    0.069    -0.285    -0.196    -0.150    -0.103    -0.013
#> r[3]        -0.051    0.061    -0.171    -0.091    -0.051    -0.010     0.069
#> r[4]        -0.023    0.034    -0.090    -0.046    -0.023    -0.001     0.044
#> r[5]         0.090    0.064    -0.036     0.047     0.091     0.133     0.216
#> r[6]         0.075    0.058    -0.039     0.036     0.076     0.115     0.189
#> r[7]        -0.022    0.051    -0.120    -0.056    -0.022     0.012     0.081
#> r[8]        -0.092    0.050    -0.186    -0.126    -0.093    -0.058     0.006
#> r[9]        -0.025    0.044    -0.111    -0.055    -0.025     0.005     0.061
#> r[10]        0.000    0.067    -0.134    -0.044     0.001     0.046     0.128
#> r[11]       -0.050    0.066    -0.180    -0.094    -0.050    -0.005     0.079
#> r[12]       -0.086    0.016    -0.115    -0.097    -0.086    -0.076    -0.055
#> r[13]        0.090    0.036     0.021     0.066     0.090     0.114     0.160
#> r[14]        0.140    0.027     0.087     0.122     0.141     0.159     0.193
#> sdr          0.112    0.009     0.094     0.106     0.112     0.118     0.131
#> vrrmax       0.461    0.038     0.388     0.435     0.460     0.486     0.539
#> deviance   236.237    5.158   227.500   232.585   235.764   239.411   247.580
#>           Rhat n.eff
#> N[1]     1.004   550
#> N[2]     1.001  5400
#> N[3]     1.001 27000
#> N[4]     1.002  2500
#> N[5]     1.001 10000
#> N[6]     1.001 14000
#> N[7]     1.001 27000
#> N[8]     1.002  1500
#> N[9]     1.001 27000
#> N[10]    1.001 27000
#> N[11]    1.001  2800
#> N[12]    1.003   820
#> N[13]    1.001 27000
#> N[14]    1.001 27000
#> N[15]    1.001 27000
#> meanr    1.003   840
#> r[1]     1.003   700
#> r[2]     1.002   990
#> r[3]     1.002  2300
#> r[4]     1.001 11000
#> r[5]     1.001 27000
#> r[6]     1.001  5700
#> r[7]     1.001  3300
#> r[8]     1.002  1200
#> r[9]     1.001 27000
#> r[10]    1.003   890
#> r[11]    1.003   870
#> r[12]    1.002  1100
#> r[13]    1.001 20000
#> r[14]    1.001 27000
#> sdr      1.001 12000
#> vrrmax   1.001 12000
#> deviance 1.001 27000
#> 
#> For each parameter, n.eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor (at convergence, Rhat=1).
#> 
#> DIC info (using the rule, pD = var(deviance)/2)
#> pD = 13.3 and DIC = 249.5
#> DIC is an estimate of expected predictive error (lower deviance is better).
#> 
# }