Import a list of BUGS outputs previously exported

This function imports a list of BUGS outputs previously exported by fit_trend(). Users can import one, several, or all models.

read_bugs(series = NULL, path = ".")

Arguments

series

a vector of character strings. One or several count series names. If NULL (default) BUGS outputs for all count series will be imported. Users can run list_series() to get the correct spelling of count series names.

path

a character string. The directory in which BUGS outputs have been saved by the function fit_trend().

Value

An n-element list (where n is the number of count series). See fit_trend() for further information.

Examples

## Load Garamba raw dataset ----
file_path <- system.file("extdata", "garamba_survey.csv", 
                         package = "popbayes")
                         
garamba <- read.csv(file = file_path)

## Create temporary folder ----
temp_path <- tempdir()

## Format dataset ----
garamba_formatted <- popbayes::format_data(
  data              = garamba, 
  path              = temp_path,
  field_method      = "field_method",
  pref_field_method = "pref_field_method",
  conversion_A2G    = "conversion_A2G",
  rmax              = "rmax")
#> ✔ Detecting 10 count series

## Select one serie ----
a_buselaphus <- popbayes::filter_series(garamba_formatted, 
                                        location = "Garamba",
                                        species  = "Alcelaphus buselaphus")
#> ✔ Found 1 series with 'Alcelaphus buselaphus' and 'Garamba'.
# \donttest{
## Fit population trends (requires JAGS) ----
a_buselaphus_mod <- popbayes::fit_trend(a_buselaphus, path = temp_path)
#> Compiling data graph
#>    Resolving undeclared variables
#>    Allocating nodes
#>    Initializing
#>    Reading data back into data table
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 15
#>    Unobserved stochastic nodes: 15
#>    Total graph size: 227
#> 
#> Initializing model
#> 

## Import BUGS outputs for one count series ----
popbayes::read_bugs(series = "garamba__alcelaphus_buselaphus", 
                    path   = temp_path)
#> $garamba__alcelaphus_buselaphus
#> Inference for Bugs model at "model.bug", fit using jags,
#>  2 chains, each with 50000 iterations (first 10000 discarded), n.thin = 3
#>  n.sims = 26666 iterations saved
#>            mu.vect  sd.vect      2.5%       25%       50%       75%     97.5%
#> N[1]     17661.074 1701.159 14309.889 16508.102 17672.885 18826.564 20933.355
#> N[2]      3861.579  494.476  2934.146  3523.530  3848.635  4186.298  4868.800
#> N[3]      3315.685  386.678  2573.366  3051.015  3315.503  3573.661  4079.259
#> N[4]      3008.722  427.603  2201.413  2712.282  2997.262  3289.630  3876.839
#> N[5]      2680.864  326.829  2054.202  2457.771  2675.103  2900.955  3336.577
#> N[6]      3237.317  551.752  2277.476  2837.421  3203.629  3588.930  4417.243
#> N[7]      3760.564  678.594  2578.346  3279.455  3708.088  4192.553  5213.219
#> N[8]      3483.056  369.140  2779.127  3233.202  3476.114  3728.303  4227.297
#> N[9]      2886.147  208.942  2485.415  2742.313  2884.355  3025.871  3302.931
#> N[10]     2748.918  204.851  2350.796  2612.495  2746.773  2884.251  3158.869
#> N[11]     2757.072  255.916  2266.481  2582.545  2755.253  2927.801  3272.840
#> N[12]     2629.967  296.138  2069.721  2425.715  2619.169  2827.448  3230.959
#> N[13]     1311.119   73.065  1169.076  1261.705  1310.403  1360.405  1455.064
#> N[14]     1571.190   88.743  1398.418  1510.687  1570.019  1630.961  1747.670
#> N[15]     2394.290  151.672  2094.655  2292.424  2394.813  2496.242  2691.404
#> meanr       -0.049    0.003    -0.054    -0.051    -0.049    -0.047    -0.043
#> r[1]        -0.218    0.022    -0.262    -0.233    -0.218    -0.203    -0.174
#> r[2]        -0.151    0.069    -0.285    -0.197    -0.151    -0.105    -0.016
#> r[3]        -0.050    0.059    -0.168    -0.089    -0.049    -0.010     0.063
#> r[4]        -0.023    0.033    -0.087    -0.045    -0.023    -0.001     0.043
#> r[5]         0.091    0.064    -0.034     0.048     0.091     0.134     0.221
#> r[6]         0.074    0.057    -0.038     0.037     0.074     0.112     0.187
#> r[7]        -0.022    0.052    -0.121    -0.057    -0.023     0.013     0.079
#> r[8]        -0.092    0.051    -0.191    -0.127    -0.093    -0.058     0.011
#> r[9]        -0.024    0.044    -0.111    -0.054    -0.024     0.005     0.062
#> r[10]        0.001    0.066    -0.129    -0.043     0.002     0.046     0.130
#> r[11]       -0.049    0.065    -0.177    -0.093    -0.049    -0.006     0.079
#> r[12]       -0.086    0.015    -0.116    -0.097    -0.087    -0.076    -0.055
#> r[13]        0.090    0.035     0.021     0.067     0.090     0.114     0.159
#> r[14]        0.140    0.027     0.087     0.122     0.140     0.159     0.193
#> sdr          0.112    0.009     0.094     0.105     0.112     0.118     0.131
#> vrrmax       0.461    0.039     0.387     0.434     0.460     0.486     0.541
#> deviance   236.012    5.174   227.233   232.320   235.569   239.192   247.232
#>           Rhat n.eff
#> N[1]     1.002  1100
#> N[2]     1.002  2700
#> N[3]     1.001  4900
#> N[4]     1.001 12000
#> N[5]     1.001 27000
#> N[6]     1.001 27000
#> N[7]     1.001 27000
#> N[8]     1.001 26000
#> N[9]     1.001 27000
#> N[10]    1.001  5100
#> N[11]    1.001  9600
#> N[12]    1.001 27000
#> N[13]    1.001 27000
#> N[14]    1.001  9900
#> N[15]    1.001  4600
#> meanr    1.003   910
#> r[1]     1.003   830
#> r[2]     1.001  7000
#> r[3]     1.001 27000
#> r[4]     1.001 27000
#> r[5]     1.001 27000
#> r[6]     1.001 14000
#> r[7]     1.001 27000
#> r[8]     1.001 16000
#> r[9]     1.001  5500
#> r[10]    1.001 27000
#> r[11]    1.001  3300
#> r[12]    1.001 27000
#> r[13]    1.001 13000
#> r[14]    1.001  3300
#> sdr      1.001  3400
#> vrrmax   1.001  3400
#> deviance 1.001 27000
#> 
#> For each parameter, n.eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor (at convergence, Rhat=1).
#> 
#> DIC info (using the rule, pD = var(deviance)/2)
#> pD = 13.4 and DIC = 249.4
#> DIC is an estimate of expected predictive error (lower deviance is better).
#> 

## Import BUGS outputs for all count series ----
popbayes::read_bugs(path = temp_path)
#> $garamba__alcelaphus_buselaphus
#> Inference for Bugs model at "model.bug", fit using jags,
#>  2 chains, each with 50000 iterations (first 10000 discarded), n.thin = 3
#>  n.sims = 26666 iterations saved
#>            mu.vect  sd.vect      2.5%       25%       50%       75%     97.5%
#> N[1]     17661.074 1701.159 14309.889 16508.102 17672.885 18826.564 20933.355
#> N[2]      3861.579  494.476  2934.146  3523.530  3848.635  4186.298  4868.800
#> N[3]      3315.685  386.678  2573.366  3051.015  3315.503  3573.661  4079.259
#> N[4]      3008.722  427.603  2201.413  2712.282  2997.262  3289.630  3876.839
#> N[5]      2680.864  326.829  2054.202  2457.771  2675.103  2900.955  3336.577
#> N[6]      3237.317  551.752  2277.476  2837.421  3203.629  3588.930  4417.243
#> N[7]      3760.564  678.594  2578.346  3279.455  3708.088  4192.553  5213.219
#> N[8]      3483.056  369.140  2779.127  3233.202  3476.114  3728.303  4227.297
#> N[9]      2886.147  208.942  2485.415  2742.313  2884.355  3025.871  3302.931
#> N[10]     2748.918  204.851  2350.796  2612.495  2746.773  2884.251  3158.869
#> N[11]     2757.072  255.916  2266.481  2582.545  2755.253  2927.801  3272.840
#> N[12]     2629.967  296.138  2069.721  2425.715  2619.169  2827.448  3230.959
#> N[13]     1311.119   73.065  1169.076  1261.705  1310.403  1360.405  1455.064
#> N[14]     1571.190   88.743  1398.418  1510.687  1570.019  1630.961  1747.670
#> N[15]     2394.290  151.672  2094.655  2292.424  2394.813  2496.242  2691.404
#> meanr       -0.049    0.003    -0.054    -0.051    -0.049    -0.047    -0.043
#> r[1]        -0.218    0.022    -0.262    -0.233    -0.218    -0.203    -0.174
#> r[2]        -0.151    0.069    -0.285    -0.197    -0.151    -0.105    -0.016
#> r[3]        -0.050    0.059    -0.168    -0.089    -0.049    -0.010     0.063
#> r[4]        -0.023    0.033    -0.087    -0.045    -0.023    -0.001     0.043
#> r[5]         0.091    0.064    -0.034     0.048     0.091     0.134     0.221
#> r[6]         0.074    0.057    -0.038     0.037     0.074     0.112     0.187
#> r[7]        -0.022    0.052    -0.121    -0.057    -0.023     0.013     0.079
#> r[8]        -0.092    0.051    -0.191    -0.127    -0.093    -0.058     0.011
#> r[9]        -0.024    0.044    -0.111    -0.054    -0.024     0.005     0.062
#> r[10]        0.001    0.066    -0.129    -0.043     0.002     0.046     0.130
#> r[11]       -0.049    0.065    -0.177    -0.093    -0.049    -0.006     0.079
#> r[12]       -0.086    0.015    -0.116    -0.097    -0.087    -0.076    -0.055
#> r[13]        0.090    0.035     0.021     0.067     0.090     0.114     0.159
#> r[14]        0.140    0.027     0.087     0.122     0.140     0.159     0.193
#> sdr          0.112    0.009     0.094     0.105     0.112     0.118     0.131
#> vrrmax       0.461    0.039     0.387     0.434     0.460     0.486     0.541
#> deviance   236.012    5.174   227.233   232.320   235.569   239.192   247.232
#>           Rhat n.eff
#> N[1]     1.002  1100
#> N[2]     1.002  2700
#> N[3]     1.001  4900
#> N[4]     1.001 12000
#> N[5]     1.001 27000
#> N[6]     1.001 27000
#> N[7]     1.001 27000
#> N[8]     1.001 26000
#> N[9]     1.001 27000
#> N[10]    1.001  5100
#> N[11]    1.001  9600
#> N[12]    1.001 27000
#> N[13]    1.001 27000
#> N[14]    1.001  9900
#> N[15]    1.001  4600
#> meanr    1.003   910
#> r[1]     1.003   830
#> r[2]     1.001  7000
#> r[3]     1.001 27000
#> r[4]     1.001 27000
#> r[5]     1.001 27000
#> r[6]     1.001 14000
#> r[7]     1.001 27000
#> r[8]     1.001 16000
#> r[9]     1.001  5500
#> r[10]    1.001 27000
#> r[11]    1.001  3300
#> r[12]    1.001 27000
#> r[13]    1.001 13000
#> r[14]    1.001  3300
#> sdr      1.001  3400
#> vrrmax   1.001  3400
#> deviance 1.001 27000
#> 
#> For each parameter, n.eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor (at convergence, Rhat=1).
#> 
#> DIC info (using the rule, pD = var(deviance)/2)
#> pD = 13.4 and DIC = 249.4
#> DIC is an estimate of expected predictive error (lower deviance is better).
#> 
# }