Import a list of BUGS outputs previously exported

This function imports a list of BUGS outputs previously exported by fit_trend(). Users can import one, several, or all models.

read_bugs(series = NULL, path = ".")

Arguments

series

a vector of character strings. One or several count series names. If NULL (default) BUGS outputs for all count series will be imported. Users can run list_series() to get the correct spelling of count series names.

path

a character string. The directory in which BUGS outputs have been saved by the function fit_trend().

Value

An n-element list (where n is the number of count series). See fit_trend() for further information.

Examples

## Load Garamba raw dataset ----
file_path <- system.file("extdata", "garamba_survey.csv", 
                         package = "popbayes")
                         
garamba <- read.csv(file = file_path)

## Create temporary folder ----
temp_path <- tempdir()

## Format dataset ----
garamba_formatted <- popbayes::format_data(
  data              = garamba, 
  path              = temp_path,
  field_method      = "field_method",
  pref_field_method = "pref_field_method",
  conversion_A2G    = "conversion_A2G",
  rmax              = "rmax")
#> Warning: the condition has length > 1 and only the first element will be used
#> Warning: the condition has length > 1 and only the first element will be used
#> Warning: the condition has length > 1 and only the first element will be used
#> Warning: the condition has length > 1 and only the first element will be used
#> Warning: the condition has length > 1 and only the first element will be used
#> Warning: the condition has length > 1 and only the first element will be used
#> Warning: the condition has length > 1 and only the first element will be used
#> Warning: the condition has length > 1 and only the first element will be used
#> Warning: the condition has length > 1 and only the first element will be used
#> Warning: the condition has length > 1 and only the first element will be used
#> ✔ Detecting 10 count series

## Select one serie ----
a_buselaphus <- popbayes::filter_series(garamba_formatted, 
                                        location = "Garamba",
                                        species  = "Alcelaphus buselaphus")
#> ✔ Found 1 series with 'Alcelaphus buselaphus' and 'Garamba'.
# \donttest{
## Fit population trends (requires JAGS) ----
a_buselaphus_mod <- popbayes::fit_trend(a_buselaphus, path = temp_path)
#> Compiling data graph
#>    Resolving undeclared variables
#>    Allocating nodes
#>    Initializing
#>    Reading data back into data table
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 15
#>    Unobserved stochastic nodes: 15
#>    Total graph size: 227
#> 
#> Initializing model
#> 

## Import BUGS outputs for one count series ----
popbayes::read_bugs(series = "garamba__alcelaphus_buselaphus", 
                    path   = temp_path)
#> $garamba__alcelaphus_buselaphus
#> Inference for Bugs model at "model.bug", fit using jags,
#>  2 chains, each with 50000 iterations (first 10000 discarded), n.thin = 3
#>  n.sims = 26666 iterations saved
#>            mu.vect  sd.vect      2.5%       25%       50%       75%     97.5%
#> N[1]     17628.342 1749.497 14127.177 16477.078 17626.147 18815.766 21032.923
#> N[2]      3877.939  497.609  2939.413  3532.063  3867.807  4206.319  4881.145
#> N[3]      3332.164  387.247  2595.767  3065.740  3325.882  3590.591  4108.228
#> N[4]      3029.488  429.109  2226.834  2731.695  3012.426  3304.324  3924.424
#> N[5]      2672.947  326.409  2043.502  2449.650  2671.501  2892.976  3314.698
#> N[6]      3221.658  542.762  2244.574  2843.037  3190.896  3567.742  4363.613
#> N[7]      3749.076  665.783  2575.113  3283.165  3706.331  4164.941  5191.973
#> N[8]      3472.736  362.762  2775.785  3226.320  3469.065  3715.057  4200.518
#> N[9]      2884.558  208.090  2485.878  2743.385  2880.110  3022.770  3301.356
#> N[10]     2754.341  204.832  2357.380  2615.210  2751.522  2891.545  3161.784
#> N[11]     2760.163  255.511  2272.204  2584.904  2754.556  2932.807  3272.415
#> N[12]     2635.338  297.107  2067.823  2430.387  2626.459  2834.928  3235.580
#> N[13]     1310.147   73.758  1167.239  1259.853  1309.863  1360.313  1454.632
#> N[14]     1570.825   89.342  1397.673  1510.740  1570.105  1631.414  1748.812
#> N[15]     2392.404  152.735  2093.923  2288.909  2392.196  2496.518  2691.311
#> meanr       -0.049    0.003    -0.054    -0.051    -0.049    -0.047    -0.043
#> r[1]        -0.217    0.023    -0.260    -0.232    -0.217    -0.202    -0.172
#> r[2]        -0.150    0.069    -0.286    -0.197    -0.150    -0.103    -0.015
#> r[3]        -0.049    0.060    -0.168    -0.090    -0.049    -0.008     0.069
#> r[4]        -0.025    0.033    -0.090    -0.047    -0.024    -0.002     0.040
#> r[5]         0.090    0.064    -0.035     0.046     0.090     0.133     0.220
#> r[6]         0.075    0.058    -0.037     0.036     0.074     0.113     0.190
#> r[7]        -0.022    0.051    -0.122    -0.056    -0.023     0.012     0.080
#> r[8]        -0.091    0.050    -0.187    -0.125    -0.093    -0.058     0.011
#> r[9]        -0.023    0.044    -0.109    -0.052    -0.023     0.006     0.062
#> r[10]        0.001    0.066    -0.131    -0.043     0.002     0.046     0.128
#> r[11]       -0.048    0.066    -0.179    -0.092    -0.048    -0.003     0.080
#> r[12]       -0.087    0.015    -0.116    -0.097    -0.087    -0.077    -0.056
#> r[13]        0.091    0.036     0.021     0.066     0.091     0.115     0.161
#> r[14]        0.140    0.028     0.085     0.122     0.140     0.159     0.193
#> sdr          0.112    0.009     0.094     0.105     0.111     0.118     0.130
#> vrrmax       0.460    0.038     0.388     0.434     0.459     0.484     0.537
#> deviance   236.071    5.179   227.367   232.374   235.648   239.250   247.494
#>           Rhat n.eff
#> N[1]     1.001  6200
#> N[2]     1.001 27000
#> N[3]     1.001 27000
#> N[4]     1.001 14000
#> N[5]     1.001  8600
#> N[6]     1.001 22000
#> N[7]     1.001  3500
#> N[8]     1.001 27000
#> N[9]     1.001 13000
#> N[10]    1.001 27000
#> N[11]    1.001 27000
#> N[12]    1.001  4100
#> N[13]    1.001 15000
#> N[14]    1.001 21000
#> N[15]    1.001 27000
#> meanr    1.001 11000
#> r[1]     1.001 23000
#> r[2]     1.001 27000
#> r[3]     1.001  9800
#> r[4]     1.002  4300
#> r[5]     1.002  2600
#> r[6]     1.001  3800
#> r[7]     1.001  4100
#> r[8]     1.001 27000
#> r[9]     1.001 12000
#> r[10]    1.001 10000
#> r[11]    1.002  2700
#> r[12]    1.001  3000
#> r[13]    1.001  6900
#> r[14]    1.001 27000
#> sdr      1.002  2100
#> vrrmax   1.002  2100
#> deviance 1.002  2000
#> 
#> For each parameter, n.eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor (at convergence, Rhat=1).
#> 
#> DIC info (using the rule, pD = var(deviance)/2)
#> pD = 13.4 and DIC = 249.5
#> DIC is an estimate of expected predictive error (lower deviance is better).
#> 

## Import BUGS outputs for all count series ----
popbayes::read_bugs(path = temp_path)
#> $garamba__alcelaphus_buselaphus
#> Inference for Bugs model at "model.bug", fit using jags,
#>  2 chains, each with 50000 iterations (first 10000 discarded), n.thin = 3
#>  n.sims = 26666 iterations saved
#>            mu.vect  sd.vect      2.5%       25%       50%       75%     97.5%
#> N[1]     17628.342 1749.497 14127.177 16477.078 17626.147 18815.766 21032.923
#> N[2]      3877.939  497.609  2939.413  3532.063  3867.807  4206.319  4881.145
#> N[3]      3332.164  387.247  2595.767  3065.740  3325.882  3590.591  4108.228
#> N[4]      3029.488  429.109  2226.834  2731.695  3012.426  3304.324  3924.424
#> N[5]      2672.947  326.409  2043.502  2449.650  2671.501  2892.976  3314.698
#> N[6]      3221.658  542.762  2244.574  2843.037  3190.896  3567.742  4363.613
#> N[7]      3749.076  665.783  2575.113  3283.165  3706.331  4164.941  5191.973
#> N[8]      3472.736  362.762  2775.785  3226.320  3469.065  3715.057  4200.518
#> N[9]      2884.558  208.090  2485.878  2743.385  2880.110  3022.770  3301.356
#> N[10]     2754.341  204.832  2357.380  2615.210  2751.522  2891.545  3161.784
#> N[11]     2760.163  255.511  2272.204  2584.904  2754.556  2932.807  3272.415
#> N[12]     2635.338  297.107  2067.823  2430.387  2626.459  2834.928  3235.580
#> N[13]     1310.147   73.758  1167.239  1259.853  1309.863  1360.313  1454.632
#> N[14]     1570.825   89.342  1397.673  1510.740  1570.105  1631.414  1748.812
#> N[15]     2392.404  152.735  2093.923  2288.909  2392.196  2496.518  2691.311
#> meanr       -0.049    0.003    -0.054    -0.051    -0.049    -0.047    -0.043
#> r[1]        -0.217    0.023    -0.260    -0.232    -0.217    -0.202    -0.172
#> r[2]        -0.150    0.069    -0.286    -0.197    -0.150    -0.103    -0.015
#> r[3]        -0.049    0.060    -0.168    -0.090    -0.049    -0.008     0.069
#> r[4]        -0.025    0.033    -0.090    -0.047    -0.024    -0.002     0.040
#> r[5]         0.090    0.064    -0.035     0.046     0.090     0.133     0.220
#> r[6]         0.075    0.058    -0.037     0.036     0.074     0.113     0.190
#> r[7]        -0.022    0.051    -0.122    -0.056    -0.023     0.012     0.080
#> r[8]        -0.091    0.050    -0.187    -0.125    -0.093    -0.058     0.011
#> r[9]        -0.023    0.044    -0.109    -0.052    -0.023     0.006     0.062
#> r[10]        0.001    0.066    -0.131    -0.043     0.002     0.046     0.128
#> r[11]       -0.048    0.066    -0.179    -0.092    -0.048    -0.003     0.080
#> r[12]       -0.087    0.015    -0.116    -0.097    -0.087    -0.077    -0.056
#> r[13]        0.091    0.036     0.021     0.066     0.091     0.115     0.161
#> r[14]        0.140    0.028     0.085     0.122     0.140     0.159     0.193
#> sdr          0.112    0.009     0.094     0.105     0.111     0.118     0.130
#> vrrmax       0.460    0.038     0.388     0.434     0.459     0.484     0.537
#> deviance   236.071    5.179   227.367   232.374   235.648   239.250   247.494
#>           Rhat n.eff
#> N[1]     1.001  6200
#> N[2]     1.001 27000
#> N[3]     1.001 27000
#> N[4]     1.001 14000
#> N[5]     1.001  8600
#> N[6]     1.001 22000
#> N[7]     1.001  3500
#> N[8]     1.001 27000
#> N[9]     1.001 13000
#> N[10]    1.001 27000
#> N[11]    1.001 27000
#> N[12]    1.001  4100
#> N[13]    1.001 15000
#> N[14]    1.001 21000
#> N[15]    1.001 27000
#> meanr    1.001 11000
#> r[1]     1.001 23000
#> r[2]     1.001 27000
#> r[3]     1.001  9800
#> r[4]     1.002  4300
#> r[5]     1.002  2600
#> r[6]     1.001  3800
#> r[7]     1.001  4100
#> r[8]     1.001 27000
#> r[9]     1.001 12000
#> r[10]    1.001 10000
#> r[11]    1.002  2700
#> r[12]    1.001  3000
#> r[13]    1.001  6900
#> r[14]    1.001 27000
#> sdr      1.002  2100
#> vrrmax   1.002  2100
#> deviance 1.002  2000
#> 
#> For each parameter, n.eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor (at convergence, Rhat=1).
#> 
#> DIC info (using the rule, pD = var(deviance)/2)
#> pD = 13.4 and DIC = 249.5
#> DIC is an estimate of expected predictive error (lower deviance is better).
#> 
# }