Import a list of BUGS outputs previously exported

This function imports a list of BUGS outputs previously exported by fit_trend(). Users can import one, several, or all models.

read_bugs(series = NULL, path = ".")

Arguments

series

a vector of character strings. One or several count series names. If NULL (default) BUGS outputs for all count series will be imported. Users can run list_series() to get the correct spelling of count series names.

path

a character string. The directory in which BUGS outputs have been saved by the function fit_trend().

Value

An n-element list (where n is the number of count series). See fit_trend() for further information.

Examples

## Load Garamba raw dataset ----
file_path <- system.file("extdata", "garamba_survey.csv", 
                         package = "popbayes")
                         
garamba <- read.csv(file = file_path)

## Create temporary folder ----
temp_path <- tempdir()

## Format dataset ----
garamba_formatted <- popbayes::format_data(
  data              = garamba, 
  path              = temp_path,
  field_method      = "field_method",
  pref_field_method = "pref_field_method",
  conversion_A2G    = "conversion_A2G",
  rmax              = "rmax")
#> ✔ Detecting 10 count series

## Select one serie ----
a_buselaphus <- popbayes::filter_series(garamba_formatted, 
                                        location = "Garamba",
                                        species  = "Alcelaphus buselaphus")
#> ✔ Found 1 series with 'Alcelaphus buselaphus' and 'Garamba'.
# \donttest{
## Fit population trends (requires JAGS) ----
a_buselaphus_mod <- popbayes::fit_trend(a_buselaphus, path = temp_path)
#> Compiling data graph
#>    Resolving undeclared variables
#>    Allocating nodes
#>    Initializing
#>    Reading data back into data table
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 15
#>    Unobserved stochastic nodes: 15
#>    Total graph size: 227
#> 
#> Initializing model
#> 

## Import BUGS outputs for one count series ----
popbayes::read_bugs(series = "garamba__alcelaphus_buselaphus", 
                    path   = temp_path)
#> $garamba__alcelaphus_buselaphus
#> Inference for Bugs model at "model.bug", fit using jags,
#>  2 chains, each with 50000 iterations (first 10000 discarded), n.thin = 3
#>  n.sims = 26666 iterations saved
#>            mu.vect  sd.vect      2.5%       25%       50%       75%     97.5%
#> N[1]     17730.900 1678.996 14429.481 16591.119 17730.213 18856.935 21030.746
#> N[2]      3872.211  481.365  2959.722  3540.121  3859.407  4194.023  4843.818
#> N[3]      3328.251  378.024  2592.575  3071.383  3323.944  3584.369  4073.968
#> N[4]      3012.375  428.228  2227.176  2714.226  2991.022  3291.778  3907.829
#> N[5]      2660.144  331.129  2026.542  2434.606  2659.273  2883.714  3310.262
#> N[6]      3204.089  546.954  2226.618  2827.981  3167.917  3541.641  4384.298
#> N[7]      3731.143  681.018  2524.343  3257.800  3685.519  4153.352  5199.300
#> N[8]      3477.039  367.415  2781.606  3223.657  3465.849  3725.654  4214.845
#> N[9]      2887.817  207.915  2483.238  2748.246  2886.307  3028.482  3301.839
#> N[10]     2749.672  205.866  2349.752  2609.265  2748.649  2887.953  3158.468
#> N[11]     2756.313  261.537  2256.478  2577.232  2752.143  2929.845  3278.567
#> N[12]     2629.900  300.140  2068.024  2423.032  2623.748  2830.492  3236.322
#> N[13]     1310.918   73.872  1166.628  1260.922  1310.628  1361.526  1454.972
#> N[14]     1569.982   89.279  1397.320  1509.348  1569.503  1629.569  1746.690
#> N[15]     2393.812  152.724  2094.696  2289.760  2393.750  2496.434  2694.850
#> meanr       -0.049    0.003    -0.054    -0.051    -0.049    -0.047    -0.043
#> r[1]        -0.218    0.021    -0.261    -0.232    -0.217    -0.203    -0.178
#> r[2]        -0.150    0.069    -0.287    -0.197    -0.150    -0.103    -0.014
#> r[3]        -0.052    0.060    -0.171    -0.091    -0.051    -0.012     0.066
#> r[4]        -0.024    0.034    -0.091    -0.047    -0.024    -0.002     0.043
#> r[5]         0.090    0.065    -0.038     0.046     0.090     0.134     0.221
#> r[6]         0.075    0.058    -0.037     0.036     0.074     0.113     0.190
#> r[7]        -0.020    0.051    -0.119    -0.055    -0.021     0.014     0.083
#> r[8]        -0.091    0.050    -0.187    -0.126    -0.092    -0.058     0.009
#> r[9]        -0.025    0.044    -0.112    -0.055    -0.024     0.005     0.061
#> r[10]        0.001    0.067    -0.134    -0.044     0.002     0.046     0.130
#> r[11]       -0.049    0.066    -0.177    -0.093    -0.049    -0.005     0.078
#> r[12]       -0.086    0.016    -0.116    -0.097    -0.087    -0.076    -0.055
#> r[13]        0.090    0.035     0.021     0.067     0.090     0.114     0.159
#> r[14]        0.140    0.027     0.086     0.122     0.141     0.159     0.194
#> sdr          0.112    0.009     0.095     0.106     0.112     0.118     0.131
#> vrrmax       0.461    0.038     0.390     0.435     0.460     0.486     0.539
#> deviance   236.101    5.057   227.401   232.512   235.663   239.212   247.198
#>           Rhat n.eff
#> N[1]     1.001  3400
#> N[2]     1.001 18000
#> N[3]     1.001  9800
#> N[4]     1.001 27000
#> N[5]     1.001  7400
#> N[6]     1.001 27000
#> N[7]     1.001  4400
#> N[8]     1.001  4400
#> N[9]     1.001 16000
#> N[10]    1.001  3700
#> N[11]    1.002  1400
#> N[12]    1.002  1500
#> N[13]    1.001  6300
#> N[14]    1.001 25000
#> N[15]    1.001 12000
#> meanr    1.001 12000
#> r[1]     1.001  3400
#> r[2]     1.001 27000
#> r[3]     1.001 16000
#> r[4]     1.001 18000
#> r[5]     1.001 27000
#> r[6]     1.001  3300
#> r[7]     1.001 17000
#> r[8]     1.002  2200
#> r[9]     1.002  2400
#> r[10]    1.002  2500
#> r[11]    1.001 24000
#> r[12]    1.002  1100
#> r[13]    1.001  4400
#> r[14]    1.001  8200
#> sdr      1.001 27000
#> vrrmax   1.001 27000
#> deviance 1.002  1000
#> 
#> For each parameter, n.eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor (at convergence, Rhat=1).
#> 
#> DIC info (using the rule, pD = var(deviance)/2)
#> pD = 12.8 and DIC = 248.9
#> DIC is an estimate of expected predictive error (lower deviance is better).
#> 

## Import BUGS outputs for all count series ----
popbayes::read_bugs(path = temp_path)
#> $garamba__alcelaphus_buselaphus
#> Inference for Bugs model at "model.bug", fit using jags,
#>  2 chains, each with 50000 iterations (first 10000 discarded), n.thin = 3
#>  n.sims = 26666 iterations saved
#>            mu.vect  sd.vect      2.5%       25%       50%       75%     97.5%
#> N[1]     17730.900 1678.996 14429.481 16591.119 17730.213 18856.935 21030.746
#> N[2]      3872.211  481.365  2959.722  3540.121  3859.407  4194.023  4843.818
#> N[3]      3328.251  378.024  2592.575  3071.383  3323.944  3584.369  4073.968
#> N[4]      3012.375  428.228  2227.176  2714.226  2991.022  3291.778  3907.829
#> N[5]      2660.144  331.129  2026.542  2434.606  2659.273  2883.714  3310.262
#> N[6]      3204.089  546.954  2226.618  2827.981  3167.917  3541.641  4384.298
#> N[7]      3731.143  681.018  2524.343  3257.800  3685.519  4153.352  5199.300
#> N[8]      3477.039  367.415  2781.606  3223.657  3465.849  3725.654  4214.845
#> N[9]      2887.817  207.915  2483.238  2748.246  2886.307  3028.482  3301.839
#> N[10]     2749.672  205.866  2349.752  2609.265  2748.649  2887.953  3158.468
#> N[11]     2756.313  261.537  2256.478  2577.232  2752.143  2929.845  3278.567
#> N[12]     2629.900  300.140  2068.024  2423.032  2623.748  2830.492  3236.322
#> N[13]     1310.918   73.872  1166.628  1260.922  1310.628  1361.526  1454.972
#> N[14]     1569.982   89.279  1397.320  1509.348  1569.503  1629.569  1746.690
#> N[15]     2393.812  152.724  2094.696  2289.760  2393.750  2496.434  2694.850
#> meanr       -0.049    0.003    -0.054    -0.051    -0.049    -0.047    -0.043
#> r[1]        -0.218    0.021    -0.261    -0.232    -0.217    -0.203    -0.178
#> r[2]        -0.150    0.069    -0.287    -0.197    -0.150    -0.103    -0.014
#> r[3]        -0.052    0.060    -0.171    -0.091    -0.051    -0.012     0.066
#> r[4]        -0.024    0.034    -0.091    -0.047    -0.024    -0.002     0.043
#> r[5]         0.090    0.065    -0.038     0.046     0.090     0.134     0.221
#> r[6]         0.075    0.058    -0.037     0.036     0.074     0.113     0.190
#> r[7]        -0.020    0.051    -0.119    -0.055    -0.021     0.014     0.083
#> r[8]        -0.091    0.050    -0.187    -0.126    -0.092    -0.058     0.009
#> r[9]        -0.025    0.044    -0.112    -0.055    -0.024     0.005     0.061
#> r[10]        0.001    0.067    -0.134    -0.044     0.002     0.046     0.130
#> r[11]       -0.049    0.066    -0.177    -0.093    -0.049    -0.005     0.078
#> r[12]       -0.086    0.016    -0.116    -0.097    -0.087    -0.076    -0.055
#> r[13]        0.090    0.035     0.021     0.067     0.090     0.114     0.159
#> r[14]        0.140    0.027     0.086     0.122     0.141     0.159     0.194
#> sdr          0.112    0.009     0.095     0.106     0.112     0.118     0.131
#> vrrmax       0.461    0.038     0.390     0.435     0.460     0.486     0.539
#> deviance   236.101    5.057   227.401   232.512   235.663   239.212   247.198
#>           Rhat n.eff
#> N[1]     1.001  3400
#> N[2]     1.001 18000
#> N[3]     1.001  9800
#> N[4]     1.001 27000
#> N[5]     1.001  7400
#> N[6]     1.001 27000
#> N[7]     1.001  4400
#> N[8]     1.001  4400
#> N[9]     1.001 16000
#> N[10]    1.001  3700
#> N[11]    1.002  1400
#> N[12]    1.002  1500
#> N[13]    1.001  6300
#> N[14]    1.001 25000
#> N[15]    1.001 12000
#> meanr    1.001 12000
#> r[1]     1.001  3400
#> r[2]     1.001 27000
#> r[3]     1.001 16000
#> r[4]     1.001 18000
#> r[5]     1.001 27000
#> r[6]     1.001  3300
#> r[7]     1.001 17000
#> r[8]     1.002  2200
#> r[9]     1.002  2400
#> r[10]    1.002  2500
#> r[11]    1.001 24000
#> r[12]    1.002  1100
#> r[13]    1.001  4400
#> r[14]    1.001  8200
#> sdr      1.001 27000
#> vrrmax   1.001 27000
#> deviance 1.002  1000
#> 
#> For each parameter, n.eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor (at convergence, Rhat=1).
#> 
#> DIC info (using the rule, pD = var(deviance)/2)
#> pD = 12.8 and DIC = 248.9
#> DIC is an estimate of expected predictive error (lower deviance is better).
#> 
# }