Import a list of BUGS outputs previously exported

This function imports a list of BUGS outputs previously exported by fit_trend(). Users can import one, several, or all models.

read_bugs(series = NULL, path = ".")

Arguments

series

a vector of character strings. One or several count series names. If NULL (default) BUGS outputs for all count series will be imported. Users can run list_series() to get the correct spelling of count series names.

path

a character string. The directory in which BUGS outputs have been saved by the function fit_trend().

Value

An n-element list (where n is the number of count series). See fit_trend() for further information.

Examples

## Load Garamba raw dataset ----
file_path <- system.file("extdata", "garamba_survey.csv", 
                         package = "popbayes")
                         
garamba <- read.csv(file = file_path)

## Create temporary folder ----
temp_path <- tempdir()

## Format dataset ----
garamba_formatted <- popbayes::format_data(
  data              = garamba, 
  path              = temp_path,
  field_method      = "field_method",
  pref_field_method = "pref_field_method",
  conversion_A2G    = "conversion_A2G",
  rmax              = "rmax")
#> ✔ Detecting 10 count series

## Select one serie ----
a_buselaphus <- popbayes::filter_series(garamba_formatted, 
                                        location = "Garamba",
                                        species  = "Alcelaphus buselaphus")
#> ✔ Found 1 series with 'Alcelaphus buselaphus' and 'Garamba'.
# \donttest{
## Fit population trends (requires JAGS) ----
a_buselaphus_mod <- popbayes::fit_trend(a_buselaphus, path = temp_path)
#> Compiling data graph
#>    Resolving undeclared variables
#>    Allocating nodes
#>    Initializing
#>    Reading data back into data table
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 15
#>    Unobserved stochastic nodes: 15
#>    Total graph size: 227
#> 
#> Initializing model
#> 

## Import BUGS outputs for one count series ----
popbayes::read_bugs(series = "garamba__alcelaphus_buselaphus", 
                    path   = temp_path)
#> $garamba__alcelaphus_buselaphus
#> Inference for Bugs model at "model.bug", fit using jags,
#>  2 chains, each with 50000 iterations (first 10000 discarded), n.thin = 3
#>  n.sims = 26666 iterations saved
#>            mu.vect  sd.vect      2.5%       25%       50%       75%     97.5%
#> N[1]     17653.580 1730.708 14293.590 16482.057 17649.444 18821.616 21070.941
#> N[2]      3871.916  503.463  2922.952  3520.481  3862.799  4209.422  4884.243
#> N[3]      3328.790  388.305  2573.761  3061.519  3328.370  3592.905  4086.847
#> N[4]      3024.678  428.357  2234.958  2727.341  3007.800  3300.223  3912.569
#> N[5]      2662.135  325.385  2039.749  2441.059  2658.166  2878.316  3313.200
#> N[6]      3211.630  532.224  2241.394  2840.639  3189.542  3561.475  4318.470
#> N[7]      3752.600  669.931  2554.177  3278.095  3719.641  4183.119  5175.053
#> N[8]      3479.440  369.143  2783.057  3220.889  3470.503  3726.885  4220.924
#> N[9]      2887.397  207.596  2487.039  2744.986  2884.240  3026.314  3301.847
#> N[10]     2751.292  204.977  2353.633  2610.787  2750.018  2889.170  3156.982
#> N[11]     2757.969  260.244  2261.522  2580.172  2754.329  2931.580  3280.249
#> N[12]     2630.888  301.253  2064.062  2424.149  2624.386  2830.413  3245.468
#> N[13]     1311.073   73.380  1168.578  1261.404  1310.985  1360.453  1455.560
#> N[14]     1571.404   90.254  1396.228  1510.828  1570.937  1631.687  1749.916
#> N[15]     2394.661  152.030  2096.964  2292.921  2394.861  2496.805  2690.762
#> meanr       -0.049    0.003    -0.054    -0.051    -0.049    -0.047    -0.043
#> r[1]        -0.217    0.023    -0.263    -0.232    -0.217    -0.202    -0.174
#> r[2]        -0.149    0.069    -0.285    -0.196    -0.150    -0.103    -0.014
#> r[3]        -0.049    0.061    -0.173    -0.090    -0.049    -0.007     0.069
#> r[4]        -0.025    0.033    -0.090    -0.047    -0.025    -0.003     0.039
#> r[5]         0.091    0.063    -0.037     0.048     0.091     0.133     0.217
#> r[6]         0.077    0.057    -0.036     0.038     0.077     0.115     0.192
#> r[7]        -0.022    0.051    -0.118    -0.057    -0.023     0.013     0.080
#> r[8]        -0.092    0.051    -0.190    -0.126    -0.092    -0.058     0.010
#> r[9]        -0.024    0.044    -0.110    -0.054    -0.024     0.005     0.060
#> r[10]        0.001    0.067    -0.131    -0.044     0.001     0.046     0.130
#> r[11]       -0.049    0.066    -0.181    -0.093    -0.049    -0.005     0.078
#> r[12]       -0.086    0.016    -0.117    -0.097    -0.087    -0.076    -0.055
#> r[13]        0.091    0.036     0.021     0.066     0.091     0.115     0.160
#> r[14]        0.140    0.028     0.086     0.122     0.140     0.159     0.194
#> sdr          0.112    0.009     0.095     0.106     0.112     0.118     0.131
#> vrrmax       0.461    0.038     0.389     0.435     0.460     0.486     0.540
#> deviance   236.050    5.152   227.378   232.363   235.600   239.245   247.319
#>           Rhat n.eff
#> N[1]     1.001 27000
#> N[2]     1.001 27000
#> N[3]     1.001 27000
#> N[4]     1.002  2400
#> N[5]     1.001 13000
#> N[6]     1.001  5700
#> N[7]     1.001  4800
#> N[8]     1.001 27000
#> N[9]     1.001 27000
#> N[10]    1.001 27000
#> N[11]    1.001 27000
#> N[12]    1.001 27000
#> N[13]    1.001 27000
#> N[14]    1.001  4000
#> N[15]    1.001 24000
#> meanr    1.001 27000
#> r[1]     1.001 27000
#> r[2]     1.001 27000
#> r[3]     1.002  2000
#> r[4]     1.002  1700
#> r[5]     1.002  1500
#> r[6]     1.001 27000
#> r[7]     1.002  2700
#> r[8]     1.001 27000
#> r[9]     1.001 27000
#> r[10]    1.001 27000
#> r[11]    1.001 27000
#> r[12]    1.001 27000
#> r[13]    1.001  7100
#> r[14]    1.001  3500
#> sdr      1.001 24000
#> vrrmax   1.001 24000
#> deviance 1.001 27000
#> 
#> For each parameter, n.eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor (at convergence, Rhat=1).
#> 
#> DIC info (using the rule, pD = var(deviance)/2)
#> pD = 13.3 and DIC = 249.3
#> DIC is an estimate of expected predictive error (lower deviance is better).
#> 

## Import BUGS outputs for all count series ----
popbayes::read_bugs(path = temp_path)
#> $garamba__alcelaphus_buselaphus
#> Inference for Bugs model at "model.bug", fit using jags,
#>  2 chains, each with 50000 iterations (first 10000 discarded), n.thin = 3
#>  n.sims = 26666 iterations saved
#>            mu.vect  sd.vect      2.5%       25%       50%       75%     97.5%
#> N[1]     17653.580 1730.708 14293.590 16482.057 17649.444 18821.616 21070.941
#> N[2]      3871.916  503.463  2922.952  3520.481  3862.799  4209.422  4884.243
#> N[3]      3328.790  388.305  2573.761  3061.519  3328.370  3592.905  4086.847
#> N[4]      3024.678  428.357  2234.958  2727.341  3007.800  3300.223  3912.569
#> N[5]      2662.135  325.385  2039.749  2441.059  2658.166  2878.316  3313.200
#> N[6]      3211.630  532.224  2241.394  2840.639  3189.542  3561.475  4318.470
#> N[7]      3752.600  669.931  2554.177  3278.095  3719.641  4183.119  5175.053
#> N[8]      3479.440  369.143  2783.057  3220.889  3470.503  3726.885  4220.924
#> N[9]      2887.397  207.596  2487.039  2744.986  2884.240  3026.314  3301.847
#> N[10]     2751.292  204.977  2353.633  2610.787  2750.018  2889.170  3156.982
#> N[11]     2757.969  260.244  2261.522  2580.172  2754.329  2931.580  3280.249
#> N[12]     2630.888  301.253  2064.062  2424.149  2624.386  2830.413  3245.468
#> N[13]     1311.073   73.380  1168.578  1261.404  1310.985  1360.453  1455.560
#> N[14]     1571.404   90.254  1396.228  1510.828  1570.937  1631.687  1749.916
#> N[15]     2394.661  152.030  2096.964  2292.921  2394.861  2496.805  2690.762
#> meanr       -0.049    0.003    -0.054    -0.051    -0.049    -0.047    -0.043
#> r[1]        -0.217    0.023    -0.263    -0.232    -0.217    -0.202    -0.174
#> r[2]        -0.149    0.069    -0.285    -0.196    -0.150    -0.103    -0.014
#> r[3]        -0.049    0.061    -0.173    -0.090    -0.049    -0.007     0.069
#> r[4]        -0.025    0.033    -0.090    -0.047    -0.025    -0.003     0.039
#> r[5]         0.091    0.063    -0.037     0.048     0.091     0.133     0.217
#> r[6]         0.077    0.057    -0.036     0.038     0.077     0.115     0.192
#> r[7]        -0.022    0.051    -0.118    -0.057    -0.023     0.013     0.080
#> r[8]        -0.092    0.051    -0.190    -0.126    -0.092    -0.058     0.010
#> r[9]        -0.024    0.044    -0.110    -0.054    -0.024     0.005     0.060
#> r[10]        0.001    0.067    -0.131    -0.044     0.001     0.046     0.130
#> r[11]       -0.049    0.066    -0.181    -0.093    -0.049    -0.005     0.078
#> r[12]       -0.086    0.016    -0.117    -0.097    -0.087    -0.076    -0.055
#> r[13]        0.091    0.036     0.021     0.066     0.091     0.115     0.160
#> r[14]        0.140    0.028     0.086     0.122     0.140     0.159     0.194
#> sdr          0.112    0.009     0.095     0.106     0.112     0.118     0.131
#> vrrmax       0.461    0.038     0.389     0.435     0.460     0.486     0.540
#> deviance   236.050    5.152   227.378   232.363   235.600   239.245   247.319
#>           Rhat n.eff
#> N[1]     1.001 27000
#> N[2]     1.001 27000
#> N[3]     1.001 27000
#> N[4]     1.002  2400
#> N[5]     1.001 13000
#> N[6]     1.001  5700
#> N[7]     1.001  4800
#> N[8]     1.001 27000
#> N[9]     1.001 27000
#> N[10]    1.001 27000
#> N[11]    1.001 27000
#> N[12]    1.001 27000
#> N[13]    1.001 27000
#> N[14]    1.001  4000
#> N[15]    1.001 24000
#> meanr    1.001 27000
#> r[1]     1.001 27000
#> r[2]     1.001 27000
#> r[3]     1.002  2000
#> r[4]     1.002  1700
#> r[5]     1.002  1500
#> r[6]     1.001 27000
#> r[7]     1.002  2700
#> r[8]     1.001 27000
#> r[9]     1.001 27000
#> r[10]    1.001 27000
#> r[11]    1.001 27000
#> r[12]    1.001 27000
#> r[13]    1.001  7100
#> r[14]    1.001  3500
#> sdr      1.001 24000
#> vrrmax   1.001 24000
#> deviance 1.001 27000
#> 
#> For each parameter, n.eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor (at convergence, Rhat=1).
#> 
#> DIC info (using the rule, pD = var(deviance)/2)
#> pD = 13.3 and DIC = 249.3
#> DIC is an estimate of expected predictive error (lower deviance is better).
#> 
# }