Import a list of BUGS outputs previously exported

This function imports a list of BUGS outputs previously exported by fit_trend(). Users can import one, several, or all models.

read_bugs(series = NULL, path = ".")

Arguments

series

a vector of character strings. One or several count series names. If NULL (default) BUGS outputs for all count series will be imported. Users can run list_series() to get the correct spelling of count series names.

path

a character string. The directory in which BUGS outputs have been saved by the function fit_trend().

Value

An n-element list (where n is the number of count series). See fit_trend() for further information.

Examples

## Load Garamba raw dataset ----
file_path <- system.file("extdata", "garamba_survey.csv", 
                         package = "popbayes")
                         
garamba <- read.csv(file = file_path)

## Create temporary folder ----
temp_path <- tempdir()

## Format dataset ----
garamba_formatted <- popbayes::format_data(
  data              = garamba, 
  path              = temp_path,
  field_method      = "field_method",
  pref_field_method = "pref_field_method",
  conversion_A2G    = "conversion_A2G",
  rmax              = "rmax")
#> ✔ Detecting 10 count series

## Select one serie ----
a_buselaphus <- popbayes::filter_series(garamba_formatted, 
                                        location = "Garamba",
                                        species  = "Alcelaphus buselaphus")
#> ✔ Found 1 series with 'Alcelaphus buselaphus' and 'Garamba'.
# \donttest{
## Fit population trends (requires JAGS) ----
a_buselaphus_mod <- popbayes::fit_trend(a_buselaphus, path = temp_path)
#> Compiling data graph
#>    Resolving undeclared variables
#>    Allocating nodes
#>    Initializing
#>    Reading data back into data table
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 15
#>    Unobserved stochastic nodes: 15
#>    Total graph size: 227
#> 
#> Initializing model
#> 

## Import BUGS outputs for one count series ----
popbayes::read_bugs(series = "garamba__alcelaphus_buselaphus", 
                    path   = temp_path)
#> $garamba__alcelaphus_buselaphus
#> Inference for Bugs model at "model.bug", fit using jags,
#>  2 chains, each with 50000 iterations (first 10000 discarded), n.thin = 3
#>  n.sims = 26666 iterations saved
#>            mu.vect  sd.vect      2.5%       25%       50%       75%     97.5%
#> N[1]     17676.176 1754.571 14280.329 16476.677 17667.967 18853.856 21130.261
#> N[2]      3866.926  496.623  2911.524  3529.349  3859.963  4189.818  4862.853
#> N[3]      3324.657  392.409  2561.976  3062.254  3324.009  3580.436  4113.429
#> N[4]      3014.585  428.340  2224.495  2716.616  3002.811  3296.960  3890.363
#> N[5]      2661.159  332.827  2027.621  2431.864  2658.324  2883.180  3325.268
#> N[6]      3199.225  545.736  2228.172  2817.783  3156.155  3548.289  4369.881
#> N[7]      3721.544  667.890  2537.206  3251.366  3672.008  4149.841  5136.607
#> N[8]      3472.311  375.399  2750.523  3216.666  3469.322  3721.691  4214.941
#> N[9]      2889.217  208.225  2485.898  2748.898  2886.977  3027.746  3308.406
#> N[10]     2748.790  205.491  2351.420  2609.430  2746.724  2885.809  3155.655
#> N[11]     2754.248  259.789  2259.089  2577.347  2750.537  2924.410  3280.450
#> N[12]     2628.063  301.719  2063.389  2419.347  2620.715  2828.629  3245.335
#> N[13]     1311.162   73.321  1168.670  1261.436  1311.074  1360.626  1455.264
#> N[14]     1570.486   89.361  1397.684  1509.167  1570.091  1631.133  1746.905
#> N[15]     2392.987  151.515  2099.043  2290.081  2391.788  2494.502  2690.650
#> meanr       -0.049    0.003    -0.054    -0.051    -0.049    -0.047    -0.043
#> r[1]        -0.218    0.023    -0.264    -0.232    -0.217    -0.202    -0.174
#> r[2]        -0.150    0.069    -0.286    -0.196    -0.150    -0.104    -0.014
#> r[3]        -0.051    0.059    -0.167    -0.091    -0.050    -0.011     0.066
#> r[4]        -0.024    0.034    -0.090    -0.048    -0.024    -0.001     0.041
#> r[5]         0.089    0.065    -0.036     0.045     0.088     0.132     0.221
#> r[6]         0.075    0.058    -0.039     0.036     0.075     0.113     0.190
#> r[7]        -0.020    0.051    -0.118    -0.055    -0.020     0.015     0.082
#> r[8]        -0.090    0.051    -0.188    -0.125    -0.091    -0.056     0.012
#> r[9]        -0.025    0.044    -0.112    -0.054    -0.025     0.004     0.062
#> r[10]        0.000    0.067    -0.133    -0.044     0.001     0.046     0.130
#> r[11]       -0.049    0.066    -0.179    -0.093    -0.049    -0.004     0.079
#> r[12]       -0.086    0.016    -0.116    -0.097    -0.087    -0.076    -0.055
#> r[13]        0.090    0.035     0.021     0.066     0.090     0.114     0.160
#> r[14]        0.140    0.027     0.088     0.122     0.140     0.159     0.193
#> sdr          0.112    0.010     0.094     0.105     0.111     0.118     0.131
#> vrrmax       0.460    0.040     0.386     0.433     0.459     0.486     0.541
#> deviance   236.316    5.236   227.499   232.594   235.849   239.478   247.866
#>           Rhat n.eff
#> N[1]     1.003   810
#> N[2]     1.002  2300
#> N[3]     1.002  1200
#> N[4]     1.005   370
#> N[5]     1.003   880
#> N[6]     1.003   710
#> N[7]     1.003   740
#> N[8]     1.001 12000
#> N[9]     1.001 22000
#> N[10]    1.001  9700
#> N[11]    1.001  3000
#> N[12]    1.002  1600
#> N[13]    1.001 27000
#> N[14]    1.001 20000
#> N[15]    1.001 27000
#> meanr    1.003   960
#> r[1]     1.001 27000
#> r[2]     1.001  7500
#> r[3]     1.003   900
#> r[4]     1.008   210
#> r[5]     1.001  3600
#> r[6]     1.001 27000
#> r[7]     1.005   410
#> r[8]     1.001  4700
#> r[9]     1.001  4900
#> r[10]    1.001  4800
#> r[11]    1.001  3200
#> r[12]    1.002  1900
#> r[13]    1.001 26000
#> r[14]    1.001 27000
#> sdr      1.001  3400
#> vrrmax   1.001  3400
#> deviance 1.002  1500
#> 
#> For each parameter, n.eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor (at convergence, Rhat=1).
#> 
#> DIC info (using the rule, pD = var(deviance)/2)
#> pD = 13.7 and DIC = 250.0
#> DIC is an estimate of expected predictive error (lower deviance is better).
#> 

## Import BUGS outputs for all count series ----
popbayes::read_bugs(path = temp_path)
#> $garamba__alcelaphus_buselaphus
#> Inference for Bugs model at "model.bug", fit using jags,
#>  2 chains, each with 50000 iterations (first 10000 discarded), n.thin = 3
#>  n.sims = 26666 iterations saved
#>            mu.vect  sd.vect      2.5%       25%       50%       75%     97.5%
#> N[1]     17676.176 1754.571 14280.329 16476.677 17667.967 18853.856 21130.261
#> N[2]      3866.926  496.623  2911.524  3529.349  3859.963  4189.818  4862.853
#> N[3]      3324.657  392.409  2561.976  3062.254  3324.009  3580.436  4113.429
#> N[4]      3014.585  428.340  2224.495  2716.616  3002.811  3296.960  3890.363
#> N[5]      2661.159  332.827  2027.621  2431.864  2658.324  2883.180  3325.268
#> N[6]      3199.225  545.736  2228.172  2817.783  3156.155  3548.289  4369.881
#> N[7]      3721.544  667.890  2537.206  3251.366  3672.008  4149.841  5136.607
#> N[8]      3472.311  375.399  2750.523  3216.666  3469.322  3721.691  4214.941
#> N[9]      2889.217  208.225  2485.898  2748.898  2886.977  3027.746  3308.406
#> N[10]     2748.790  205.491  2351.420  2609.430  2746.724  2885.809  3155.655
#> N[11]     2754.248  259.789  2259.089  2577.347  2750.537  2924.410  3280.450
#> N[12]     2628.063  301.719  2063.389  2419.347  2620.715  2828.629  3245.335
#> N[13]     1311.162   73.321  1168.670  1261.436  1311.074  1360.626  1455.264
#> N[14]     1570.486   89.361  1397.684  1509.167  1570.091  1631.133  1746.905
#> N[15]     2392.987  151.515  2099.043  2290.081  2391.788  2494.502  2690.650
#> meanr       -0.049    0.003    -0.054    -0.051    -0.049    -0.047    -0.043
#> r[1]        -0.218    0.023    -0.264    -0.232    -0.217    -0.202    -0.174
#> r[2]        -0.150    0.069    -0.286    -0.196    -0.150    -0.104    -0.014
#> r[3]        -0.051    0.059    -0.167    -0.091    -0.050    -0.011     0.066
#> r[4]        -0.024    0.034    -0.090    -0.048    -0.024    -0.001     0.041
#> r[5]         0.089    0.065    -0.036     0.045     0.088     0.132     0.221
#> r[6]         0.075    0.058    -0.039     0.036     0.075     0.113     0.190
#> r[7]        -0.020    0.051    -0.118    -0.055    -0.020     0.015     0.082
#> r[8]        -0.090    0.051    -0.188    -0.125    -0.091    -0.056     0.012
#> r[9]        -0.025    0.044    -0.112    -0.054    -0.025     0.004     0.062
#> r[10]        0.000    0.067    -0.133    -0.044     0.001     0.046     0.130
#> r[11]       -0.049    0.066    -0.179    -0.093    -0.049    -0.004     0.079
#> r[12]       -0.086    0.016    -0.116    -0.097    -0.087    -0.076    -0.055
#> r[13]        0.090    0.035     0.021     0.066     0.090     0.114     0.160
#> r[14]        0.140    0.027     0.088     0.122     0.140     0.159     0.193
#> sdr          0.112    0.010     0.094     0.105     0.111     0.118     0.131
#> vrrmax       0.460    0.040     0.386     0.433     0.459     0.486     0.541
#> deviance   236.316    5.236   227.499   232.594   235.849   239.478   247.866
#>           Rhat n.eff
#> N[1]     1.003   810
#> N[2]     1.002  2300
#> N[3]     1.002  1200
#> N[4]     1.005   370
#> N[5]     1.003   880
#> N[6]     1.003   710
#> N[7]     1.003   740
#> N[8]     1.001 12000
#> N[9]     1.001 22000
#> N[10]    1.001  9700
#> N[11]    1.001  3000
#> N[12]    1.002  1600
#> N[13]    1.001 27000
#> N[14]    1.001 20000
#> N[15]    1.001 27000
#> meanr    1.003   960
#> r[1]     1.001 27000
#> r[2]     1.001  7500
#> r[3]     1.003   900
#> r[4]     1.008   210
#> r[5]     1.001  3600
#> r[6]     1.001 27000
#> r[7]     1.005   410
#> r[8]     1.001  4700
#> r[9]     1.001  4900
#> r[10]    1.001  4800
#> r[11]    1.001  3200
#> r[12]    1.002  1900
#> r[13]    1.001 26000
#> r[14]    1.001 27000
#> sdr      1.001  3400
#> vrrmax   1.001  3400
#> deviance 1.002  1500
#> 
#> For each parameter, n.eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor (at convergence, Rhat=1).
#> 
#> DIC info (using the rule, pD = var(deviance)/2)
#> pD = 13.7 and DIC = 250.0
#> DIC is an estimate of expected predictive error (lower deviance is better).
#> 
# }